% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Clean.R
\name{add_shooter_biometrics}
\alias{add_shooter_biometrics}
\title{Add shooter biometrics to (a) play-by-play(s)}
\usage{
add_shooter_biometrics(play_by_play)
}
\arguments{
\item{play_by_play}{data.frame of play-by-play(s) using the current public schema returned by \code{\link[=gc_play_by_play]{gc_play_by_play()}}, \code{\link[=gc_play_by_plays]{gc_play_by_plays()}}, \code{\link[=wsc_play_by_play]{wsc_play_by_play()}}, or \code{\link[=wsc_play_by_plays]{wsc_play_by_plays()}}}
}
\value{
data.frame with one row per event (play) and added columns: \code{shooterHeight}, \code{shooterWeight}, \code{shooterHandCode}, \code{shooterAge}, and \code{shooterPositionCode}
}
\description{
\code{add_shooter_biometrics()} adds shooter biometrics (height, weight, hand, age at game date, and position) to (a) play-by-play(s) for shot attempts.
}
