% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Award.R
\name{award_winners}
\alias{award_winners}
\title{Access all the award winners/finalists}
\usage{
award_winners()
}
\value{
data.frame with one row per winner/finalist
}
\description{
\code{award_winners()} scrapes all the award winners/finalists.
}
\examples{
all_award_winners <- award_winners()
}
