% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Award.R
\name{awards}
\alias{awards}
\title{Access all the awards}
\usage{
awards()
}
\value{
data.frame with one row per award
}
\description{
\code{awards()} scrapes all the awards.
}
\examples{
all_awards <- awards()
}
