% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Coach.R
\name{coach_career_statistics}
\alias{coach_career_statistics}
\alias{coach_career_stats}
\title{Access the career statistics for all the coaches}
\usage{
coach_career_statistics()

coach_career_stats()
}
\value{
data.frame with one row per coach
}
\description{
\code{coach_career_statistics()} scrapes the career results for all the coaches.
}
\examples{
coach_career_stats <- coach_career_statistics()
}
