% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Draft.R
\name{combine_reports}
\alias{combine_reports}
\title{Access the draft combine reports}
\usage{
combine_reports()
}
\value{
data.frame with one row per player
}
\description{
\code{combine_reports()} scrapes the draft combine reports.
}
\examples{
combine_reports <- combine_reports()
}
