% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Game.R
\name{.build_api_html_match_table}
\alias{.build_api_html_match_table}
\title{Build API-side HTML matching features}
\usage{
.build_api_html_match_table(play_by_play)
}
\arguments{
\item{play_by_play}{data.frame play-by-play object}
}
\value{
data.frame of API-side matching features
}
\description{
\code{.build_api_html_match_table()} derives the API-side matching features used
to align HTML play-by-play rows back to API play-by-play rows.
}
\keyword{internal}
