% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Clean.R
\name{.compute_pbp_deltas_in_r}
\alias{.compute_pbp_deltas_in_r}
\title{Compute play-by-play deltas in R}
\usage{
.compute_pbp_deltas_in_r(play_by_play)
}
\arguments{
\item{play_by_play}{data.frame of play-by-play(s) using the current public schema returned by \code{\link[=gc_play_by_play]{gc_play_by_play()}}, \code{\link[=gc_play_by_plays]{gc_play_by_plays()}}, \code{\link[=wsc_play_by_play]{wsc_play_by_play()}}, or \code{\link[=wsc_play_by_plays]{wsc_play_by_plays()}}}
}
\value{
named list containing delta context vectors
}
\description{
\code{.compute_pbp_deltas_in_r()} is the pure-R fallback for event-to-event delta
calculations.
}
\keyword{internal}
