% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Clean.R
\name{.covid_round_robin_shootout_game_ids}
\alias{.covid_round_robin_shootout_game_ids}
\title{Identify Return-to-Play round-robin shootout games}
\usage{
.covid_round_robin_shootout_game_ids()
}
\value{
integer vector of game IDs
}
\description{
\code{.covid_round_robin_shootout_game_ids()} returns the 2019-20 round-robin
seeding games that used regular-season overtime/shootout rules despite
carrying playoff game IDs.
}
\keyword{internal}
