% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Game.R
\name{.has_any_on_ice_player_ids}
\alias{.has_any_on_ice_player_ids}
\title{Identify populated on-ice player rows}
\usage{
.has_any_on_ice_player_ids(play_by_play)
}
\arguments{
\item{play_by_play}{data.frame play-by-play object}
}
\value{
logical vector
}
\description{
\code{.has_any_on_ice_player_ids()} returns a logical vector indicating whether a
play-by-play row already has any scalar on-ice player IDs assigned.
}
\keyword{internal}
