% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Game.R
\name{.html_clocks_to_seconds}
\alias{.html_clocks_to_seconds}
\title{Convert clocks to seconds}
\usage{
.html_clocks_to_seconds(x)
}
\arguments{
\item{x}{character vector}
}
\value{
integer vector
}
\description{
\code{.html_clocks_to_seconds()} vectorizes \code{.html_time_to_seconds()} for shift
report clock columns.
}
\keyword{internal}
