% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Game.R
\name{.html_desc_owner_team_id}
\alias{.html_desc_owner_team_id}
\title{Resolve the owner team from an HTML description}
\usage{
.html_desc_owner_team_id(
  description,
  home_abbrev,
  away_abbrev,
  home_team_id,
  away_team_id
)
}
\arguments{
\item{description}{HTML event description}

\item{home_abbrev}{home-team abbreviation}

\item{away_abbrev}{away-team abbreviation}

\item{home_team_id}{home-team ID}

\item{away_team_id}{away-team ID}
}
\value{
integer team ID or \code{NA_integer_}
}
\description{
\code{.html_desc_owner_team_id()} infers the event owner team ID from the leading
team abbreviation embedded in an HTML event description.
}
\keyword{internal}
