% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Game.R
\name{.html_extract_actor_player_ids}
\alias{.html_extract_actor_player_ids}
\title{Extract actor player IDs from an HTML event}
\usage{
.html_extract_actor_player_ids(
  description,
  type_desc_key,
  owner_team_id,
  home_team_id,
  away_team_id,
  home_abbrev,
  away_abbrev,
  roster_lookup
)
}
\arguments{
\item{description}{HTML event description}

\item{type_desc_key}{internal event type key}

\item{owner_team_id}{owner team ID}

\item{home_team_id}{home-team ID}

\item{away_team_id}{away-team ID}

\item{home_abbrev}{home-team abbreviation}

\item{away_abbrev}{away-team abbreviation}

\item{roster_lookup}{standardized roster lookup data.frame}
}
\value{
named list of actor player IDs
}
\description{
\code{.html_extract_actor_player_ids()} resolves primary, secondary, and tertiary
actor player IDs from HTML event descriptions using the roster lookup.
}
\keyword{internal}
