% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Game.R
\name{.html_extract_tagged_players}
\alias{.html_extract_tagged_players}
\title{Extract team-tagged player references from HTML}
\usage{
.html_extract_tagged_players(
  description,
  home_team_id,
  away_team_id,
  home_abbrev,
  away_abbrev,
  roster_lookup
)
}
\arguments{
\item{description}{HTML event description}

\item{home_team_id}{home-team ID}

\item{away_team_id}{away-team ID}

\item{home_abbrev}{home-team abbreviation}

\item{away_abbrev}{away-team abbreviation}

\item{roster_lookup}{standardized roster lookup data.frame}
}
\value{
data.frame of ordered team/player references
}
\description{
\code{.html_extract_tagged_players()} resolves explicitly team-tagged HTML player
references such as \code{TOR #42} into ordered team/player pairs.
}
\keyword{internal}
