% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Game.R
\name{.html_pbp_report_url}
\alias{.html_pbp_report_url}
\title{Build the HTML play-by-play report URL}
\usage{
.html_pbp_report_url(game)
}
\arguments{
\item{game}{game ID}
}
\value{
character scalar URL
}
\description{
\code{.html_pbp_report_url()} returns the NHL HTML play-by-play report URL for a
game.
}
\keyword{internal}
