% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utility.R
\name{.html_report_request}
\alias{.html_report_request}
\title{Build an HTML report request}
\usage{
.html_report_request(url)
}
\arguments{
\item{url}{character scalar}
}
\value{
httr2 request object
}
\description{
\code{.html_report_request()} constructs an HTML report request with the same retry
policy used for API calls.
}
\keyword{internal}
