% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Game.R
\name{.html_shift_report_url}
\alias{.html_shift_report_url}
\title{Build the HTML shift report URL}
\usage{
.html_shift_report_url(game, team_tag)
}
\arguments{
\item{game}{game ID}

\item{team_tag}{HTML report team tag}
}
\value{
character scalar URL
}
\description{
\code{.html_shift_report_url()} returns a team-specific NHL HTML shift report URL
for a game. \code{team_tag} is usually \code{TH} for the home report or \code{TV} for the
away report.
}
\keyword{internal}
