% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Game.R
\name{.html_time_to_seconds}
\alias{.html_time_to_seconds}
\title{Convert a clock duration to seconds}
\usage{
.html_time_to_seconds(x)
}
\arguments{
\item{x}{character scalar}
}
\value{
integer scalar or \code{NA_integer_}
}
\description{
\code{.html_time_to_seconds()} converts \code{MM:SS} or \code{HH:MM:SS} strings to seconds.
}
\keyword{internal}
