% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Clean.R
\name{.is_short_missed_shot}
\alias{.is_short_missed_shot}
\title{Identify short missed shots}
\usage{
.is_short_missed_shot(play_by_play)
}
\arguments{
\item{play_by_play}{data.frame play-by-play object}
}
\value{
logical vector
}
\description{
\code{.is_short_missed_shot()} flags missed shots with reason \code{short}, which are
excluded from Fenwick and Corsi calculations.
}
\keyword{internal}
