% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utility.R
\name{.nhl_json_from_response}
\alias{.nhl_json_from_response}
\title{Parse an NHL API response as JSON}
\usage{
.nhl_json_from_response(resp)
}
\arguments{
\item{resp}{httr2 response object}
}
\value{
parsed JSON (i.e., data.frame or list)
}
\description{
\code{.nhl_json_from_response()} converts an \code{httr2} response object into a parsed
JSON object using the package's standard flattening settings.
}
\keyword{internal}
