% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Clean.R
\name{.on_ice_int_col}
\alias{.on_ice_int_col}
\title{Extract an integer on-ice column}
\usage{
.on_ice_int_col(df, name)
}
\arguments{
\item{df}{data.frame source object}

\item{name}{column name to extract}
}
\value{
integer vector
}
\description{
\code{.on_ice_int_col()} safely extracts an integer column from a data frame and
returns \code{NA} values when the column is absent.
}
\keyword{internal}
