% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Game.R
\name{.on_ice_scalar_column_names}
\alias{.on_ice_scalar_column_names}
\title{List all scalar on-ice columns}
\usage{
.on_ice_scalar_column_names()
}
\value{
character vector of scalar on-ice columns
}
\description{
\code{.on_ice_scalar_column_names()} returns the scalar on-ice player ID and
shift-timing columns used by the public play-by-play schema.
}
\keyword{internal}
