% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Game.R
\name{.on_ice_skater_slots}
\alias{.on_ice_skater_slots}
\title{Return the number of on-ice skater slots}
\usage{
.on_ice_skater_slots()
}
\value{
integer scalar
}
\description{
\code{.on_ice_skater_slots()} returns the maximum number of skater slots tracked
per team in play-by-play outputs.
}
\keyword{internal}
