% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Game.R
\name{.parse_html_elapsed_clock}
\alias{.parse_html_elapsed_clock}
\title{Parse an HTML elapsed clock}
\usage{
.parse_html_elapsed_clock(x)
}
\arguments{
\item{x}{HTML time field}
}
\value{
integer seconds elapsed in period or \code{NA_integer_}
}
\description{
\code{.parse_html_elapsed_clock()} extracts the elapsed \code{MM:SS} clock from an HTML
report time field and converts it to elapsed seconds in period.
}
\keyword{internal}
