% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Game.R
\name{.parse_html_on_ice_cell}
\alias{.parse_html_on_ice_cell}
\title{Parse an HTML on-ice cell}
\usage{
.parse_html_on_ice_cell(text, team_id, roster_lookup)
}
\arguments{
\item{text}{on-ice cell text}

\item{team_id}{team ID for the cell}

\item{roster_lookup}{standardized roster lookup data.frame}
}
\value{
named list containing goalie and skater player IDs
}
\description{
\code{.parse_html_on_ice_cell()} resolves a goalie and skater player-ID list from
a single team on-ice HTML cell.
}
\keyword{internal}
