% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Game.R
\name{.parse_html_period_label}
\alias{.parse_html_period_label}
\title{Parse an HTML period label}
\usage{
.parse_html_period_label(x, is_playoffs = FALSE)
}
\arguments{
\item{x}{HTML period label}

\item{is_playoffs}{logical; whether the game is a playoff game}
}
\value{
integer period number or \code{NA_integer_}
}
\description{
\code{.parse_html_period_label()} converts HTML report period labels such as \code{OT}
and \code{SO} into integer period numbers.
}
\keyword{internal}
