% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Clean.R
\name{.pbp_delta_public_cols}
\alias{.pbp_delta_public_cols}
\title{List public delta column names}
\usage{
.pbp_delta_public_cols()
}
\value{
character vector of public delta column names
}
\description{
\code{.pbp_delta_public_cols()} returns the public-facing delta columns inserted
by \code{\link[=add_deltas]{add_deltas()}}.
}
\keyword{internal}
