% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utility.R
\name{.perform_parallel_requests}
\alias{.perform_parallel_requests}
\title{Perform multiple requests concurrently}
\usage{
.perform_parallel_requests(reqs, on_error = c("stop", "return", "continue"))
}
\arguments{
\item{reqs}{list of httr2 request objects}

\item{on_error}{forwarded to \code{\link[httr2:req_perform_parallel]{httr2::req_perform_parallel()}}}
}
\value{
list of responses or httr2 failure objects
}
\description{
\code{.perform_parallel_requests()} executes a list of \code{httr2} requests with
libcurl multi support and preserves the input names on the output.
}
\keyword{internal}
