% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Share.R
\name{.share_game_labels}
\alias{.share_game_labels}
\title{Read game labels for share plots}
\usage{
.share_game_labels(game)
}
\arguments{
\item{game}{game ID}
}
\value{
named list of game labels
}
\description{
\code{.share_game_labels()} fetches the GameCenter summary and extracts the team
abbreviations and game date used in plot titles.
}
\keyword{internal}
