% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Clean.R
\name{.shift_situation_add_metadata}
\alias{.shift_situation_add_metadata}
\title{Add team and player metadata}
\usage{
.shift_situation_add_metadata(summary, roster = NULL)
}
\arguments{
\item{summary}{data.frame summary input}

\item{roster}{optional data.frame roster input}
}
\value{
Internal helper output.
}
\description{
\code{.shift_situation_add_metadata()} is an internal helper for \code{calculate_shift_times_by_situation()}.
}
\keyword{internal}
