% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Clean.R
\name{.shift_situation_add_totals}
\alias{.shift_situation_add_totals}
\title{Add situation total columns}
\usage{
.shift_situation_add_totals(summary, overlaps)
}
\arguments{
\item{summary}{data.frame summary input}

\item{overlaps}{data.frame overlap totals}
}
\value{
Internal helper output.
}
\description{
\code{.shift_situation_add_totals()} is an internal helper for \code{calculate_shift_times_by_situation()}.
}
\keyword{internal}
