% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Clean.R
\name{.shift_situation_intervals}
\alias{.shift_situation_intervals}
\title{Build shift-situation intervals}
\usage{
.shift_situation_intervals(play_by_play, shift_chart)
}
\arguments{
\item{play_by_play}{data.frame play-by-play input}

\item{shift_chart}{data.frame shift-chart input}
}
\value{
Internal helper output.
}
\description{
\code{.shift_situation_intervals()} is an internal helper for \code{calculate_shift_times_by_situation()}.
}
\keyword{internal}
