% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Clean.R
\name{.shift_situation_order_output}
\alias{.shift_situation_order_output}
\title{Order shift-situation output columns}
\usage{
.shift_situation_order_output(summary)
}
\arguments{
\item{summary}{data.frame summary input}
}
\value{
Internal helper output.
}
\description{
\code{.shift_situation_order_output()} is an internal helper for \code{calculate_shift_times_by_situation()}.
}
\keyword{internal}
