% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Clean.R
\name{.shift_situation_perspective_code}
\alias{.shift_situation_perspective_code}
\title{Convert situation codes to player-team perspective}
\usage{
.shift_situation_perspective_code(situation_code, player_is_home)
}
\arguments{
\item{situation_code}{integer or character situation codes}

\item{player_is_home}{logical vector indicating home-team perspective}
}
\value{
Internal helper output.
}
\description{
\code{.shift_situation_perspective_code()} is an internal helper for \code{calculate_shift_times_by_situation()}.
}
\keyword{internal}
