% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Model.R
\name{.xg_bake_matrix}
\alias{.xg_bake_matrix}
\title{Bake xG model matrix}
\usage{
.xg_bake_matrix(data, spec)
}
\arguments{
\item{data}{data.frame feature input}

\item{spec}{xG preprocessing specification}
}
\value{
Internal helper output.
}
\description{
\code{.xg_bake_matrix()} is an internal helper for \code{calculate_expected_goals()}.
}
\keyword{internal}
