% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Model.R
\name{.xg_booster_url}
\alias{.xg_booster_url}
\title{Build xG booster URL}
\usage{
.xg_booster_url(path)
}
\arguments{
\item{path}{character file or bundle path}
}
\value{
Internal helper output.
}
\description{
\code{.xg_booster_url()} is an internal helper for \code{calculate_expected_goals()}.
}
\keyword{internal}
