% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Model.R
\name{.xg_build_skater_matrix}
\alias{.xg_build_skater_matrix}
\title{Build skater-slot matrix}
\usage{
.xg_build_skater_matrix(data, suffix, mode = "numeric")
}
\arguments{
\item{data}{data.frame feature input}

\item{suffix}{character skater-slot column suffix}

\item{mode}{character matrix storage mode}
}
\value{
Internal helper output.
}
\description{
\code{.xg_build_skater_matrix()} is an internal helper for \code{calculate_expected_goals()}.
}
\keyword{internal}
