% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Model.R
\name{.xg_cache_dir}
\alias{.xg_cache_dir}
\title{Locate xG cache directory}
\usage{
.xg_cache_dir()
}
\value{
Internal helper output.
}
\description{
\code{.xg_cache_dir()} is an internal helper for \code{calculate_expected_goals()}.
}
\keyword{internal}
