% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Model.R
\name{.xg_cached_booster_path}
\alias{.xg_cached_booster_path}
\title{Build cached xG booster path}
\usage{
.xg_cached_booster_path(path, bundle = .xg_load_bundle())
}
\arguments{
\item{path}{character file or bundle path}

\item{bundle}{xG model bundle}
}
\value{
Internal helper output.
}
\description{
\code{.xg_cached_booster_path()} is an internal helper for \code{calculate_expected_goals()}.
}
\keyword{internal}
