% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Model.R
\name{.xg_default_model_base_url}
\alias{.xg_default_model_base_url}
\title{Get default xG model base URL}
\usage{
.xg_default_model_base_url()
}
\value{
Internal helper output.
}
\description{
\code{.xg_default_model_base_url()} is an internal helper for \code{calculate_expected_goals()}.
}
\keyword{internal}
