% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Model.R
\name{.xg_download_booster}
\alias{.xg_download_booster}
\title{Download xG booster file}
\usage{
.xg_download_booster(row, bundle = .xg_load_bundle())
}
\arguments{
\item{row}{data.frame model-index row}

\item{bundle}{xG model bundle}
}
\value{
Internal helper output.
}
\description{
\code{.xg_download_booster()} is an internal helper for \code{calculate_expected_goals()}.
}
\keyword{internal}
