% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Model.R
\name{.xg_expected_booster_sha256}
\alias{.xg_expected_booster_sha256}
\title{Read expected xG booster checksum}
\usage{
.xg_expected_booster_sha256(row)
}
\arguments{
\item{row}{data.frame model-index row}
}
\value{
Internal helper output.
}
\description{
\code{.xg_expected_booster_sha256()} is an internal helper for \code{calculate_expected_goals()}.
}
\keyword{internal}
