% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Model.R
\name{.xg_extdata_dir}
\alias{.xg_extdata_dir}
\title{Locate xG extdata directory}
\usage{
.xg_extdata_dir()
}
\value{
Internal helper output.
}
\description{
\code{.xg_extdata_dir()} is an internal helper for \code{calculate_expected_goals()}.
}
\keyword{internal}
