% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Model.R
\name{.xg_extract_slot_indices}
\alias{.xg_extract_slot_indices}
\title{Extract skater slot indices}
\usage{
.xg_extract_slot_indices(data, suffix)
}
\arguments{
\item{data}{data.frame feature input}

\item{suffix}{character skater-slot column suffix}
}
\value{
Internal helper output.
}
\description{
\code{.xg_extract_slot_indices()} is an internal helper for \code{calculate_expected_goals()}.
}
\keyword{internal}
