% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Model.R
\name{.xg_fetch_shift_data}
\alias{.xg_fetch_shift_data}
\title{Fetch shift data for xG features}
\usage{
.xg_fetch_shift_data(game_ids)
}
\arguments{
\item{game_ids}{integer game ID vector}
}
\value{
Internal helper output.
}
\description{
\code{.xg_fetch_shift_data()} is an internal helper for \code{calculate_expected_goals()}.
}
\keyword{internal}
