% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Model.R
\name{.xg_fill_goalie_against_fallback}
\alias{.xg_fill_goalie_against_fallback}
\title{Fill goalie-against fallback columns}
\usage{
.xg_fill_goalie_against_fallback(play_by_play)
}
\arguments{
\item{play_by_play}{data.frame play-by-play input}
}
\value{
Internal helper output.
}
\description{
\code{.xg_fill_goalie_against_fallback()} is an internal helper for \code{calculate_expected_goals()}.
}
\keyword{internal}
