% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Model.R
\name{.xg_is_inner_slot_shot}
\alias{.xg_is_inner_slot_shot}
\title{Flag inner-slot shots}
\usage{
.xg_is_inner_slot_shot(x_coord_norm, abs_y_coord_norm)
}
\arguments{
\item{x_coord_norm}{numeric normalized x-coordinate vector}

\item{abs_y_coord_norm}{numeric absolute normalized y-coordinate vector}
}
\value{
Internal helper output.
}
\description{
\code{.xg_is_inner_slot_shot()} is an internal helper for \code{calculate_expected_goals()}.
}
\keyword{internal}
