% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Model.R
\name{.xg_load_booster}
\alias{.xg_load_booster}
\title{Load xG booster model}
\usage{
.xg_load_booster(model_key, bundle = .xg_load_bundle())
}
\arguments{
\item{model_key}{character xG model key}

\item{bundle}{xG model bundle}
}
\value{
Internal helper output.
}
\description{
\code{.xg_load_booster()} is an internal helper for \code{calculate_expected_goals()}.
}
\keyword{internal}
