% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Model.R
\name{.xg_load_bundle}
\alias{.xg_load_bundle}
\title{Load xG model bundle}
\usage{
.xg_load_bundle()
}
\value{
Internal helper output.
}
\description{
\code{.xg_load_bundle()} is an internal helper for \code{calculate_expected_goals()}.
}
\keyword{internal}
