% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Model.R
\name{.xg_make_type_desc_key_prev}
\alias{.xg_make_type_desc_key_prev}
\title{Build previous event type key}
\usage{
.xg_make_type_desc_key_prev(
  type_desc_key_prev,
  reason_prev,
  shot_type_prev,
  event_owner_team_id_prev,
  event_owner_team_id
)
}
\arguments{
\item{type_desc_key_prev}{character previous event type keys}

\item{reason_prev}{character previous missed-shot reasons}

\item{shot_type_prev}{character previous shot types}

\item{event_owner_team_id_prev}{integer previous event owner team IDs}

\item{event_owner_team_id}{integer current event owner team IDs}
}
\value{
Internal helper output.
}
\description{
\code{.xg_make_type_desc_key_prev()} is an internal helper for \code{calculate_expected_goals()}.
}
\keyword{internal}
