% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Model.R
\name{.xg_model_index_row}
\alias{.xg_model_index_row}
\title{Find xG model index row}
\usage{
.xg_model_index_row(model_key, bundle = .xg_load_bundle())
}
\arguments{
\item{model_key}{character xG model key}

\item{bundle}{xG model bundle}
}
\value{
Internal helper output.
}
\description{
\code{.xg_model_index_row()} is an internal helper for \code{calculate_expected_goals()}.
}
\keyword{internal}
