% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Model.R
\name{.xg_model_key}
\alias{.xg_model_key}
\title{Build xG model key}
\usage{
.xg_model_key(target_season, partition)
}
\arguments{
\item{target_season}{integer target season}

\item{partition}{character partition code}
}
\value{
Internal helper output.
}
\description{
\code{.xg_model_key()} is an internal helper for \code{calculate_expected_goals()}.
}
\keyword{internal}
