% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Model.R
\name{.xg_normalize_missed_reason}
\alias{.xg_normalize_missed_reason}
\title{Normalize xG missed-shot reason}
\usage{
.xg_normalize_missed_reason(x)
}
\arguments{
\item{x}{vector input}
}
\value{
Internal helper output.
}
\description{
\code{.xg_normalize_missed_reason()} is an internal helper for \code{calculate_expected_goals()}.
}
\keyword{internal}
