% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Model.R
\name{.xg_normalize_shot_type}
\alias{.xg_normalize_shot_type}
\title{Normalize xG shot type}
\usage{
.xg_normalize_shot_type(x)
}
\arguments{
\item{x}{vector input}
}
\value{
Internal helper output.
}
\description{
\code{.xg_normalize_shot_type()} is an internal helper for \code{calculate_expected_goals()}.
}
\keyword{internal}
