% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Model.R
\name{.xg_partition_shots}
\alias{.xg_partition_shots}
\title{Partition shots by game state}
\usage{
.xg_partition_shots(shots)
}
\arguments{
\item{shots}{data.frame shot rows}
}
\value{
Internal helper output.
}
\description{
\code{.xg_partition_shots()} is an internal helper for \code{calculate_expected_goals()}.
}
\keyword{internal}
