% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Model.R
\name{.xg_percent_rank}
\alias{.xg_percent_rank}
\title{Compute xG percentile rank}
\usage{
.xg_percent_rank(x)
}
\arguments{
\item{x}{vector input}
}
\value{
Internal helper output.
}
\description{
\code{.xg_percent_rank()} is an internal helper for \code{calculate_expected_goals()}.
}
\keyword{internal}
