% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Model.R
\name{.xg_prepare_play_by_play}
\alias{.xg_prepare_play_by_play}
\title{Prepare play-by-play for xG scoring}
\usage{
.xg_prepare_play_by_play(play_by_play)
}
\arguments{
\item{play_by_play}{data.frame play-by-play input}
}
\value{
Internal helper output.
}
\description{
\code{.xg_prepare_play_by_play()} is an internal helper for \code{calculate_expected_goals()}.
}
\keyword{internal}
