% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Model.R
\name{.xg_required_shift_cols}
\alias{.xg_required_shift_cols}
\title{Identify required xG shift columns}
\usage{
.xg_required_shift_cols(play_by_play)
}
\arguments{
\item{play_by_play}{data.frame play-by-play input}
}
\value{
Internal helper output.
}
\description{
\code{.xg_required_shift_cols()} is an internal helper for \code{calculate_expected_goals()}.
}
\keyword{internal}
