% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Model.R
\name{.xg_resolve_booster_path}
\alias{.xg_resolve_booster_path}
\title{Resolve xG booster file path}
\usage{
.xg_resolve_booster_path(
  model_key,
  bundle = .xg_load_bundle(),
  download = TRUE
)
}
\arguments{
\item{model_key}{character xG model key}

\item{bundle}{xG model bundle}

\item{download}{logical whether downloading is allowed}
}
\value{
Internal helper output.
}
\description{
\code{.xg_resolve_booster_path()} is an internal helper for \code{calculate_expected_goals()}.
}
\keyword{internal}
