% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Model.R
\name{.xg_row_median}
\alias{.xg_row_median}
\title{Compute row medians with missing values}
\usage{
.xg_row_median(mat)
}
\arguments{
\item{mat}{numeric matrix}
}
\value{
Internal helper output.
}
\description{
\code{.xg_row_median()} is an internal helper for \code{calculate_expected_goals()}.
}
\keyword{internal}
