% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Model.R
\name{.xg_row_min}
\alias{.xg_row_min}
\title{Compute row minimums with missing values}
\usage{
.xg_row_min(mat)
}
\arguments{
\item{mat}{numeric matrix}
}
\value{
Internal helper output.
}
\description{
\code{.xg_row_min()} is an internal helper for \code{calculate_expected_goals()}.
}
\keyword{internal}
