% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Model.R
\name{.xg_score_xgboost}
\alias{.xg_score_xgboost}
\title{Score shots with xGBoost}
\usage{
.xg_score_xgboost(df, model_key, bundle = .xg_load_bundle())
}
\arguments{
\item{df}{data.frame model-frame input}

\item{model_key}{character xG model key}

\item{bundle}{xG model bundle}
}
\value{
Internal helper output.
}
\description{
\code{.xg_score_xgboost()} is an internal helper for \code{calculate_expected_goals()}.
}
\keyword{internal}
