% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Model.R
\name{.xg_select_target_season}
\alias{.xg_select_target_season}
\title{Select xG target season}
\usage{
.xg_select_target_season(game_id, bundle = .xg_load_bundle())
}
\arguments{
\item{game_id}{integer game ID vector}

\item{bundle}{xG model bundle}
}
\value{
Internal helper output.
}
\description{
\code{.xg_select_target_season()} is an internal helper for \code{calculate_expected_goals()}.
}
\keyword{internal}
