% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Model.R
\name{.xg_to_logical}
\alias{.xg_to_logical}
\title{Coerce xG feature to logical}
\usage{
.xg_to_logical(x)
}
\arguments{
\item{x}{vector input}
}
\value{
Internal helper output.
}
\description{
\code{.xg_to_logical()} is an internal helper for \code{calculate_expected_goals()}.
}
\keyword{internal}
