% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Model.R
\name{.xg_validate_bundle}
\alias{.xg_validate_bundle}
\title{Validate xG model bundle}
\usage{
.xg_validate_bundle(bundle)
}
\arguments{
\item{bundle}{xG model bundle}
}
\value{
Internal helper output.
}
\description{
\code{.xg_validate_bundle()} is an internal helper for \code{calculate_expected_goals()}.
}
\keyword{internal}
