% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utility.R
\name{dot_to_camel}
\alias{dot_to_camel}
\title{Convert dot-delimited names to camelCase}
\usage{
dot_to_camel(x)
}
\arguments{
\item{x}{character vector}
}
\value{
character vector
}
\description{
Convert dot-delimited names to camelCase
}
\keyword{internal}
