% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Draft.R
\name{draft_prospects}
\alias{draft_prospects}
\title{Access all the draft prospects}
\usage{
draft_prospects()
}
\value{
data.frame with one row per player
}
\description{
\code{draft_prospects()} retrieves all the draft prospects as a \code{data.frame} where each row represents player and includes detail on player identity, role, handedness, and biographical profile plus broadcast carriage, media availability, and viewing-link metadata.
}
\examples{
# May take >5s, so skip.
\donttest{all_prospects <- draft_prospects()}
}
