% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Draft.R
\name{drafts}
\alias{drafts}
\title{Access all the drafts}
\usage{
drafts()
}
\value{
data.frame with one row per draft
}
\description{
\code{drafts()} merges the records draft master and stats draft-round endpoints,
returning one row per draft year with IDs, dates, location, and round
metadata.
}
\examples{
all_drafts <- drafts()
}
