% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Graph.R
\name{draw_NHL_rink}
\alias{draw_NHL_rink}
\title{Draw a full NHL rink}
\usage{
draw_NHL_rink()
}
\value{
\code{NULL}
}
\description{
\code{draw_NHL_rink()} draws a full NHL rink such that the x and y coordinates
span -100 to 100 and -43 to +43, respectively. Use \code{\link[graphics:points]{graphics::points()}}
to create custom graphs; check out an example on the online documentation!
}
\examples{
draw_NHL_rink()
}
