% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Graph.R
\name{draw_rink_arc}
\alias{draw_rink_arc}
\title{Draw an arc on the rink}
\usage{
draw_rink_arc(cx, cy, r, theta1, theta2, n = 100, col = "black", lwd = 1)
}
\arguments{
\item{cx}{numeric x coordinate of the circle center}

\item{cy}{numeric y coordinate of the circle center}

\item{r}{numeric radius}

\item{theta1}{numeric start angle of arc}

\item{theta2}{numeric end angle of arc}

\item{n}{integer number of points used to approximate arc}

\item{col}{character color}

\item{lwd}{numeric line width}
}
\value{
\code{NULL}
}
\description{
\code{draw_rink_arc()} draws an arc on the rink.
}
\keyword{internal}
