% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Graph.R
\name{draw_rink_circle}
\alias{draw_rink_circle}
\title{Draw a circle on the rink}
\usage{
draw_rink_circle(x, y, r, n = 200, col = "black", lwd = 1)
}
\arguments{
\item{x}{numeric x coordinate of the circle center}

\item{y}{numeric y coordinate of the circle center}

\item{r}{numeric radius}

\item{n}{integer number of points used to approximate circle}

\item{col}{character color}

\item{lwd}{numeric line width}
}
\value{
\code{NULL}
}
\description{
\code{draw_rink_circle()} draws a circle on the rink.
}
\keyword{internal}
