% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Other.R
\name{espn_lookup}
\alias{espn_lookup}
\title{Look up team or player by ESPN ID}
\usage{
espn_lookup(
  year = nhl_season_to_espn_season(get_season_now$seasonId),
  type = "teams",
  id = 1
)
}
\arguments{
\item{year}{integer Year in YYYY (start of season)}

\item{type}{string Type of 'team' or 'player'}

\item{id}{integer ESPN ID}
}
\value{
list with various items
}
\description{
Look up team or player by ESPN ID
}
\examples{
espn_20252026_FLA <- espn_lookup(2025, 'team', 26)
}
