% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ESPNTeam.R
\name{espn_teams}
\alias{espn_teams}
\title{Access all the ESPN teams}
\usage{
espn_teams()
}
\value{
data.frame with one row per ESPN team
}
\description{
\code{espn_teams()} scrapes all the ESPN teams.
}
\examples{
all_ESPN_teams <- espn_teams()
}
