% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Draft.R
\name{expansion_drafts}
\alias{expansion_drafts}
\title{Access all the expansion drafts}
\usage{
expansion_drafts()
}
\value{
data.frame with one row per expansion draft
}
\description{
\code{expansion_drafts()} scrapes all the expansion drafts.
}
\examples{
all_expansion_drafts <- expansion_drafts()
}
