% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Clean.R
\name{flag_is_rush}
\alias{flag_is_rush}
\title{Flag if the shot attempt is a rush attempt or not for all the shots in a
play-by-play}
\usage{
flag_is_rush(
  data,
  is_home_name = "isHome",
  seconds_elapsed_in_game_name = "secondsElapsedInGame",
  game_id_name = "gameId",
  sort_order_name = "sortOrder",
  type_description_name = "typeDescKey",
  zone_code_name = "zoneCode",
  is_rush_name = "isRush"
)
}
\arguments{
\item{data}{data.frame of play-by-play(s); see \code{\link[=gc_play_by_play]{gc_play_by_play()}} and/or
\code{\link[=wsc_play_by_play]{wsc_play_by_play()}} for reference}

\item{is_home_name}{name of column that contains home/not logical
indicator; see \code{\link[=flag_is_home]{flag_is_home()}} for reference}

\item{seconds_elapsed_in_game_name}{name of column that contains seconds
elapsed in game; see \code{\link[=strip_time_period]{strip_time_period()}} for reference}

\item{sort_order_name}{name of column that contains sort order}

\item{type_description_name}{name of column that contains event type
description}

\item{zone_code_name}{name of column that contains zone code}

\item{is_rush_name}{name of column that you want contain rush/not
logical indicator}
}
\value{
data.frame with one row per event (play)
}
\description{
\code{flag_is_rush()} flags if the shot attempt is a rush attempt or not for all
the shots in a play-by-play.
}
\examples{
# May take >5s, so skip.
\donttest{
  test                      <- gc_play_by_play()
  test_time_period_stripped <- strip_time_period(test)
  test_is_home_flagged      <- flag_is_home(test_time_period_stripped)
  test_is_rush_flagged      <- flag_is_rush(test_is_home_flagged)
}
}
