% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Franchise.R
\name{franchise_playoff_situational_results}
\alias{franchise_playoff_situational_results}
\title{Access the playoff series results for all the franchises by situation}
\usage{
franchise_playoff_situational_results()
}
\value{
data.frame with one row per franchise per situation
}
\description{
\code{franchise_playoff_situational_results()} scrapes the playoff series
results for all the franchises by situation.
}
\examples{
franchise_playoff_situational_results <- 
  franchise_playoff_situational_results()
}
