% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Franchise.R
\name{franchise_statistics}
\alias{franchise_statistics}
\alias{franchise_stats}
\title{Access the all-time statistics for all the franchises by game type}
\usage{
franchise_statistics()

franchise_stats()
}
\value{
data.frame with one row per franchise per game type
}
\description{
\code{franchise_statistics()} retrieves the all-time statistics for all the franchises by game type as a \code{data.frame} where each row represents franchise per game type and includes detail on date/season filtering windows and chronological context, team identity, affiliation, and matchup-side context, and production, workload, efficiency, and result-level performance outcomes.
}
\examples{
franchise_stats <- franchise_statistics()
}
