% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Other.R
\name{game_partner_odds}
\alias{game_partner_odds}
\title{Get the real-time game odds for a country by partnered bookmaker}
\usage{
game_partner_odds(country = "US")
}
\arguments{
\item{country}{two-letter code (e.g., 'CA'); see \code{\link[=countries]{countries()}} for
reference}
}
\value{
data.frame with one row per game
}
\description{
\code{game_partner_odds()} scrapes the real-time game odds for a given
\code{country} by partnered bookmaker.
}
\examples{
game_partner_odds_CA <- game_partner_odds(country = 'CA')
}
