% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Game.R
\name{game_rosters}
\alias{game_rosters}
\title{Access the rosters for a game}
\usage{
game_rosters(game = 2023030417)
}
\arguments{
\item{game}{integer ID (e.g., 2025020275); see \code{\link[=games]{games()}} for reference}
}
\value{
data.frame with one row per player
}
\description{
\code{game_rosters()} scrapes the rosters for a given \code{game}.
}
\examples{
rosters_Martin_Necas_legacy_game <- game_rosters(game = 2025020275)
}
