% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/League.R
\name{game_type}
\alias{game_type}
\title{Access the game type as of now}
\usage{
game_type()
}
\value{
integer in 1:3 (where 1 = pre-season, 2 = regular season, 3
= playoff/post-season)
}
\description{
\code{game_type()} scrapes the current game type.
}
\examples{
game_type_now <- game_type()
}
