% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Load.R
\name{gc_play_by_plays_raw}
\alias{gc_play_by_plays_raw}
\alias{gc_pbps_raw}
\title{Access the raw GameCenter (GC) play-by-plays for a season}
\usage{
gc_play_by_plays_raw(season = 20242025)

gc_pbps_raw(season = 20242025)
}
\arguments{
\item{season}{integer in YYYYYYYY (e.g., 20242025); see \code{\link[=seasons]{seasons()}} for
reference}
}
\value{
data.frame with one row per raw event (play) per game
}
\description{
\code{gc_play_by_plays_raw()} loads the raw GC play-by-plays for a given
\code{season}.
}
\examples{
# May take >5s, so skip.
\donttest{gc_pbps_raw_20212022 <- gc_play_by_plays_raw(season = 20212022)}
}
