% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Award.R
\name{get_award_winners}
\alias{get_award_winners}
\title{Get all award winners/finalists}
\usage{
get_award_winners()
}
\value{
tibble with one row per winner/finalist
}
\description{
\code{get_award_winners()} retrieves information on each award winner or finalist, including but not limited to their player, trophy, and season IDs; name; and vote count.
}
\examples{
all_award_winners <- get_award_winners()
}
