% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AwardOld.R
\name{get_awards}
\alias{get_awards}
\title{Access all the awards}
\usage{
get_awards()
}
\value{
data.frame with one row per award
}
\description{
\code{get_awards()} is deprecated. Use \code{\link[=awards]{awards()}} instead.
}
