% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Other.R
\name{get_configuration}
\alias{get_configuration}
\title{Get configuration for skater, goalie, and team statistics}
\usage{
get_configuration()
}
\value{
list with 5 items
}
\description{
\code{get_configuration()} retrieves information on the outputs of the possible combinations of inputs for \code{get_team_statistics()}, \code{get_skater_statistics()}, and \code{get_goalie_statistics()}.
}
\examples{
config <- get_configuration()
}
