% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OtherOld.R
\name{get_countries}
\alias{get_countries}
\title{Access all the countries}
\usage{
get_countries()
}
\value{
data.frame with one row per country
}
\description{
\code{get_countries()} is deprecated. Use \code{\link[=countries]{countries()}} instead.
}
