% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Draft.R
\name{get_draft_information}
\alias{get_draft_information}
\title{Get draft information for all seasons}
\usage{
get_draft_information()
}
\value{
tibble with one row per season
}
\description{
Get draft information for all seasons
}
\examples{
draft_info <- get_draft_information()
}
