% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Draft.R
\name{get_draft_picks}
\alias{get_draft_picks}
\title{Get all draft picks}
\usage{
get_draft_picks()
}
\value{
tibble with one row per pick
}
\description{
\code{get_draft_picks()} retrieves information on each selection, including but not limited to their player ID, name, draft year, overall number, bio-metrics, and the pick's team history.
}
\examples{
all_draft_picks <- get_draft_picks()
}
