% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Draft.R
\name{get_draft_rankings}
\alias{get_draft_rankings}
\title{Get draft rankings by year and player-type}
\usage{
get_draft_rankings(year = get_season_now()$seasonId\%\%10000, player_type = 1)
}
\arguments{
\item{year}{integer Year in YYYY}

\item{player_type}{integer Player-type where 1=NA Skaters, 2=Int. Skaters,
3=NA Goalies, and 4=Int. Goalies}
}
\value{
tibble with one row per player
}
\description{
Get draft rankings by year and player-type
}
\examples{
draft_rankings_2025 <- get_draft_rankings(year=2025)
}
