% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DraftOld.R
\name{get_draft_tracker}
\alias{get_draft_tracker}
\title{Get the real-time draft tracker}
\usage{
get_draft_tracker()
}
\description{
\code{get_draft_tracker()} is deprecated. Use \code{\link[=ns_draft_tracker]{ns_draft_tracker()}} instead.
}
