% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DraftOld.R
\name{get_drafts}
\alias{get_drafts}
\title{Access all the drafts}
\usage{
get_drafts()
}
\value{
data.frame with one row per draft
}
\description{
\code{get_drafts()} is deprecated. Use \code{\link[=drafts]{drafts()}} instead.
}
