% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ESPNPlayerOld.R
\name{get_espn_athlete}
\alias{get_espn_athlete}
\title{Access the ESPN summary for an athlete (player) and season}
\usage{
get_espn_athlete()
}
\description{
\code{get_espn_athlete()} is defunct. Use \code{\link[=espn_player_summary]{espn_player_summary()}} instead.
}
