% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ESPNAthlete.R
\name{get_espn_athletes}
\alias{get_espn_athletes}
\title{Get all ESPN athletes}
\usage{
get_espn_athletes()
}
\value{
tibble with one row per athlete
}
\description{
Get all ESPN athletes
}
\examples{
all_ESPN_athletes <- get_espn_athletes()
}
