% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ESPNCoachOld.R
\name{get_espn_coach}
\alias{get_espn_coach}
\title{Access the ESPN statistics for a coach and (multiple) season(s)}
\usage{
get_espn_coach()
}
\description{
\code{get_espn_coach()} is defunct. Use \code{\link[=coach_career_statistics]{coach_career_statistics()}} instead.
}
