% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ESPNCoachOld.R
\name{get_espn_coach_career}
\alias{get_espn_coach_career}
\title{Access the career ESPN statistics for a coach}
\usage{
get_espn_coach_career()
}
\description{
\code{get_espn_coach_career()} is defunct. Use \code{\link[=coach_career_statistics]{coach_career_statistics()}} instead.
}
