% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ESPNGameOld.R
\name{get_espn_event}
\alias{get_espn_event}
\title{Access the ESPN summary for an event (game)}
\usage{
get_espn_event(event = 401777460)
}
\arguments{
\item{event}{integer ID (e.g., 401777460); see \code{\link[=espn_games]{espn_games()}} for
reference}
}
\value{
data.frame with one row per event (game)
}
\description{
\code{get_espn_event()} is deprecated. Use \code{\link[=espn_game_summary]{espn_game_summary()}} instead.
}
