% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ESPNEvent.R
\name{get_espn_event_odds}
\alias{get_espn_event_odds}
\title{Get event (game) odds by ESPN Event (Game) ID}
\usage{
get_espn_event_odds(event = 401687600)
}
\arguments{
\item{event}{integer ESPN Event (Game) ID}
}
\value{
tibble with one row per provider
}
\description{
\code{get_espn_event_odds()} retrieves information on each provider for a given \code{event}, including but not limited to its name, favorite and underdog teams, and money-line and spread odds. Access \code{get_espn_events()} for \code{event} reference.
}
\examples{
NJD_BUF_2024_10_04_odds <- get_espn_event_odds(event=401687600)
}
