% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ESPNEvent.R
\name{get_espn_event_officials}
\alias{get_espn_event_officials}
\title{Get event (game) officials by ESPN Event (Game) ID}
\usage{
get_espn_event_officials(event = 401687600)
}
\arguments{
\item{event}{integer ESPN Event (Game) ID}
}
\value{
tibble with one row per official
}
\description{
\code{get_espn_event_officials()} retrieves information on each official for a given \code{event}, including but not limited to its ESPN ID, name, and position. Access \code{get_espn_events()} for \code{event} reference.
}
\examples{
NJD_BUF_2024_10_04_officials <- get_espn_event_officials(event=401687600)
}
