% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ESPNGameOld.R
\name{get_espn_event_play_by_play}
\alias{get_espn_event_play_by_play}
\title{Access the ESPN play-by-play for an event (game)}
\usage{
get_espn_event_play_by_play(event = 401777460)
}
\arguments{
\item{event}{integer ID (e.g., 401777460); see \code{\link[=espn_games]{espn_games()}} for
reference}
}
\value{
data.frame with one row per play
}
\description{
\code{get_espn_event_play_by_play()} is deprecated. Use \code{\link[=espn_play_by_play]{espn_play_by_play()}}
instead.
}
