% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ESPNEvent.R
\name{get_espn_event_stars}
\alias{get_espn_event_stars}
\title{Get the three stars of an ESPN event}
\usage{
get_espn_event_stars(event = 401687600)
}
\description{
\code{get_espn_event_stars()} is defunct. Use \code{\link[=ns_gc_summary]{ns_gc_summary()}} instead.
}
