% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ESPNOther.R
\name{get_espn_futures}
\alias{get_espn_futures}
\title{Get ESPN futures by season}
\usage{
get_espn_futures(season = get_season_now()$seasonId\%\%10000)
}
\arguments{
\item{season}{integer Season in YYYY}
}
\value{
nested tibble with one row per type and book
}
\description{
\code{get_espn_futures()} retrieves futures by type for a given \code{season}.
}
\examples{
ESPN_futures_20252026 <- get_espn_futures(2026)
}
