% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ESPNTeam.R
\name{get_espn_team}
\alias{get_espn_team}
\title{Get team by season and ESPN Team ID}
\usage{
get_espn_team(season = get_season_now()$seasonId\%\%10000, team = 1)
}
\arguments{
\item{season}{integer Season in YYYY}

\item{team}{integer ESPN Team ID}
}
\value{
list with various items
}
\description{
Get team by season and ESPN Team ID
}
\examples{
ESPN_BOS_20242025 <- get_espn_team(season=2025, team=1)
}
