% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ESPNTeam.R
\name{get_espn_teams}
\alias{get_espn_teams}
\title{Get ESPN teams by season}
\usage{
get_espn_teams(season = get_season_now()$seasonId\%\%10000)
}
\arguments{
\item{season}{integer Season in YYYY}
}
\value{
tibble with one row per team
}
\description{
Get ESPN teams by season
}
\examples{
ESPN_teams_20242025 <- get_espn_teams(2025)
}
