% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FranchiseOld.R
\name{get_franchise_season_by_season}
\alias{get_franchise_season_by_season}
\title{Get the season-by-season results for all the franchises}
\usage{
get_franchise_season_by_season()
}
\description{
\code{get_franchise_season_by_season()} is deprecated. Use
\code{\link[=ns_franchise_season_by_season]{ns_franchise_season_by_season()}} instead.
}
