% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Betting.R
\name{get_futures}
\alias{get_futures}
\title{Get ESPN futures by season}
\usage{
get_futures(season = get_season_now$seasonId)
}
\arguments{
\item{season}{integer Season in YYYYYYY}
}
\value{
tibble with one row per future
}
\description{
Get ESPN futures by season
}
\examples{
futures_20242025 <- get_futures(20242025)
}
