% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GameOld.R
\name{get_game_boxscore}
\alias{get_game_boxscore}
\title{Get the boxscore of a game for a team and player type}
\usage{
get_game_boxscore(game = 2025020275, team = "home", player_type = "forwards")
}
\description{
\code{get_game_boxscore()} is deprecated. Use \code{\link[=ns_boxscore]{ns_boxscore()}} instead.
}
