% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GameOld.R
\name{get_game_landing}
\alias{get_game_landing}
\title{Get the GameCenter (GC) summary of a game}
\usage{
get_game_landing(game = 2025020275)
}
\description{
\code{get_game_landing()} is deprecated. Use \code{\link[=ns_gc_summary]{ns_gc_summary()}} instead.
}
