% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Game.R
\name{get_game_story}
\alias{get_game_story}
\title{Get WSC game story by game}
\usage{
get_game_story(game = 2024020602)
}
\arguments{
\item{game}{integer Game ID}
}
\value{
list of 24 items
}
\description{
Get WSC game story by game
}
\examples{
game_story_2024030411 <- get_game_story(game=2024030411)
}
