% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GameOld.R
\name{get_games}
\alias{get_games}
\title{Access all the games}
\usage{
get_games()
}
\value{
data.frame with one row per game
}
\description{
\code{get_games()} is deprecated. Use \code{\link[=games]{games()}} instead.
}
