% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Other.R
\name{get_glossary}
\alias{get_glossary}
\title{Get glossary}
\usage{
get_glossary()
}
\value{
tibble with one row per terminology
}
\description{
\code{get_glossary()} retrieves information on each terminology, including but not limited to their definition and abbreviation.
}
\examples{
glossary <- get_glossary()
}
