% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Other.R
\name{get_injuries}
\alias{get_injuries}
\title{Get injury reports as of now}
\usage{
get_injuries()
}
\value{
nested tibble with one row per team and one row per player
}
\description{
Get injury reports as of now
}
\examples{
injuries_now <- get_injuries()
}
