% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OfficialOld.R
\name{get_officials}
\alias{get_officials}
\title{Access all the officials}
\usage{
get_officials()
}
\value{
data.frame with one row per official
}
\description{
\code{get_officials()} is deprecated. Use \code{\link[=officials]{officials()}} instead.
}
