% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Other.R
\name{get_partner_odds}
\alias{get_partner_odds}
\title{Get partner odds as of now}
\usage{
get_partner_odds(country = "US")
}
\arguments{
\item{country}{string 2-letter country code e.g. 'US'}
}
\value{
tibble with one row per game
}
\description{
Get partner odds as of now
}
\examples{
partner_odds_now_CA <- get_partner_odds(country='CA')
}
