% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Player.R
\name{get_player_game_log}
\alias{get_player_game_log}
\title{Get game-log by player, season, and game-type}
\usage{
get_player_game_log(
  player = 8480039,
  season = get_season_now()$seasonId,
  game_type = 2
)
}
\arguments{
\item{player}{integer Player ID}

\item{season}{integer Season in YYYYYYYY}

\item{game_type}{integer Game-type where 2=regular and 3=playoffs}
}
\value{
tibble with one row per game
}
\description{
Get game-log by player, season, and game-type
}
\examples{
playoff_Mikko_Rantanen_gl_20242025 <- get_player_game_log(
  player=8478420,
  season=20242025,
  game_type=3
)
}
