% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Player.R
\name{get_player_information}
\alias{get_player_information}
\title{Get a player's information}
\usage{
get_player_information(player_id)
}
\arguments{
\item{player_id}{integer Player ID}
}
\value{
list with 36 items
}
\description{
Get a player's information
}
