% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Player.R
\name{get_player_landing}
\alias{get_player_landing}
\title{Get landing by player}
\usage{
get_player_landing(player = 8480039)
}
\arguments{
\item{player}{integer Player ID}
}
\value{
list with 36 items
}
\description{
Get landing by player
}
\examples{
Mikko_Rantanen_landing <- get_player_landing(player=8478420)
}
