% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Players.R
\name{get_players}
\alias{get_players}
\title{Get all players' ids, first and last names, and multiplicities from
specified seasons}
\usage{
get_players(start_year = 1917, end_year = 2025)
}
\arguments{
\item{start_year}{integer Year to start search}

\item{end_year}{integer Year to end search}
}
\value{
dataframe with one row per player
}
\description{
Get all players' ids, first and last names, and multiplicities from
specified seasons
}
