% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GameOld.R
\name{get_scoreboards}
\alias{get_scoreboards}
\title{Get the scoreboard(s) of the game(s) for a date}
\usage{
get_scoreboards(date = "2025-01-01")
}
\description{
\code{get_scoreboards()} is deprecated. Use \code{\link[=ns_scores]{ns_scores()}} instead.
}
