% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GameOld.R
\name{get_scores}
\alias{get_scores}
\title{Access the scores for a date}
\usage{
get_scores(date = "now")
}
\arguments{
\item{date}{character in 'YYYY-MM-DD' (e.g., '2025-01-01'); see
\code{\link[=seasons]{seasons()}} for reference}
}
\value{
data.frame with one row per game
}
\description{
\code{get_scores()} is deprecated. Use \code{\link[=scores]{scores()}} instead.
}
