% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OtherOld.R
\name{get_season_now}
\alias{get_season_now}
\title{Get the real-time season}
\usage{
get_season_now()
}
\description{
\code{get_season_now()} is defunct. Use \code{\link[=ns_season]{ns_season()}} and/or \code{\link[=ns_game_type]{ns_game_type()}}
instead.
}
