% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Playoff.R
\name{get_series}
\alias{get_series}
\title{Get series by season and round}
\usage{
get_series(season = get_season_now()$seasonId, round = 1)
}
\arguments{
\item{season}{integer in YYYYYYYY}

\item{round}{integer of 1:4}
}
\value{
tibble with one row per series
}
\description{
\code{get_series()} retrieves information on each series for a given set of \code{season} and \code{round}, including but not limited to their label, 1-letter code, top and bottom seeds, and winning and losing teams' IDs. Access \code{get_seasons()} for \code{season} reference.
}
\examples{
CF_series_20242025 <- get_series(season=20242025, round=3)
}
