% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Playoff.R
\name{get_series_carousel}
\alias{get_series_carousel}
\title{Get playoff series carousel by season and round}
\usage{
get_series_carousel(season = get_season_now()$seasonId, round = 1)
}
\arguments{
\item{season}{integer Season in YYYYYYYY}

\item{round}{integer Round of 1:4}
}
\value{
tibble with one row per match-up
}
\description{
Get playoff series carousel by season and round
}
\examples{
carousel_20242025_2 <- get_series_carousel(season=20242025, round=2)
}
