% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GameOld.R
\name{get_shift_charts}
\alias{get_shift_charts}
\title{Get the shifts of a game}
\usage{
get_shift_charts(game = 2025020275)
}
\description{
\code{get_shift_charts()} is deprecated. Use \code{\link[=ns_shifts]{ns_shifts()}} instead.
}
