% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SkaterOld.R
\name{get_skater_leaders}
\alias{get_skater_leaders}
\title{Get the skater statistics leaders for a season, game type, and category}
\usage{
get_skater_leaders(season = "current", game_type = "", category = "points")
}
\description{
\code{get_skater_leaders()} is deprecated. Use \code{\link[=skater_leaders]{skater_leaders()}} instead.
}
