% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SkaterOld.R
\name{get_skater_statistics}
\alias{get_skater_statistics}
\title{Get skater statistics}
\usage{
get_skater_statistics(
  season = 20242025,
  teams = 1:100,
  game_types = 1:3,
  dates = c("2025-01-01"),
  report = "summary",
  is_aggregate = FALSE,
  is_game = FALSE
)
}
\description{
\code{get_skater_statistics()} is defunct. Use \code{\link[=ns_skater_statistics]{ns_skater_statistics()}} instead.
}
