% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SkaterOld.R
\name{get_skaters}
\alias{get_skaters}
\title{Access all the skaters for a range of seasons}
\usage{
get_skaters()
}
\description{
\code{get_skaters()} is defunct. Use \code{\link[=players]{players()}} instead.
}
