% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Players.R
\name{get_skaters_old}
\alias{get_skaters_old}
\title{OLD: Get all skaters' ids, first and last names, and multiplicities from
specified seasons (who have at least scored 1 point)}
\usage{
get_skaters_old(start_year = 1917, end_year = 2025)
}
\arguments{
\item{start_year}{integer Year to start search}

\item{end_year}{integer Year to end search}
}
\value{
dataframe with one row per player
}
\description{
OLD: Get all skaters' ids, first and last names, and multiplicities from
specified seasons (who have at least scored 1 point)
}
\keyword{internal}
