% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlayerOld.R
\name{get_spotlight_players}
\alias{get_spotlight_players}
\title{Access the spotlight players}
\usage{
get_spotlight_players()
}
\value{
data.frame with one row per player
}
\description{
\code{get_spotlight_players()} is deprecated. Use \code{\link[=spotlight_players]{spotlight_players()}}
instead.
}
