% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/League.R
\name{get_standings_information}
\alias{get_standings_information}
\title{Get standings information for all seasons}
\usage{
get_standings_information()
}
\value{
tibble with one row per season
}
\description{
Get standings information for all seasons
}
\examples{
standings_info <- get_standings_information()
}
