% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Other.R
\name{get_streams}
\alias{get_streams}
\title{Get all streams}
\usage{
get_streams()
}
\value{
tibble with one row per stream
}
\description{
\code{get_streams()} retrieves information on each stream, including but not limited to their ID, name, and URL.
}
\examples{
all_streams <- get_streams()
}
