% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TeamOld.R
\name{get_team_prospects}
\alias{get_team_prospects}
\title{Get the prospects of a team for a position}
\usage{
get_team_prospects(team = 1, player_type = "forwards")
}
\description{
\code{get_team_prospects()} is deprecated. Use \code{\link[=ns_team_prospects]{ns_team_prospects()}} instead.
}
