% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Team.R
\name{get_team_scoreboard}
\alias{get_team_scoreboard}
\title{Get team scoreboard as of now}
\usage{
get_team_scoreboard(team = "BOS")
}
\arguments{
\item{team}{string 3-letter Code}
}
\value{
tibble with one row per game
}
\description{
\code{get_team_scoreboard()} retrieves information on the current game for a given \code{team}, including but not limited to their ID, season, type, start date and time, and home and visiting teams' IDs and scores. Access \code{get_teams()} for \code{team} reference.
}
\examples{
FLA_scoreboard_now <- get_team_scoreboard(team='FLA')
}
