% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Team.R
\name{get_teams}
\alias{get_teams}
\title{Get all teams}
\usage{
get_teams()
}
\value{
tibble with one row per team
}
\description{
\code{get_teams()} retrieves information on each team, including but not limited to their ID, name, and 3-letter code.
}
\examples{
all_teams <- get_teams()
}
