% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/League.R
\name{get_transactions}
\alias{get_transactions}
\title{Get transactions by season}
\usage{
get_transactions(season = get_season_now()$seasonId)
}
\arguments{
\item{season}{integer Season in YYYYYYYY}
}
\value{
tibble with one row per transaction
}
\description{
Get transactions by season
}
\examples{
transactions_20242025 <- get_transactions(20242025)
}
