% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OtherOld.R
\name{get_tv_schedule}
\alias{get_tv_schedule}
\title{Get the TV schedule of the NHL Network for a date}
\usage{
get_tv_schedule(date = "now")
}
\description{
\code{get_tv_schedule()} is deprecated. Use \code{\link[=ns_tv_schedule]{ns_tv_schedule()}} instead.
}
