% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Goalie.R
\name{goalie_career_statistics}
\alias{goalie_career_statistics}
\alias{goalie_career_stats}
\title{Get the career statistics for all the goalies}
\usage{
goalie_career_statistics()

goalie_career_stats()
}
\value{
data.frame with one row per goalie
}
\description{
\code{goalie_career_statistics()} retrieves information on ...
}
\examples{
goalie_career_stats <- goalie_career_statistics()
}
