% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GoalieEDGE.R
\name{goalie_edge_5_vs_5}
\alias{goalie_edge_5_vs_5}
\title{Get the EDGE 5 vs. 5 statistics of a goalie for a season, game type, and
report type}
\usage{
goalie_edge_5_vs_5(
  player = 8478406,
  season = "now",
  game_type = "",
  report_type = "details"
)
}
\arguments{
\item{player}{integer ID (e.g., 8478406)}

\item{season}{integer in YYYYYYYY (e.g., 20242025)}

\item{game_type}{integer in 1:3 (where 1 = pre-season, 2 = regular season, 3
= playoff/post-season) OR character of 'pre', 'regular', or 'playoff'/'post'}

\item{report_type}{character of 'd'/'details' or 'l'/'l10'/'last 10'}
}
\value{
list with four items (report_type = 'details') or data.frame with
one row per game (report_type = 'last 10')
}
\description{
\code{goalie_edge_5_vs_5()} retrieves information on each
}
\examples{
Mackenzie_Blackwood_L10_5_vs_5_regular_20242025 <- goalie_edge_5_vs_5(
  player      = 8478406,
  season      = 20242025,
  game_type   = 2,
  report_type = 'L'
 )
}
