% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Goalie.R
\name{goalie_game_scoring}
\alias{goalie_game_scoring}
\title{Access the scoring statistics for all the goalies by game}
\usage{
goalie_game_scoring()
}
\value{
data.frame with one row per player per game
}
\description{
\code{goalie_game_scoring()} retrieves the scoring statistics for all the goalies by game as a \code{data.frame} with detail on game timeline state, period/clock progression, and matchup flow, date/season filtering windows and chronological context, and team identity, affiliation, and matchup-side context.
}
\examples{
goalie_game_scoring <- goalie_game_scoring()
}
