% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Goalie.R
\name{goalie_regular_statistics}
\alias{goalie_regular_statistics}
\alias{goalie_regular_stats}
\title{Access the career regular season statistics for all the goalies}
\usage{
goalie_regular_statistics()

goalie_regular_stats()
}
\value{
data.frame with one row per goalie
}
\description{
\code{goalie_regular_statistics()} returns records-site regular-season career
goalie totals, with one row per goalie.
}
\examples{
goalie_career_regular_statistics <- goalie_regular_statistics()
}
