% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Goalie.R
\name{goalie_report_configurations}
\alias{goalie_report_configurations}
\alias{goalie_report_configs}
\title{Access the configurations for goalie reports}
\usage{
goalie_report_configurations()

goalie_report_configs()
}
\value{
list with various items
}
\description{
\code{goalie_report_configurations()} retrieves the configurations for goalie reports as a nested \code{list} that separates summary and detail blocks for situational splits across home/road, strength state, and overtime/shootout states plus configuration catalogs for valid report categories and filters.
}
\examples{
goalie_report_configs <- goalie_report_configurations()
}
