% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Goalie.R
\name{goalie_scoring}
\alias{goalie_scoring}
\title{Access the career scoring statistics for all the goalies}
\usage{
goalie_scoring()
}
\value{
data.frame with one row per player
}
\description{
\code{goalie_scoring()} scrapes the career scoring statistics for all the
goalies.
}
\examples{
goalie_scoring <- goalie_scoring()
}
