% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Goalie.R
\name{goalie_scoring_games}
\alias{goalie_scoring_games}
\title{Get all the games in which at least one goalie got at least one point}
\usage{
goalie_scoring_games()
}
\value{
data.frame with one row per goalie
}
\description{
\code{goalie_scoring_games()} retrieves information on ...
}
\examples{
goalie_scoring_games <- goalie_scoring_games()
}
