% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Goalie.R
\name{goalie_season_regular_statistics}
\alias{goalie_season_regular_statistics}
\alias{goalie_season_regular_stats}
\title{Get the regular season goalie statistics for all the seasons}
\usage{
goalie_season_regular_statistics()

goalie_season_regular_stats()
}
\value{
data.frame with one row per goalie per season
}
\description{
\code{goalie_season_regular_statistics()} retrieves information on ...
}
\examples{
goalie_season_regular_stats <- goalie_season_regular_statistics()
}
