% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Goalie.R
\name{goalie_series_statistics}
\alias{goalie_series_statistics}
\alias{goalie_series_stats}
\title{Access the playoff statistics for all the goalies by series}
\usage{
goalie_series_statistics()

goalie_series_stats()
}
\value{
data.frame with one row per player per series
}
\description{
\code{goalie_series_statistics()} retrieves the playoff statistics for all the goalies by series as a \code{data.frame} where each row represents player per series and includes detail on date/season filtering windows and chronological context, team identity, affiliation, and matchup-side context, and player identity, role, handedness, and biographical profile.
}
\examples{
goalie_series_stats <- goalie_series_statistics()
}
