% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Goalie.R
\name{goalie_statistics}
\alias{goalie_statistics}
\alias{goalie_stats}
\title{Access the career statistics for all the goalies}
\usage{
goalie_statistics()

goalie_stats()
}
\value{
data.frame with one row per player
}
\description{
\code{goalie_statistics()} returns records-site regular-season plus playoff career
goalie totals, with one row per goalie.
}
\examples{
goalie_stats <- goalie_statistics()
}
