% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Share.R
\name{ig_game_shot_location_summary}
\alias{ig_game_shot_location_summary}
\title{Save an Instagram (IG) share-able shot summary for a game}
\usage{
ig_game_shot_location_summary(game = 2023030417, team = "home", model = 1)
}
\arguments{
\item{game}{integer ID (e.g., 2025020275); see \code{\link[=games]{games()}} for reference}

\item{team}{character of 'h'/'home' or 'a'/'away'}

\item{model}{integer in 1:3 indicating which expected goals model to use
(e.g., 1); see \code{\link[=calculate_expected_goals_v1]{calculate_expected_goals_v1()}},
\code{\link[=calculate_expected_goals_v2]{calculate_expected_goals_v2()}}, and/or \code{\link[=calculate_expected_goals_v3]{calculate_expected_goals_v3()}} for
reference}
}
\value{
\code{NULL}
}
\description{
\code{ig_game_shot_location_summary()} saves an IG share-able shot summary for a game as a PNG.
}
\examples{
# Saves PNG file, so skip.
\donttest{ig_game_shot_location_summary(2023030417, model = 1, team = 'H')}
}
