% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utility.R
\name{nhl_api}
\alias{nhl_api}
\title{Call NHL API}
\usage{
nhl_api(path, query = list(), type)
}
\arguments{
\item{path}{String}

\item{query}{list}

\item{type}{integer where 1=api-web, 2=api.nhle, and 3=records.nhl}
}
\value{
parsed JSON
}
\description{
Call NHL API
}
\keyword{internal}
