% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utility.R
\name{nhl_season_to_espn_season}
\alias{nhl_season_to_espn_season}
\title{Convert NHL season to ESPN season}
\usage{
nhl_season_to_espn_season(season = get_season_now$seasonId)
}
\arguments{
\item{season}{integer Season in YYYYYYYY}
}
\value{
integer Season in YYYY
}
\description{
Convert NHL season to ESPN season
}
\keyword{internal}
