% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nhlscraper-package.R
\docType{package}
\name{nhlscraper-package}
\alias{nhlscraper}
\alias{nhlscraper-package}
\title{nhlscraper: Scraper for National Hockey League Data}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Scrapes and cleans data from the 'NHL' and 'ESPN' APIs into data.frames and lists. Wraps 125+ endpoints documented in \url{https://github.com/RentoSaijo/nhlscraper/wiki} from high-level multi-season summaries and award winners to low-level decisecond replays and bookmakers' odds, making them more accessible. Features cleaning and visualization tools, primarily for play-by-plays.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://rentosaijo.github.io/nhlscraper/}
  \item \url{https://github.com/RentoSaijo/nhlscraper}
  \item Report bugs at \url{https://github.com/RentoSaijo/nhlscraper/issues}
}

}
\author{
\strong{Maintainer}: Rento Saijo \email{rentosaijo0527@gmail.com} (\href{https://orcid.org/0009-0008-4919-7349}{ORCID}) [copyright holder]

Other contributors:
\itemize{
  \item Lars Skytte \email{lars.sunesen.skytte@gmail.com} [contributor]
  \item Jack Pallotta \email{jackjpallotta@gmail.com} (\href{https://orcid.org/0009-0004-0744-6977}{ORCID}) [contributor]
}

}
\keyword{internal}
