% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utility.R
\name{normalize_locale_names}
\alias{normalize_locale_names}
\title{Normalize locale-style dotted columns to camelCase}
\usage{
normalize_locale_names(x)
}
\arguments{
\item{x}{character vector}
}
\value{
character vector
}
\description{
Converts dotted names (e.g., firstName.default, name.cs) to camelCase and
removes trailing \code{Default} from default-language fields.
}
\keyword{internal}
