% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utility.R
\name{normalize_team_abbrev_cols}
\alias{normalize_team_abbrev_cols}
\title{Normalize team abbreviation columns to team tri-code names}
\usage{
normalize_team_abbrev_cols(x)
}
\arguments{
\item{x}{character vector}
}
\value{
character vector
}
\description{
Normalize team abbreviation columns to team tri-code names
}
\keyword{internal}
