% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utility.R
\name{normalize_team_key}
\alias{normalize_team_key}
\title{Normalize the team key}
\usage{
normalize_team_key(team)
}
\arguments{
\item{team}{integer ID (e.g., 21), character full name (e.g., 'Colorado
Avalanche'), OR three-letter code (e.g., 'COL')}
}
\value{
integer in 1:68, character full name, OR three-letter code
}
\description{
Normalize the team key
}
\keyword{internal}
