% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Award.R
\name{ns_award_winners}
\alias{ns_award_winners}
\title{Get all the award winners/finalists}
\usage{
ns_award_winners()
}
\value{
data.frame with one row per winner/finalist
}
\description{
\code{ns_award_winners()} retrieves information on each award winner or
finalist, including but not limited to their player, trophy, and season IDs;
name; and vote count.
}
\examples{
all_award_winners <- ns_award_winners()
}
