% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Award.R
\name{ns_awards}
\alias{ns_awards}
\title{Get all the awards}
\usage{
ns_awards()
}
\value{
data.frame with one row per award
}
\description{
\code{ns_awards()} retrieves information on each award, including but not
limited to their trophy ID, name, description, creation date, and image URL.
}
\examples{
all_awards <- ns_awards()
}
