% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Game.R
\name{ns_boxscore}
\alias{ns_boxscore}
\title{Access the boxscore for a game, team, and position}
\usage{
ns_boxscore(game = 2023030417, team = "home", position = "forwards")
}
\arguments{
\item{game}{integer ID (e.g., 2025020275); see \code{\link[=ns_games]{ns_games()}} for reference}

\item{team}{character of 'h'/'home' or 'a'/'away'}

\item{position}{character of 'f'/'forwards', 'd'/'defensemen', or
'g'/'goalies'}
}
\value{
data.frame with one row per player
}
\description{
\code{ns_boxscore()} scrapes the boxscore for a given set of \code{game}, \code{team}, and
\code{position}.
}
\examples{
boxscore_COL_forwards_Martin_Necas_legacy_game <- ns_boxscore(
  game     = 2025020275,
  team     = 'H',
  position = 'F'
)
}
