% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Playoff.R
\name{ns_bracket}
\alias{ns_bracket}
\title{Access the playoff bracket for a season}
\usage{
ns_bracket(season = ns_season())
}
\arguments{
\item{season}{integer in YYYYYYYY (e.g., 20242025); see \code{\link[=ns_seasons]{ns_seasons()}} for
reference}
}
\value{
data.frame with one row per series
}
\description{
\code{ns_bracket()} scrapes the playoff bracket for a given \code{season}.
}
\examples{
bracket_20242025 <- ns_bracket(season = 20242025)
}
