% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Coach.R
\name{ns_coach_career_records}
\alias{ns_coach_career_records}
\title{Get all the coaches' career records}
\usage{
ns_coach_career_records()
}
\value{
data.frame with one row per coach
}
\description{
\code{ns_coach_career_records()} retrieves information on each coach ...
}
\examples{
coach_career_records <- ns_coach_career_records()
}
