% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Coach.R
\name{ns_coach_career_statistics}
\alias{ns_coach_career_statistics}
\alias{ns_coach_career_stats}
\title{Access the career statistics for all the coaches}
\usage{
ns_coach_career_statistics()

ns_coach_career_stats()
}
\value{
data.frame with one row per coach
}
\description{
\code{ns_coach_career_statistics()} scrapes the career results for all the coaches
}
\examples{
coach_career_stats <- ns_coach_career_statistics()
}
