% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Coach.R
\name{ns_coach_franchise_records}
\alias{ns_coach_franchise_records}
\title{Get all the coaches' franchise records by game type}
\usage{
ns_coach_franchise_records()
}
\value{
data.frame with one row per coach's franchise, separated by game type
}
\description{
\code{ns_coach_franchise_records()} retrieves information on each coach ...
}
\examples{
coach_franchise_records <- ns_coach_franchise_records()
}
