% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Coach.R
\name{ns_coach_franchise_statistics}
\alias{ns_coach_franchise_statistics}
\alias{ns_coach_franchise_stats}
\title{Access the statistics for all the coaches by franchise and game type}
\usage{
ns_coach_franchise_statistics()

ns_coach_franchise_stats()
}
\value{
data.frame with one row per franchise per coach per game type
}
\description{
\code{ns_coach_franchise_statistics()} retrieves information on each coach ...
}
\examples{
coach_franchise_stats <- ns_coach_franchise_statistics()
}
