% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Coach.R
\name{ns_coaches}
\alias{ns_coaches}
\title{Access all the coaches}
\usage{
ns_coaches()
}
\value{
data.frame with one row per coach
}
\description{
\code{ns_coaches()} scrapes information on all the coaches.
}
\examples{
all_coaches <- ns_coaches()
}
