% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Draft.R
\name{ns_combine_reports}
\alias{ns_combine_reports}
\title{Get the draft combine reports}
\usage{
ns_combine_reports()
}
\value{
data.frame with one row per player
}
\description{
\code{ns_combine_reports()} retrieves information on ...
}
\examples{
combine_reports <- ns_combine_reports()
}
