% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Other.R
\name{ns_countries}
\alias{ns_countries}
\title{Access all the countries}
\usage{
ns_countries()
}
\value{
data.frame with one row per country
}
\description{
\code{ns_countries} scrapes information on all the countries.
}
\examples{
all_countries <- ns_countries()
}
