% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Draft.R
\name{ns_draft_picks}
\alias{ns_draft_picks}
\title{Access all the draft picks}
\usage{
ns_draft_picks()
}
\value{
data.frame with one row per pick
}
\description{
\code{ns_draft_picks()} scrapes information on all the draft picks.
}
\examples{
# This may take >5s, so skip.
\donttest{all_draft_picks <- ns_draft_picks()}
}
