% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Draft.R
\name{ns_drafts}
\alias{ns_drafts}
\title{Access all the drafts}
\usage{
ns_drafts()
}
\value{
data.frame with one row per draft
}
\description{
\code{ns_drafts()} scrapes information on all the drafts.
}
\examples{
all_drafts <- ns_drafts()
}
