% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Event.R
\name{ns_event_replay}
\alias{ns_event_replay}
\title{Get the replay of an event}
\usage{
ns_event_replay(game, event)
}
\value{
data.frame with one row per ...
}
\description{
\code{ns_event_replay()} retrieves ...
}
\examples{
Gabriel_Landeskog_first_regular_goal_back_replay <- ns_event_replay(
  game  = 2025020262,
  event = 751
)
}
