% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Draft.R
\name{ns_expansion_drafts}
\alias{ns_expansion_drafts}
\title{Get all the expansion drafts}
\usage{
ns_expansion_drafts()
}
\value{
data.frame with one row per pick
}
\description{
\code{ns_expansion_drafts()} retrieves information on ...
}
\examples{
all_expansion_drafts <- ns_expansion_drafts()
}
