% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Franchise.R
\name{ns_franchise_playoff_situational_results}
\alias{ns_franchise_playoff_situational_results}
\title{Access the playoff series results for all the franchises by situation}
\usage{
ns_franchise_playoff_situational_results()
}
\value{
data.frame with one row per franchise per situation
}
\description{
\code{ns_franchise_playoff_situational_results()} scrapes the playoff series
results for all the franchises by situation
}
\examples{
franchise_playoff_situational_results <- 
  ns_franchise_playoff_situational_results()
}
