% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Franchise.R
\name{ns_franchise_season_by_season}
\alias{ns_franchise_season_by_season}
\alias{ns_franchise_sbs}
\title{Get the season-by-season results for all the franchises}
\usage{
ns_franchise_season_by_season()

ns_franchise_sbs()
}
\value{
data.frame with one row per franchise's season
}
\description{
\code{ns_franchise_season_by_season()} retrieves information on each
franchise's season, including but not limited to their ID, decision, final
playoff round, and statistics.
}
\examples{
sbs_all_franchises <- ns_franchise_season_by_season()
}
