% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Franchise.R
\name{ns_franchise_season_statistics}
\alias{ns_franchise_season_statistics}
\alias{ns_franchise_season_stats}
\title{Access the statistics for all the franchises by season and game type}
\usage{
ns_franchise_season_statistics()

ns_franchise_season_stats()
}
\value{
data.frame with one row per franchise per season per game type
}
\description{
\code{ns_franchise_season_statistics()} scrapes the statistics for all the
franchises by season and game type.
}
\examples{
# This may take >5s, so skip.
\donttest{franchise_season_stats <- ns_franchise_season_statistics()}
}
