% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Franchise.R
\name{ns_franchise_team_records}
\alias{ns_franchise_team_records}
\title{Get all the franchises' teams' all-time records}
\usage{
ns_franchise_team_records()
}
\value{
data.frame with one row per franchise's team
}
\description{
\code{ns_franchise_team_records()} retrieves information on each team, including
but not limited to their ID, first and last seasons' IDs, and all-time
statistics.
}
\examples{
franchise_team_records <- ns_franchise_team_records()
}
