% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Franchise.R
\name{ns_franchise_team_statistics}
\alias{ns_franchise_team_statistics}
\alias{ns_franchise_team_stats}
\title{Get the all-time statistics for all the franchises by team and game type}
\usage{
ns_franchise_team_statistics()

ns_franchise_team_stats()
}
\value{
data.frame with one row per team per franchise per game type
}
\description{
\code{ns_franchise_team_statistics()} returns the all-time statistics for all the franchises by team and game type, including but not limited to each team's ID, wins, ties, and losses.
}
\examples{
franchise_team_statistics <- ns_franchise_team_statistics()
}
