% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Franchise.R
\name{ns_franchise_team_totals}
\alias{ns_franchise_team_totals}
\title{Get all the franchises' teams' all-time totals}
\usage{
ns_franchise_team_totals()
}
\value{
data.frame with one row per team
}
\description{
\code{ns_franchise_team_totals()} retrieves information on each team, including
but not limited to their ID, first and last seasons' IDs, and all-time
statistics.
}
\examples{
all_franchise_team_totals <- ns_franchise_team_totals()
}
