% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Franchise.R
\name{ns_franchise_versus_franchise}
\alias{ns_franchise_versus_franchise}
\alias{ns_franchise_vs_franchise}
\title{Get the all-time statistics versus other franchises for all the franchises}
\usage{
ns_franchise_versus_franchise()

ns_franchise_vs_franchise()
}
\value{
data.frame with one row per franchise per franchise
}
\description{
\code{ns_franchise_versus_franchise()} returns information on the all-time statistics versus other franchises for all the franchises, including but not limited to each franchise's ID, wins, ties, and losses.
}
\examples{
franchise_vs_franchise <- ns_franchise_versus_franchise()
}
