% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Franchise.R
\name{ns_franchises}
\alias{ns_franchises}
\title{Get all the franchises}
\usage{
ns_franchises()
}
\value{
data.frame with one row per franchise
}
\description{
\code{ns_franchises()} returns information on all the franchises, including but not limited to each franchise's ID, name, and history.
}
\examples{
all_franchises <- ns_franchises()
}
