% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Other.R
\name{ns_game_partner_odds}
\alias{ns_game_partner_odds}
\title{Get the real-time game odds for a country by partnered bookmaker}
\usage{
ns_game_partner_odds(country = "US")
}
\arguments{
\item{country}{two-letter code (e.g., 'CA'); see \code{\link[=ns_countries]{ns_countries()}} for
reference}
}
\value{
data.frame with one row per game
}
\description{
\code{ns_game_partner_odds()} scrapes the real-time game odds for a given
\code{country} by partnered bookmaker.
}
\examples{
game_partner_odds_CA <- ns_game_partner_odds(country = 'CA')
}
