% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Game.R
\name{ns_games}
\alias{ns_games}
\title{Access all the games}
\usage{
ns_games()
}
\value{
data.frame with one row per game
}
\description{
\code{ns_games()} scrapes information on all the games.
}
\examples{
# May take >5s, so skip.
\donttest{all_games <- ns_games()}
}
