% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Game.R
\name{ns_gc_play_by_play}
\alias{ns_gc_play_by_play}
\alias{ns_gc_pbp}
\title{Get the GameCenter (GC) play-by-play of a game}
\usage{
ns_gc_play_by_play(game = 2025020275)

ns_gc_pbp(game = 2025020275)
}
\arguments{
\item{game}{integer ID (e.g., 2025020275)}
}
\value{
data.frame with one row per event/play
}
\description{
\code{ns_gc_play_by_play()} retrieves GC-provided information on each play for a
given \code{game}, including but not limited to their ID; type; time of
occurrence; winning, losing, blocking, shooting, hitting, hit, scoring,
assisting, committed-by, drawn-by, and/or served-by player IDs; and X and Y
coordinates. Access \code{get_games()} for \code{game} reference.
}
\examples{
gc_pbp_SCF_game_7_20232024 <- ns_gc_play_by_play(game = 2023030417)
}
