% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Game.R
\name{ns_gc_summary}
\alias{ns_gc_summary}
\title{Access the GameCenter (GC) summary for a game}
\usage{
ns_gc_summary(game = 2023030417)
}
\arguments{
\item{game}{integer ID (e.g., 2025020275); see \code{\link[=ns_games]{ns_games()}} for reference}
}
\value{
list of various items
}
\description{
\code{ns_gc_summary()} scrapes the GC summary for a given \code{game}.
}
\examples{
gc_summary_Martin_Necas_legacy_game <- ns_gc_summary(game = 2025020275)
}
