% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Manager.R
\name{ns_general_managers}
\alias{ns_general_managers}
\alias{ns_gms}
\title{Access all the general managers}
\usage{
ns_general_managers()

ns_gms()
}
\value{
data.frame with one row per general manager
}
\description{
\code{ns_general_managers()} scrapes information on all the general managers.
}
\examples{
all_GMs <- ns_general_managers()
}
