% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Goalie.R
\name{ns_goalie_career_scoring}
\alias{ns_goalie_career_scoring}
\title{Get the career scoring totals for all the goalies}
\usage{
ns_goalie_career_scoring()
}
\value{
data.frame with one row per goalie
}
\description{
\code{ns_goalie_career_scoring()} retrieves information on ...
}
\examples{
goalie_career_scoring <- ns_goalie_career_scoring()
}
