% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Goalie.R
\name{ns_goalie_career_statistics}
\alias{ns_goalie_career_statistics}
\alias{ns_goalie_career_stats}
\title{Get the career statistics for all the goalies}
\usage{
ns_goalie_career_statistics()

ns_goalie_career_stats()
}
\value{
data.frame with one row per goalie
}
\description{
\code{ns_goalie_career_statistics()} retrieves information on ...
}
\examples{
goalie_career_stats <- ns_goalie_career_statistics()
}
