% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Goalie.R
\name{ns_goalie_game_statistics}
\alias{ns_goalie_game_statistics}
\alias{ns_goalie_game_stats}
\title{Get the goalie statistics for all the games}
\usage{
ns_goalie_game_statistics()

ns_goalie_game_stats()
}
\value{
data.frame with one row per goalie per game
}
\description{
\code{ns_goalie_game_statistics()} retrieves information on ...
}
\examples{
goalie_game_stats <- ns_goalie_game_statistics()
}
