% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Goalie.R
\name{ns_goalie_leaders}
\alias{ns_goalie_leaders}
\title{Get goalie statistics leaders by season, game-type, and category}
\usage{
ns_goalie_leaders(season = "current", game_type = "", report_type = "wins")
}
\arguments{
\item{season}{integer in YYYYYYYY (e.g., 20242025)}

\item{game_type}{integer in 1:3 (where 1 = pre-season, 2 = regular season, 3
= playoff/post-season) OR character of 'pre', 'regular', or 'playoff'/'post'}

\item{report_type}{string of 'w'/'wins', 's'/shutouts',
's\%'/'sP'/'save \%'/'save percentage', or 'gaa'/'goals against average'}
}
\value{
data.frame with one row per goalie
}
\description{
\code{get_goalie_leaders()} retrieves information on each goalie for a given set of \code{season}, \code{game_type}, and \code{category}, including but not limited to their ID, name, and statistics. Access \code{get_seasons()} for \code{season} reference.
}
\examples{
gaa_leaders_regular_20242025 <- ns_goalie_leaders(
  season      = 20242025,
  game_type   = 2,
  report_type = 'goals against average'
)
}
