% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Goalie.R
\name{ns_goalie_season_regular_statistics}
\alias{ns_goalie_season_regular_statistics}
\alias{ns_goalie_season_regular_stats}
\title{Get the regular season goalie statistics for all the seasons}
\usage{
ns_goalie_season_regular_statistics()

ns_goalie_season_regular_stats()
}
\value{
data.frame with one row per goalie per season
}
\description{
\code{ns_goalie_season_regular_statistics()} retrieves information on ...
}
\examples{
goalie_season_regular_stats <- ns_goalie_season_regular_statistics()
}
