% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Goalie.R
\name{ns_goalie_series_statistics}
\alias{ns_goalie_series_statistics}
\alias{ns_goalie_series_stats}
\title{Get the statistics for all the goalies by playoff series}
\usage{
ns_goalie_series_statistics()

ns_goalie_series_stats()
}
\value{
data.frame with one row per goalie per series
}
\description{
\code{ns_goalie_series_statistics()} retrieves information on ...
}
\examples{
goalie_series_stats <- ns_goalie_series_statistics()
}
