% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Other.R
\name{ns_location}
\alias{ns_location}
\title{Get the location of a zip code}
\usage{
ns_location(zip = 80204)
}
\arguments{
\item{zip}{integer (e.g., 80204)}
}
\value{
data.frame with one row per team
}
\description{
\code{ns_location()} retrieves ...
}
\examples{
Cranbrook_Schools <- ns_location(48304)
}
