% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Official.R
\name{ns_officials}
\alias{ns_officials}
\title{Access all the officials}
\usage{
ns_officials()
}
\value{
data.frame with one row per official
}
\description{
\code{ns_officials()} scrapes information on all the officials.
}
\examples{
all_officials <- ns_officials()
}
