% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Other.R
\name{ns_partner_odds}
\alias{ns_partner_odds}
\title{Get the partnered bookmakers' real-time game odds}
\usage{
ns_partner_odds(country = "US")
}
\arguments{
\item{country}{two-letter code}
}
\value{
data.frame with one row per game
}
\description{
\code{ns_partner_odds()} retrieves partner-provided information on each game for
a given \code{country}, including but not limited to their ID and home and away
team odds. Access \code{get_countries()} for \code{country} reference.
}
\examples{
partner_odds_CA <- ns_partner_odds(country = 'CA')
}
