% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Player.R
\name{ns_players}
\alias{ns_players}
\title{Get all the players}
\usage{
ns_players()
}
\value{
data.frame with one row per player
}
\description{
\code{ns_players()} returns all the players, including but not limited to each player's ID, name, and bio-metrics.
}
\examples{
# This may take >5s, so skip.
\donttest{all_players <- ns_players()}
}
