% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Playoff.R
\name{ns_playoff_season_statistics}
\alias{ns_playoff_season_statistics}
\alias{ns_playoff_season_stats}
\title{Get the playoff statistics for all the seasons}
\usage{
ns_playoff_season_statistics()

ns_playoff_season_stats()
}
\value{
data.frame with one row per season
}
\description{
\code{ns_playoff_season_statistics()} retrieves information on ...
}
\examples{
playoff_season_stats <- ns_playoff_season_statistics()
}
