% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/League.R
\name{ns_seasons}
\alias{ns_seasons}
\title{Get all the seasons}
\usage{
ns_seasons()
}
\value{
data.frame with one row per season
}
\description{
\code{ns_seasons()} returns information about all the seasons, including but not limited to each season's ID, start & end dates, and rules.
}
\examples{
all_seasons <- ns_seasons()
}
