% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Playoff.R
\name{ns_series}
\alias{ns_series}
\title{Access all the playoff series by game}
\usage{
ns_series()
}
\value{
data.frame with one row per game per series
}
\description{
\code{ns_series()} scrapes information on all the playoff series by game.
}
\examples{
# This may take >5s, so skip.
\donttest{all_series <- ns_series()}
}
