% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Playoff.R
\name{ns_series_schedule}
\alias{ns_series_schedule}
\title{Get the schedule of a playoff series for a season}
\usage{
ns_series_schedule(season = ns_season()$seasonId, series = "a")
}
\arguments{
\item{season}{integer in YYYYYYYY (e.g., 20242025)}

\item{series}{one-letter code (e.g., 'F')}
}
\value{
data.frame with one row per game
}
\description{
\code{ns_series_schedule()} retrieves information on each game for a given set of
\code{season} and \code{series}, including but not limited to their ID; venue; start
date and time; and home and away teams' IDs, names, and scores. Access
\code{get_seasons()} for \code{season} and \code{get_bracket()} for \code{series} references.
}
\examples{
COL_DAL_schedule_20242025 <- ns_series_schedule(
  season = 20242025, 
  series = 'F'
)
}
