% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Playoff.R
\name{ns_series_situational_records}
\alias{ns_series_situational_records}
\title{Get all the playoff series records by situation}
\usage{
ns_series_situational_records()
}
\value{
data.frame with one row per series situation
}
\description{
\code{ns_series_situational_records()} retrieves ...
}
\examples{
series_situational_records <- ns_series_situational_records()
}
