% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Game.R
\name{ns_shifts}
\alias{ns_shifts}
\title{Get the shifts of a game}
\usage{
ns_shifts(game = 2025020275)
}
\arguments{
\item{game}{integer ID (e.g., 2025020275)}
}
\value{
data.frame with one row per shift
}
\description{
\code{ns_shifts()} retrieves information on each shift for a given \code{game},
including but not limited to their period, start and end times, and player's
ID and name. Access \code{get_games()} for \code{game} reference.
}
\examples{
shifts_SCF_game_7_20232024 <- ns_shifts(game = 2023030417)
}
