% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Skater.R
\name{ns_skater_career_playoff_statistics}
\alias{ns_skater_career_playoff_statistics}
\alias{ns_skater_career_playoff_stats}
\title{Get the career playoff statistics for all the skaters}
\usage{
ns_skater_career_playoff_statistics()

ns_skater_career_playoff_stats()
}
\value{
data.frame with one row per skater
}
\description{
\code{ns_skater_career_playoff_statistics()} retrieves information on ...
}
\examples{
skater_career_playoff_statistics <- ns_skater_career_playoff_statistics()
}
