% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Skater.R
\name{ns_skater_career_regular_statistics}
\alias{ns_skater_career_regular_statistics}
\alias{ns_skater_career_regular_stats}
\title{Get the career regular season statistics for all the skaters}
\usage{
ns_skater_career_regular_statistics()

ns_skater_career_regular_stats()
}
\value{
data.frame with one row per skater
}
\description{
\code{ns_skater_career_regular_statistics()} retrieves information on ...
}
\examples{
skater_career_regular_statistics <- ns_skater_career_regular_statistics()
}
