% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Skater.R
\name{ns_skater_career_statistics}
\alias{ns_skater_career_statistics}
\alias{ns_skater_career_stats}
\title{Get the career statistics for all the skaters}
\usage{
ns_skater_career_statistics()

ns_skater_career_stats()
}
\value{
data.frame with one row per skater
}
\description{
\code{ns_skater_career_statistics()} retrieves information on ...
}
\examples{
skater_career_stats <- ns_skater_career_statistics()
}
