% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SkaterEDGE.R
\name{ns_skater_edge_skating_distance}
\alias{ns_skater_edge_skating_distance}
\title{Get the EDGE skating distance statistics of a skater for a season, game
type, and report type}
\usage{
ns_skater_edge_skating_distance(
  player = 8480039,
  season = "now",
  game_type = "",
  report_type = "details"
)
}
\arguments{
\item{player}{integer ID (e.g., 8480039)}

\item{season}{integer in YYYYYYYY (e.g., 20242025)}

\item{game_type}{integer in 1:3 (where 1 = pre-season, 2 = regular season, 3
= playoff/post-season) OR character of 'pre', 'regular', or 'playoff'/'post'}

\item{report_type}{character of 'd'/'details' or 'l'/'l10'/'last 10'}
}
\value{
data.frame with one row per strength state (report_type = 'details')
or game (report_type = 'last 10')
}
\description{
\code{ns_skater_edge_skating_distance()} retrieves information on each
}
\examples{
Martin_Necas_L10_skating_distance_regular_20242025 <- 
  ns_skater_edge_skating_distance(
    player      = 8480039,
    season      = 20242025,
    game_type   = 2,
    report_type = 'L'
  )
}
