% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Skater.R
\name{ns_skater_milestones}
\alias{ns_skater_milestones}
\title{Get the skaters on milestone watch}
\usage{
ns_skater_milestones()
}
\value{
data.frame with one row per skater
}
\description{
\code{ns_skater_milestones()} retrieves information on each skater close to a
milestone, including but not limited to their ID, name, and statistics.
}
\examples{
skater_milestones <- ns_skater_milestones()
}
