% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Skater.R
\name{ns_skater_season_statistics}
\alias{ns_skater_season_statistics}
\alias{ns_skater_season_stats}
\title{Get the statistics for all the skaters by season, game type, and team}
\usage{
ns_skater_season_statistics()

ns_skater_season_stats()
}
\value{
data.frame with one row per skater per game type per season,
separated by team when necessary
}
\description{
\code{ns_skater_season_statistics()} retrieves information on ...
}
\examples{
skater_season_stats <- ns_skater_season_statistics()
}
