% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Skater.R
\name{ns_skater_series_statistics}
\alias{ns_skater_series_statistics}
\alias{ns_skater_series_stats}
\title{Get the statistics for all the skaters by playoff series}
\usage{
ns_skater_series_statistics()

ns_skater_series_stats()
}
\value{
data.frame with one row per skater per series
}
\description{
\code{ns_skater_series_statistics()} retrieves information on ...
}
\examples{
skater_series_stats <- ns_skater_series_statistics()
}
