% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/League.R
\name{ns_standings}
\alias{ns_standings}
\title{Get the standings for a date}
\usage{
ns_standings(date = "now")
}
\arguments{
\item{date}{character in 'YYYY-MM-DD' (e.g., '2025-01-01')}
}
\value{
data.frame with one row per team
}
\description{
\code{ns_standings()} retrieves information on each team for a given \code{date},
including but not limited to their ID; name; conference; division; season,
recent, and home-away statistics; and waiver sequence. Access \code{ns_seasons()}
for \code{date} reference.
}
\examples{
standings_Halloween_2025 <- ns_standings(date = '2025-10-31')
}
