% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/League.R
\name{ns_standings_information}
\alias{ns_standings_information}
\alias{ns_standings_info}
\title{Get information about the standings for all the seasons}
\usage{
ns_standings_information()

ns_standings_info()
}
\value{
data.frame with one row per season
}
\description{
\code{ns_standings_information()} retrieves information on each season, including
but not limited to their ID; start and end dates for standings; and
conference-division, win-tie-loss, and wildcard regulations.
}
\examples{
standings_info <- ns_standings_information()
}
