% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/League.R
\name{ns_standings_rules}
\alias{ns_standings_rules}
\title{Access the standings rules by season}
\usage{
ns_standings_rules()
}
\value{
data.frame with one row per season
}
\description{
\code{ns_standings_rules()} scrapes the standings rules by season.
}
\examples{
standings_rules <- ns_standings_rules()
}
