% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Other.R
\name{ns_statistics_configuration}
\alias{ns_statistics_configuration}
\alias{ns_stats_config}
\title{Get the configuration for skater, goalie, and team statistics}
\usage{
ns_statistics_configuration()

ns_stats_config()
}
\value{
list with 5 items
}
\description{
\code{ns_statistics_configuration()} retrieves information on the outputs of the
possible combinations of inputs for \code{get_team_statistics()},
\code{get_skater_statistics()}, and \code{get_goalie_statistics()}.
}
\examples{
stats_config <- ns_statistics_configuration()
}
