% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Other.R
\name{ns_statistics_report_configuration}
\alias{ns_statistics_report_configuration}
\alias{ns_stats_report_config}
\title{Access the configurations for skater, goalie, and team statistics reports}
\usage{
ns_statistics_report_configuration()

ns_stats_report_config()
}
\value{
list with 5 items
}
\description{
\code{ns_statistics_report_configuration()} scrapes the configurations for
\code{\link[=ns_team_season_report]{ns_team_season_report()}}, \code{\link[=ns_team_game_report]{ns_team_game_report()}},
\code{\link[=ns_skater_season_report]{ns_skater_season_report()}}, \code{\link[=ns_skater_game_report]{ns_skater_game_report()}},
\code{\link[=ns_goalie_season_report]{ns_goalie_season_report()}}, and \code{\link[=ns_goalie_game_report]{ns_goalie_game_report()}}.
}
\examples{
stats_report_config <- ns_statistics_report_configuration()
}
