% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Other.R
\name{ns_streams}
\alias{ns_streams}
\title{Access all the streams}
\usage{
ns_streams()
}
\value{
data.frame with one row per stream
}
\description{
\code{ns_streams()} scrapes information on all the streams.
}
\examples{
all_streams <- ns_streams()
}
