% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TeamEDGE.R
\name{ns_team_edge_leaders}
\alias{ns_team_edge_leaders}
\title{Access the team EDGE statistics leaders for a season and game type}
\usage{
ns_team_edge_leaders(season = "now", game_type = "")
}
\arguments{
\item{season}{integer in YYYYYYYY (e.g., 20242025); see
\code{\link[=ns_team_edge_seasons]{ns_team_edge_seasons()}} for reference}

\item{game_type}{integer in 1:3 (where 1 = pre-season, 2 = regular season, 3
= playoff/post-season) OR character of 'pre', 'regular', or
'playoff'/'post'; see \code{\link[=ns_team_edge_seasons]{ns_team_edge_seasons()}} for reference}
}
\value{
list of various items
}
\description{
\code{ns_team_edge_leaders()} scrapes the team EDGE statistics leaders for a
given set of \code{season} and \code{game_type}.
}
\examples{
team_EDGE_leaders_regular_20242025 <- ns_team_edge_leaders(
  season    = 20242025,
  game_type = 2
)
}
