% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TeamEDGE.R
\name{ns_team_edge_seasons}
\alias{ns_team_edge_seasons}
\title{Access the season(s) and game type(s) in which there exists team EDGE
statistics}
\usage{
ns_team_edge_seasons()
}
\value{
data.frame with one row per season
}
\description{
\code{ns_team_edge_seasons} scrapes the season(s) and game type(s) in which the
NHL recorded team EDGE statistics.
}
\examples{
team_EDGE_seasons <- ns_team_edge_seasons()
}
