% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TeamEDGE.R
\name{ns_team_edge_shot_location}
\alias{ns_team_edge_shot_location}
\title{Get the EDGE shot location statistics for a team, season, game type, and
report type}
\usage{
ns_team_edge_shot_location(
  team = 1,
  season = "now",
  game_type = "",
  report_type = "details"
)
}
\arguments{
\item{team}{integer ID (e.g., 21), character full name (e.g., 'Colorado
Avalanche'), OR three-letter code (e.g., 'COL')}

\item{season}{integer in YYYYYYYY (e.g., 20242025)}

\item{game_type}{integer in 1:3 (where 1 = pre-season, 2 = regular season, 3
= playoff/post-season) OR character of 'pre', 'regular', or 'playoff'/'post'}

\item{report_type}{character of 'd'/details' or 't'/'totals'}
}
\value{
data.frame with one row per location (report_type = 'details') or
combination of strength state and position (report_type = 'totals')
}
\description{
\code{ns_team_edge_shot_location()} returns the EDGE shot location statistics for a given set of \code{team}, \code{season}, \code{game_type}, and \code{report_type}. Use \code{\link[=ns_teams]{ns_teams()}} for \code{team} reference and \code{\link[=ns_team_edge_seasons]{ns_team_edge_seasons()}} for \code{season} & \code{game_type} references.
}
\examples{
COL_shot_location_totals_regular_20242025 <- ns_team_edge_shot_location(
  team        = 21,
  season      = 20242025,
  game_type   = 2,
  report_type = 'T'
)
}
