% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EDGETeam.R
\name{ns_team_edge_shot_location_top_10}
\alias{ns_team_edge_shot_location_top_10}
\title{Get top 10 EDGE shot location teams by season, game type, position, report
type, and location}
\usage{
ns_team_edge_shot_location_top_10(
  season = "now",
  game_type = "",
  position = "all",
  location = "all",
  report_type = "sog"
)
}
\arguments{
\item{season}{integer in YYYYYYYY (e.g., 20242025)}

\item{game_type}{integer in 1:3 (where 1 = pre-season, 2 = regular season, 3
= playoff/post-season) OR character of 'pre', 'regular', or 'playoff'/'post'}

\item{position}{character of 'a'/'all', 'f'/'forward'/forwards', or
'd'/'defense'/defensemen'}

\item{location}{character 'a'/'all', 'h'/'high'/'high-danger'/'high danger',
'm'/'mid'/'mid-range'/'mid range', or 'l'/'long'/'long-range'/'long range'}

\item{report_type}{character of
'shot'/'shots'/'sog'/'shots-on-goal'/'shots on goal', 'goal'/'goals', or
'shooting\%'/'shootingP'/'shooting-percentage'/'shooting percentage'}
}
\value{
data.frame with one row per team
}
\description{
\code{ns_team_edge_shot_location_top_10()} retrieves information on each
}
\examples{
top_10_defensemen_shootingP_high_danger_regular_20242025 <- 
  ns_team_edge_shot_location_top_10(
    season      = 20242025,
    game_type   = 2,
    position    = 'defensemen',
    location    = 'high danger',
    report_type = 'shooting\%'
  )
}
