% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EDGETeam.R
\name{ns_team_edge_skating_distance_top_10}
\alias{ns_team_edge_skating_distance_top_10}
\title{Get top 10 EDGE skating distance teams by season, game type, position,
strength, and report type}
\usage{
ns_team_edge_skating_distance_top_10(
  season = "now",
  game_type = "",
  position = "all",
  strength = "all",
  report_type = "total"
)
}
\arguments{
\item{season}{integer in YYYYYYYY (e.g., 20242025)}

\item{game_type}{integer in 1:3 (where 1 = pre-season, 2 = regular season, 3
= playoff/post-season) OR character of 'pre', 'regular', or 'playoff'/'post'}

\item{position}{character of 'a'/'all', 'f'/'forward'/forwards', or
'd'/'defense'/defensemen'}

\item{strength}{character of 'a'/'all',
'pp'/'powerplay'/'power-play'/'power play', or
'pk'/'penaltykill'/'penalty-kill'/'penalty kill'}

\item{report_type}{character of 't'/'total'/'totals',
'a'/'average'/'per60'/'per-60'/'per 60',
'g'/'game'/'most-in-game'/'most in game', or
'p'/'period'/'most-in-period'/'most in period'}
}
\value{
data.frame with one row per team
}
\description{
\code{ns_team_edge_skating_distance_top_10()} retrieves information on each
}
\examples{
top_10_defensemen_avg_skating_distance_per_60_pp_regular_20242025 <- 
  ns_team_edge_skating_distance_top_10(
    season      = 20242025,
    game_type   = 2,
    position    = 'defensemen',
    strength    = 'pp',
    report_type = 'per 60'
  )
}
