% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TeamEDGE.R
\name{ns_team_edge_summary}
\alias{ns_team_edge_summary}
\title{Get the EDGE summary of a team for a season and game type}
\usage{
ns_team_edge_summary(team = 1, season = "now", game_type = "")
}
\arguments{
\item{team}{integer ID (e.g., 21), character full name (e.g., 'Colorado
Avalanche'), OR three-letter code (e.g., 'COL')}

\item{season}{integer in YYYYYYYY (e.g., 20242025)}

\item{game_type}{integer in 1:3 (where 1 = pre-season, 2 = regular season, 3
= playoff/post-season) OR character of 'pre', 'regular', or 'playoff'/'post'}
}
\value{
list of various items
}
\description{
\code{ns_team_edge_summary()} retrieves information on each
}
\examples{
COL_EDGE_summary_regular_20242025 <- ns_team_edge_summary(
  team      = 21, 
  season    = 20242025,
  game_type = 2
)
}
