% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Team.R
\name{ns_team_logos}
\alias{ns_team_logos}
\title{Access all the team logos}
\usage{
ns_team_logos()
}
\value{
data.frame with one row per logo
}
\description{
\code{ns_team_logos()} scrapes information on all the team logos.
}
\examples{
all_team_logos <- ns_team_logos()
}
