% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Team.R
\name{ns_team_month_schedule}
\alias{ns_team_month_schedule}
\title{Get the schedule for a team and month}
\usage{
ns_team_month_schedule(team = 1, month = "now")
}
\arguments{
\item{team}{integer ID (e.g., 21), character full name (e.g., 'Colorado
Avalanche'), OR three-letter code (e.g., 'COL')}

\item{month}{character in 'YYYY-MM' (e.g., '2025-01')}
}
\value{
data.frame with one row per game
}
\description{
\code{ns_team_month_schedule()} returns the schedule for a given set of \code{team} and \code{month}, including but not limited to each game's ID, competing teams, start time, and venue. Use \code{\link[=ns_teams]{ns_teams()}} for \code{team} reference and \code{\link[=ns_seasons]{ns_seasons()}} for \code{month} reference.
}
\examples{
COL_schedule_December_2025 <- ns_team_month_schedule(
  team  = 21, 
  month = '2025-12'
)
}
