% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Team.R
\name{ns_team_prospects}
\alias{ns_team_prospects}
\title{Get the prospects of a team for a position}
\usage{
ns_team_prospects(team = 1, position = "forwards")
}
\arguments{
\item{team}{integer ID (e.g., 21), character full name (e.g., 'Colorado
Avalanche'), OR three-letter code (e.g., 'COL')}

\item{position}{string of 'f'/'forwards', d'/'defensemen', or 'g'/goalies'}
}
\value{
data.frame with one row per player
}
\description{
\code{ns_team_prospects()} retrieves information on each prospect for a given set
of \code{team} and \code{position}, including but not limited to their ID, name,
bio-metrics, and birth date and location. Access \code{ns_teams()} for \code{team}
reference.
}
\examples{
COL_defensemen_prospects <- ns_team_prospects(
  team     = 21,
  position = 'defensemen'
)
}
