% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Team.R
\name{ns_team_roster}
\alias{ns_team_roster}
\title{Get roster by team, season, and position}
\usage{
ns_team_roster(team = 1, season = "current", position = "forwards")
}
\arguments{
\item{team}{integer ID (e.g., 21), character full name (e.g., 'Colorado
Avalanche'), OR three-letter code (e.g., 'COL')}

\item{season}{integer in YYYYYYYY (e.g., 20242025)}

\item{position}{string of 'f'/'forward'/'forwards',
'd'/'defense'/'defensemen', or 'g'/goalie'/goalies'}
}
\value{
data.frame with one row per player
}
\description{
\code{ns_team_roster()} retrieves information on each player for a given set of
\code{team}, \code{season}, and \code{position}, including but not limited to their ID,
name, bio-metrics, and birth date and location. Access \code{ns_teams()} for
\code{team} and \code{ns_team_seasons()} for \code{season} references.
}
\examples{
COL_defensemen_20242025 <- ns_team_roster(
  team        = 1,
  season      = 20242025,
  position = 'defensemen'
)
}
