% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Team.R
\name{ns_team_schedule_month}
\alias{ns_team_schedule_month}
\title{Get schedule by team and month}
\usage{
ns_team_schedule_month(team = 1, month = "now")
}
\arguments{
\item{team}{integer ID (e.g., 21), character full name (e.g., 'Colorado
Avalanche'), OR three-letter code (e.g., 'COL')}

\item{month}{character in 'YYYY-MM' (e.g., '2025-12')}
}
\value{
data.frame with one row per game
}
\description{
\code{ns_team_schedule_month()} retrieves information on each ...
}
\examples{
COL_schedule_December_2025 <- ns_team_schedule_month(
  team  = 21, 
  month = '2025-12'
)
}
