% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Team.R
\name{ns_team_schedule_season}
\alias{ns_team_schedule_season}
\title{Get schedule by team and season}
\usage{
ns_team_schedule_season(team = 1, season = "now")
}
\arguments{
\item{team}{integer ID (e.g., 21), character full name (e.g., 'Colorado
Avalanche'), OR three-letter code (e.g., 'COL')}

\item{season}{integer in YYYYYYYY (e.g., 20242025)}
}
\value{
data.frame with one row per game
}
\description{
\code{ns_team_schedule_season()} retrieves information on each
}
\examples{
COL_schedule_20252026 <- ns_team_schedule_season(
  team   = 21, 
  season = 20252026
)
}
