% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Team.R
\name{ns_team_schedule_week}
\alias{ns_team_schedule_week}
\title{Get the schedule of a team for a week since a date}
\usage{
ns_team_schedule_week(team = 1, date = "now")
}
\arguments{
\item{team}{integer ID (e.g., 21), character full name (e.g., 'Colorado
Avalanche'), OR three-letter code (e.g., 'COL')}

\item{date}{character in 'YYYY-MM-DD' (e.g., '2025-01-01')}
}
\value{
data.frame with one row per game
}
\description{
\code{ns_team_schedule_week()} retrieves information on each ...
}
\examples{
COL_schedule_Family_Week_2025 <- ns_team_schedule_week(
  team = 21,
  date = '2025-10-06'
)
}
