% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Team.R
\name{ns_team_season_statistics}
\alias{ns_team_season_statistics}
\alias{ns_team_season_stats}
\title{Get statistics for all the teams by season and game type}
\usage{
ns_team_season_statistics()

ns_team_season_stats()
}
\value{
data.frame with one row per team per game type per season
}
\description{
\code{ns_team_season_statistics()} returns statistics for all the teams by season and game type, including but not limited to each team's ID, wins, ties, and losses.
}
\examples{
# This may take >5s, so skip.
\donttest{team_season_statistics <- ns_team_season_statistics()}
}
