% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Team.R
\name{ns_team_seasons}
\alias{ns_team_seasons}
\title{Get season(s) for which team played in}
\usage{
ns_team_seasons(team = 1)
}
\arguments{
\item{team}{integer ID (e.g., 21), character full name (e.g., 'Colorado
Avalanche'), OR three-letter code (e.g., 'COL')}
}
\value{
data.frame with one row per season
}
\description{
\code{ns_team_seasons()} retrieves information on each season for a given \code{team},
including but not limited to their ID and game-type(s). Access \code{ns_teams()}
for \code{team} reference.
}
\examples{
COL_seasons <- ns_team_seasons(team = 21)
}
