% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Team.R
\name{ns_team_statistics}
\alias{ns_team_statistics}
\title{Get statistics by team, season, game type, and position}
\usage{
ns_team_statistics(
  team = 1,
  season = "now",
  game_type = "",
  position = "skaters"
)
}
\arguments{
\item{team}{integer ID (e.g., 21), character full name (e.g., 'Colorado
Avalanche'), OR three-letter code (e.g., 'COL')}

\item{season}{integer in YYYYYYYY (e.g., 20242025)}

\item{game_type}{integer in 1:3 (where 1 = pre-season, 2 = regular season, 3
= playoff/post-season) OR character of 'pre', 'regular', or 'playoff'/'post'}

\item{position}{string of 's'/'skater'/'skaters' or 'g'/'goalie'/goalies'}
}
\value{
data.frame with one row per player
}
\description{
\code{ns_team_statistics()} retrieves information on each player for a given set
of \code{team}, \code{season}, \code{game_type} and \code{position}, including but not
limited to their ID, name, and statistics. Access \code{ns_teams()} for \code{team}
and \code{ns_team_seasons()} for \code{season} and \code{game_type} references.
}
\examples{
COL_goalies_statistics_regular_20242025 <- ns_team_statistics(
  team      = 21,
  season    = 20242025,
  game_type = 2,
  position  = 'goalies'
)
}
