% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Team.R
\name{ns_teams}
\alias{ns_teams}
\title{Get all the teams}
\usage{
ns_teams()
}
\value{
data.frame with one row per team
}
\description{
\code{ns_teams()} returns information on all the teams, including but not limited to each team's ID, name, and three-letter code.
}
\examples{
all_teams <- ns_teams()
}
