% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Other.R
\name{ns_tv_schedule}
\alias{ns_tv_schedule}
\title{Get the TV schedule of the NHL Network for a date}
\usage{
ns_tv_schedule(date = "now")
}
\arguments{
\item{date}{character in 'YYYY-MM-DD' (e.g., '2025-01-01')}
}
\value{
data.frame with one row per program
}
\description{
\code{ns_tv_schedule()} retrieves information on each TV program for a given
\code{date}, including but not limited to their title, description, start and end
times, and broadcast status. Access \code{get_seasons()} for \code{date} reference.
}
\examples{
tv_schedule_Halloween_2025 <- ns_tv_schedule(date = '2025-10-31')
}
