% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Other.R
\name{ns_venues}
\alias{ns_venues}
\title{Get all the venues}
\usage{
ns_venues()
}
\value{
data.frame with one row per venue
}
\description{
\code{ns_venues()} retrieves information on each venue, including but not limited
to their ID, name, and location.
}
\examples{
all_venues <- ns_venues()
}
