% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Game.R
\name{ns_wsc_summary}
\alias{ns_wsc_summary}
\title{Get the World Showcase (WSC) summary of a game}
\usage{
ns_wsc_summary(game = 2025020275)
}
\arguments{
\item{game}{integer ID (e.g., 2025020275)}
}
\value{
list of various items
}
\description{
\code{ns_wsc_summary()} retrieves WSC-provided information on a \code{game}, including
but not limited to its type, venue, start time, clock, home and away teams,
and TV broadcast(s). Access \code{get_games()} for \code{game} reference.
}
\examples{
wsc_summary_SCF_game_7_20232024 <- ns_wsc_summary(game = 2023030417)
}
