% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Player.R
\name{player_game_log}
\alias{player_game_log}
\title{Access the game log for a player, season, and game type}
\usage{
player_game_log(player = 8478402, season = "now", game_type = "")
}
\arguments{
\item{player}{integer ID (e.g., 8480039); see \code{\link[=players]{players()}} for reference}

\item{season}{integer in YYYYYYYY (e.g., 20242025); see \code{\link[=seasons]{seasons()}} for
reference}

\item{game_type}{integer in 1:3 (where 1 = pre-season, 2 = regular season, 3
= playoff/post-season) OR character of 'pre', 'regular', or
playoff'/'post'; see \code{\link[=seasons]{seasons()}} for reference; most functions will NOT
support pre-season}
}
\value{
data.frame with one row per game
}
\description{
\code{player_game_log()} retrieves the game log for a player, season, and game type as a \code{data.frame} where each row represents game and includes detail on game timeline state, period/clock progression, and matchup flow plus production, workload, efficiency, and result-level performance outcomes.
}
\examples{
Martin_Necas_game_log_regular_20242025 <- player_game_log(
  player    = 8480039,
  season    = 20242025,
  game_type = 2
)
}
