% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Player.R
\name{player_seasons}
\alias{player_seasons}
\title{Access the season(s) and game type(s) in which a player played}
\usage{
player_seasons(player = 8478402)
}
\arguments{
\item{player}{integer ID (e.g., 8480039); see \code{\link[=players]{players()}} for reference}
}
\value{
data.frame with one row per season
}
\description{
\code{player_seasons()} scrapes the season(s) and game type(s) in which a
player played in the NHL.
}
\examples{
Martin_Necas_seasons <- player_seasons(player = 8480039)
}
