% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Player.R
\name{players}
\alias{players}
\title{Access all the players}
\usage{
players()
}
\value{
data.frame with one row per player
}
\description{
\code{players()} retrieves all the players as a \code{data.frame} where each row represents player and includes detail on player identity, role, handedness, and biographical profile.
}
\examples{
# May take >5s, so skip.
\donttest{all_players <- players()}
}
