% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Playoff.R
\name{playoff_season_statistics}
\alias{playoff_season_statistics}
\alias{playoff_season_stats}
\title{Access the playoff statistics by season}
\usage{
playoff_season_statistics()

playoff_season_stats()
}
\value{
data.frame with one row per season
}
\description{
\code{playoff_season_statistics()} retrieves the playoff statistics by season as a \code{data.frame} where each row represents season and includes detail on date/season filtering windows and chronological context.
}
\examples{
playoff_season_stats <- playoff_season_statistics()
}
