% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EventOld.R
\name{ps}
\alias{ps}
\title{Access all the penalty shots}
\usage{
ps()
}
\value{
data.frame with one row per penalty shot
}
\description{
\code{ps()} is deprecated. Use \code{\link[=penalty_shots]{penalty_shots()}} instead.
}
