% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Event.R
\name{replay}
\alias{replay}
\title{Access the replay for an event}
\usage{
replay(game = 2023030417, event = 866)
}
\arguments{
\item{game}{integer ID (e.g., 2025020262); see \code{\link[=games]{games()}} for reference}

\item{event}{integer ID (e.g., 751); see \code{\link[=gc_play_by_play]{gc_play_by_play()}} and/or
\code{\link[=wsc_play_by_play]{wsc_play_by_play()}} for reference; must be a 'goal' event}
}
\value{
data.frame with one row per decisecond
}
\description{
\code{replay()} scrapes the replay for a given \code{event}.
}
\examples{
Gabriel_Landeskog_first_regular_goal_back_replay <- replay(
  game  = 2025020262,
  event = 751
)
}
