% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Other.R
\name{report_configurations}
\alias{report_configurations}
\alias{report_configs}
\title{Access the configurations for team, skater, and goalie reports}
\usage{
report_configurations()

report_configs()
}
\value{
list with 5 items
}
\description{
\code{report_configurations()} scrapes the configurations for
\code{\link[=team_season_report]{team_season_report()}}, \code{\link[=team_game_report]{team_game_report()}},
\code{\link[=skater_season_report]{skater_season_report()}}, \code{\link[=skater_game_report]{skater_game_report()}},
\code{\link[=goalie_season_report]{goalie_season_report()}}, and \code{\link[=goalie_game_report]{goalie_game_report()}}.
}
\examples{
report_configs <- report_configurations()
}
