% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utils.R
\name{req_nhl}
\alias{req_nhl}
\title{Internal: call NHL API}
\usage{
req_nhl(path, query = list())
}
\arguments{
\item{path}{string API path}

\item{query}{list Query parameters}
}
\value{
parsed JSON
}
\description{
Internal: call NHL API
}
\keyword{internal}
