% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/League.R
\name{schedule}
\alias{schedule}
\title{Access the schedule for a date}
\usage{
schedule(date = Sys.Date())
}
\arguments{
\item{date}{character in 'YYYY-MM-DD' (e.g., '2025-01-01'); see
\code{\link[=seasons]{seasons()}} for reference}
}
\value{
data.frame with one row per game
}
\description{
\code{schedule()} scrapes the schedule for a given \code{date}.
}
\examples{
schedule_Halloween_2025 <- schedule(date = '2025-10-31')
}
