% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utility.R
\name{scope_person_name_cols}
\alias{scope_person_name_cols}
\title{Scope generic person-name columns to an entity}
\usage{
scope_person_name_cols(x, prefix)
}
\arguments{
\item{x}{character vector}

\item{prefix}{character scalar (e.g., 'player', 'goalie', 'skater')}
}
\value{
character vector
}
\description{
Scope generic person-name columns to an entity
}
\keyword{internal}
