% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Game.R
\name{scores}
\alias{scores}
\title{Access the scores for a date}
\usage{
scores(date = "now")
}
\arguments{
\item{date}{character in 'YYYY-MM-DD' (e.g., '2025-01-01'); see
\code{\link[=seasons]{seasons()}} for reference}
}
\value{
data.frame with one row per game
}
\description{
\code{scores()} retrieves the scores for a date as a \code{data.frame} where each row represents game and includes detail on game timeline state, period/clock progression, and matchup flow, date/season filtering windows and chronological context, and team identity, affiliation, and matchup-side context.
}
\examples{
scores_Halloween_2025 <- scores(date = '2025-10-31')
}
