% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/League.R
\name{season}
\alias{season}
\title{Access the season as of now}
\usage{
season()
}
\value{
integer in YYYYYYYY (e.g., 20242025)
}
\description{
\code{season} scrapes the current season ID.
}
\examples{
season_now <- season()
}
