% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Load.R
\name{shift_chart_summaries}
\alias{shift_chart_summaries}
\title{Access the shift chart summaries for a season}
\usage{
shift_chart_summaries(season = 20242025)
}
\arguments{
\item{season}{integer in YYYYYYYY (e.g., 20242025); see \code{\link[=seasons]{seasons()}} for
reference}
}
\value{
data.frame with one row per player per period
}
\description{
\code{shift_chart_summaries()} loads the shift chart summaries for a given
\code{season}.
}
\examples{
# May take >5s, so skip.
\donttest{shift_chart_summaries_20212022 <- shift_chart_summaries(season = 20212022)}
}
