% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GameOld.R
\name{shifts}
\alias{shifts}
\title{Access the shift charts for a game}
\usage{
shifts(game = 2023030417)
}
\arguments{
\item{game}{integer ID (e.g., 2025020275); see \code{\link[=games]{games()}} for reference}
}
\value{
data.frame with one row per shift
}
\description{
\code{shifts()} is deprecated. Use \code{\link[=shift_chart]{shift_chart()}} instead.
}
