% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Skater.R
\name{skater_career_statistics}
\alias{skater_career_statistics}
\alias{skater_career_stats}
\title{Get the career statistics for all the skaters}
\usage{
skater_career_statistics()

skater_career_stats()
}
\value{
data.frame with one row per skater
}
\description{
\code{skater_career_statistics()} retrieves information on ...
}
\examples{
skater_career_stats <- skater_career_statistics()
}
