% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SkaterEDGE.R
\name{skater_edge_seasons}
\alias{skater_edge_seasons}
\title{Access the season(s) and game type(s) in which there exists skater EDGE
statistics}
\usage{
skater_edge_seasons()
}
\value{
data.frame with one row per season
}
\description{
\code{skater_edge_seasons()} returns the seasons and game type IDs for which the
NHL EDGE skater endpoints expose data.
}
\examples{
skater_EDGE_seasons <- skater_edge_seasons()
}
