% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SkaterEDGE.R
\name{skater_edge_shot_location}
\alias{skater_edge_shot_location}
\title{Access the EDGE shot location statistics for a skater, season, game type, and
category}
\usage{
skater_edge_shot_location(
  player = 8478402,
  season = "now",
  game_type = "",
  category = "details"
)
}
\arguments{
\item{player}{integer ID (e.g., 8480039); see \code{\link[=players]{players()}} for reference}

\item{season}{integer in YYYYYYYY (e.g., 20242025); see
\code{\link[=skater_edge_seasons]{skater_edge_seasons()}} for reference}

\item{game_type}{integer in 1:3 (where 1 = pre-season, 2 = regular season, 3
= playoff/post-season) OR character of 'pre', 'regular', or
'playoff'/'post'; see \code{\link[=skater_edge_seasons]{skater_edge_seasons()}} for reference; most functions
will NOT support pre-season}

\item{category}{character of 'd'/details' or 't'/'totals'}
}
\value{
data.frame with one row per shot location
}
\description{
\code{skater_edge_shot_location()} retrieves the EDGE shot location statistics for a skater, season, game type, and category as a \code{data.frame} where each row represents shot location and includes detail on production, workload, efficiency, and result-level performance outcomes plus NHL EDGE style tracking outputs and relative-performance context.
}
\examples{
Martin_Necas_shot_location_totals_regular_20242025 <- 
  skater_edge_shot_location(
    player    = 8480039,
    season    = 20242025,
    game_type = 2,
    category  = 'T'
  )
}
