% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SkaterEDGE.R
\name{skater_edge_summary}
\alias{skater_edge_summary}
\title{Get the EDGE summary of a skater for a season and game type}
\usage{
skater_edge_summary(player = 8480039, season = "now", game_type = "")
}
\arguments{
\item{player}{integer ID (e.g., 8480039)}

\item{season}{integer in YYYYYYYY (e.g., 20242025)}

\item{game_type}{integer in 1:3 (where 1 = pre-season, 2 = regular season, 3
= playoff/post-season) OR character of 'pre', 'regular', or 'playoff'/'post'}
}
\value{
list of various items
}
\description{
\code{skater_edge_summary()} retrieves information on each
}
\examples{
Martin_Necas_EDGE_summary_regular_20242025 <- skater_edge_summary(
  player    = 8480039, 
  season    = 20242025,
  game_type = 2
)
}
