% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SkaterEDGE.R
\name{skater_edge_zone_time}
\alias{skater_edge_zone_time}
\title{Get the EDGE zone time statistics of a skater for a season, game type, and
report type}
\usage{
skater_edge_zone_time(
  player = 8480039,
  season = "now",
  game_type = "",
  report_type = "details"
)
}
\arguments{
\item{player}{integer ID (e.g., 8480039)}

\item{season}{integer in YYYYYYYY (e.g., 20242025)}

\item{game_type}{integer in 1:3 (where 1 = pre-season, 2 = regular season, 3
= playoff/post-season) OR character of 'pre', 'regular', or 'playoff'/'post'}

\item{report_type}{character of 'd'/'details' or 's'/'starts'}
}
\value{
data.frame with one row per strength state (report_type = 'details')
or list with six items (report_type = 'starts')
}
\description{
\code{skater_edge_zone_time()} retrieves information on each
}
\examples{
Martin_Necas_starts_regular_20242025 <- skater_edge_zone_time(
  player      = 8480039,
  season      = 20242025,
  game_type   = 2,
  report_type = 'S'
)
}
