% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Skater.R
\name{skater_playoff_statistics}
\alias{skater_playoff_statistics}
\alias{skater_playoff_stats}
\title{Access the career playoff statistics for all the skaters}
\usage{
skater_playoff_statistics()

skater_playoff_stats()
}
\value{
data.frame with one row per player
}
\description{
\code{skater_playoff_statistics()} scrapes the career playoff statistics for all
the skaters.
}
\examples{
skater_playoff_stats <- skater_playoff_statistics()
}
