% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Skater.R
\name{skater_regular_statistics}
\alias{skater_regular_statistics}
\alias{skater_regular_stats}
\title{Access the career regular season statistics for all the skaters}
\usage{
skater_regular_statistics()

skater_regular_stats()
}
\value{
data.frame with one row per player
}
\description{
\code{skater_regular_statistics()} retrieves the career regular season statistics for all the skaters as a \code{data.frame} where each row represents player and includes detail on team identity, affiliation, and matchup-side context, player identity, role, handedness, and biographical profile, and production, workload, efficiency, and result-level performance outcomes.
}
\examples{
skater_regular_stats <- skater_regular_statistics()
}
