% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Skater.R
\name{skater_season_statistics}
\alias{skater_season_statistics}
\alias{skater_season_stats}
\title{Access the statistics for all the skaters by season, game type, and team}
\usage{
skater_season_statistics()

skater_season_stats()
}
\value{
data.frame with one row per player per season per game type,
separated by team if applicable
}
\description{
\code{skater_season_statistics()} returns records-site player stat rows by
player, team, season, and game type, preserving separate rows when a player
changed teams.
}
\examples{
# May take >5s, so skip.
\donttest{skater_season_stats <- skater_season_statistics()}
}
