% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Skater.R
\name{skater_series_statistics}
\alias{skater_series_statistics}
\alias{skater_series_stats}
\title{Get the statistics for all the skaters by playoff series}
\usage{
skater_series_statistics()

skater_series_stats()
}
\value{
data.frame with one row per skater per series
}
\description{
\code{skater_series_statistics()} retrieves information on ...
}
\examples{
skater_series_stats <- skater_series_statistics()
}
