% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Player.R
\name{spotlight_players}
\alias{spotlight_players}
\title{Access the spotlight players}
\usage{
spotlight_players()
}
\value{
data.frame with one row per player
}
\description{
\code{spotlight_players()} scrapes the spotlight players.
}
\examples{
spotlight_players <- spotlight_players()
}
