% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/League.R
\name{standings_rules}
\alias{standings_rules}
\title{Access the standings rules by season}
\usage{
standings_rules()
}
\value{
data.frame with one row per season
}
\description{
\code{standings_rules()} scrapes the standings rules by season.
}
\examples{
standings_rules <- standings_rules()
}
