% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Other.R
\name{statistics_report_configurations}
\alias{statistics_report_configurations}
\alias{stats_report_configs}
\title{Access the configurations for skater, goalie, and team statistics reports}
\usage{
statistics_report_configurations()

stats_report_configs()
}
\value{
list with 5 items
}
\description{
\code{statistics_report_configurations()} scrapes the configurations for
\code{\link[=team_season_report]{team_season_report()}}, \code{\link[=team_game_report]{team_game_report()}},
\code{\link[=skater_season_report]{skater_season_report()}}, \code{\link[=skater_game_report]{skater_game_report()}},
\code{\link[=goalie_season_report]{goalie_season_report()}}, and \code{\link[=goalie_game_report]{goalie_game_report()}}.
}
\examples{
stats_report_config <- statistics_report_configurations()
}
