% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Clean.R
\name{strip_game_id}
\alias{strip_game_id}
\title{Strip the game ID into the season ID, game type ID, and game number for all
the events (plays) in a play-by-play}
\usage{
strip_game_id(
  data,
  game_id_name = "gameId",
  season_id_name = "seasonId",
  game_type_id_name = "gameTypeId",
  game_number_name = "gameNumber"
)
}
\arguments{
\item{data}{data.frame of play-by-play(s); see \code{\link[=gc_play_by_play]{gc_play_by_play()}} and/or
\code{\link[=wsc_play_by_play]{wsc_play_by_play()}} for reference}

\item{game_id_name}{name of column that contains game ID}

\item{season_id_name}{name of column that you want contain season ID}

\item{game_type_id_name}{name of column that you want contain game type ID}

\item{game_number_name}{name of column that you want contain game number}
}
\value{
data.frame with one row per event (play)
}
\description{
\code{strip_game_id()} strips the game ID into the season ID, game type ID, and
game number for all the events (plays) in a play-by-play.
}
\examples{
# May take >5s, so skip.
\donttest{
  test                 <- gc_play_by_play()
  test_gameId_stripped <- strip_game_id(test)
}
}
