% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Clean.R
\name{strip_time_period}
\alias{strip_time_period}
\title{Strip the timestamp and period number into the time elapsed in the period and game for all the events (plays) in a play-by-play}
\usage{
strip_time_period(play_by_play)
}
\arguments{
\item{play_by_play}{data.frame of play-by-play(s); see \code{\link[=gc_play_by_play]{gc_play_by_play()}} and/or \code{\link[=wsc_play_by_play]{wsc_play_by_play()}} for reference}
}
\value{
data.frame with one row per event (play) and added columns \code{secondsElapsedInPeriod} and \code{secondsElapsedInGame}
}
\description{
\code{strip_time_period()} strips the timestamp and period number into the time elapsed in the period and game for all the events (plays) in a play-by-play.
}
\keyword{internal}
