% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TeamEDGE.R
\name{team_edge_leaders}
\alias{team_edge_leaders}
\title{Access the team EDGE statistics leaders for a season and game type}
\usage{
team_edge_leaders(season = "now", game_type = "")
}
\arguments{
\item{season}{integer in YYYYYYYY (e.g., 20242025); see
\code{\link[=team_edge_seasons]{team_edge_seasons()}} for reference}

\item{game_type}{integer in 1:3 (where 1 = pre-season, 2 = regular season, 3
= playoff/post-season) OR character of 'pre', 'regular', or
'playoff'/'post'; see \code{\link[=team_edge_seasons]{team_edge_seasons()}} for reference; most functions
will NOT support pre-season}
}
\value{
list of various items
}
\description{
\code{team_edge_leaders()} retrieves the team EDGE statistics leaders for a season and game type as a nested \code{list} that separates summary and detail blocks for NHL EDGE style tracking outputs and relative-performance context.
}
\examples{
team_EDGE_leaders_regular_20242025 <- team_edge_leaders(
  season    = 20242025,
  game_type = 2
)
}
