% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TeamEDGE.R
\name{team_edge_zone_time}
\alias{team_edge_zone_time}
\title{Access the EDGE zone time statistics for a team, season, game type, and
category}
\usage{
team_edge_zone_time(
  team = 1,
  season = "now",
  game_type = "",
  category = "details"
)
}
\arguments{
\item{team}{integer ID (e.g., 21), character full name (e.g., 'Colorado
Avalanche'), OR three-letter code (e.g., 'COL'); see \code{\link[=teams]{teams()}} for
reference; ID is preferable as there now exists duplicate three-letter codes
(i.e., 'UTA' for 'Utah Hockey Club' and 'Utah Mammoth')}

\item{season}{integer in YYYYYYYY (e.g., 20242025); see
\code{\link[=team_edge_seasons]{team_edge_seasons()}} for reference}

\item{game_type}{integer in 1:3 (where 1 = pre-season, 2 = regular season, 3
= playoff/post-season) OR character of 'pre', 'regular', or
'playoff'/'post'; see \code{\link[=team_edge_seasons]{team_edge_seasons()}} for reference; most functions
will NOT support pre-season}

\item{category}{character of 'd'/'details' or
'dS'/'dSOG'/'dShot'/'shot differential'}
}
\value{
data.frame with one row per strength state
(category = 'details') or list with four items (category = 'shot
differential')
}
\description{
\code{team_edge_zone_time()} retrieves the EDGE zone time statistics for a team, season, game type, and category as a \code{data.frame} where each row represents strength state and includes detail on NHL EDGE style tracking outputs and relative-performance context.
}
\examples{
COL_dS_regular_20242025 <- team_edge_zone_time(
  team      = 21,
  season    = 20242025,
  game_type = 2,
  category  = 'dS'
)
}
