% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Team.R
\name{team_prospects}
\alias{team_prospects}
\title{Access the prospects for a team and position}
\usage{
team_prospects(team = 1, position = "forwards")
}
\arguments{
\item{team}{integer ID (e.g., 21), character full name (e.g., 'Colorado
Avalanche'), OR three-letter code (e.g., 'COL'); see \code{\link[=teams]{teams()}} for
reference; ID is preferable as there now exists duplicate three-letter codes
(i.e., 'UTA' for 'Utah Hockey Club' and 'Utah Mammoth')}

\item{position}{character of 'f'/'forwards', 'd'/'defensemen', or
'g'/'goalies'}
}
\value{
data.frame with one row per player
}
\description{
\code{team_prospects()} scrapes the prospects for a given set of \code{team} and
\code{position}.
}
\examples{
COL_forward_prospects <- team_prospects(
  team     = 21,
  position = 'F'
)
}
