% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Team.R
\name{team_seasons}
\alias{team_seasons}
\title{Access the season(s) and game type(s) in which a team played}
\usage{
team_seasons(team = 1)
}
\arguments{
\item{team}{integer ID (e.g., 21), character full name (e.g., 'Colorado
Avalanche'), OR three-letter code (e.g., 'COL'); see \code{\link[=teams]{teams()}} for
reference; ID is preferable as there now exists duplicate three-letter codes
(i.e., 'UTA' for 'Utah Hockey Club' and 'Utah Mammoth')}
}
\value{
data.frame with one row per season
}
\description{
\code{team_seasons()} scrapes the season(s) and game type(s) in which a team
played in the NHL.
}
\examples{
COL_seasons <- team_seasons(team = 21)
}
