% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utility.R
\name{to_game_type_id}
\alias{to_game_type_id}
\title{Convert to the appropriate game type ID}
\usage{
to_game_type_id(game_type)
}
\arguments{
\item{game_type}{integer in 1:3 (where 1 = pre-season, 2 = regular season, 3
= playoff/post-season) OR character of 'pre', 'regular', or 'playoff'/'post'}
}
\value{
integer in 1:3
}
\description{
Convert to the appropriate game type ID
}
\keyword{internal}
