% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utility.R
\name{to_team_id}
\alias{to_team_id}
\title{Convert to the appropriate team ID}
\usage{
to_team_id(team, lookup = .to_team_id)
}
\arguments{
\item{team}{integer ID (e.g., 21), character full name (e.g., 'Colorado
Avalanche'), OR three-letter code (e.g., 'COL')}
}
\value{
integer in 1:68
}
\description{
Convert to the appropriate team ID
}
\keyword{internal}
