% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utility.R
\name{to_team_tri_code}
\alias{to_team_tri_code}
\title{Convert to the appropriate team three-letter code}
\usage{
to_team_tri_code(team, lookup = .to_team_tri_code)
}
\arguments{
\item{team}{integer ID (e.g., 21), character full name (e.g., 'Colorado
Avalanche'), OR three-letter code (e.g., 'COL')}
}
\value{
three-letter code
}
\description{
Convert to the appropriate team three-letter code
}
\keyword{internal}
