% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Load.R
\name{wsc_play_by_plays}
\alias{wsc_play_by_plays}
\alias{wsc_pbps}
\title{Access the World Showcase (WSC) play-by-plays for a season}
\usage{
wsc_play_by_plays(season = 20242025)

wsc_pbps(season = 20242025)
}
\arguments{
\item{season}{integer in YYYYYYYY (e.g., 20242025); see \code{\link[=seasons]{seasons()}} for
reference}
}
\value{
data.frame with one row per event (play) per game
}
\description{
\code{wsc_play_by_plays()} loads the WSC play-by-plays for a given \code{season}.
}
\examples{
# May take >5s, so skip.
\donttest{wsc_pbps_20212022 <- wsc_play_by_plays(season = 20212022)}
}
