% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Game.R
\name{wsc_summary}
\alias{wsc_summary}
\title{Access the World Showcase (WSC) summary for a game}
\usage{
wsc_summary(game = 2023030417)
}
\arguments{
\item{game}{integer ID (e.g., 2025020275); see \code{\link[=games]{games()}} for reference}
}
\value{
list of various items
}
\description{
\code{wsc_summary()} retrieves the World Showcase (WSC) summary for a game as a nested \code{list} that separates summary and detail blocks for game timeline state, period/clock progression, and matchup flow, date/season filtering windows and chronological context, and venue/location geography and regional metadata.
}
\examples{
wsc_summary_Martin_Necas_legacy_game <- wsc_summary(game = 2025020275)
}
