% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/League.R
\name{attendance}
\alias{attendance}
\title{Access the attendance by season and game type}
\usage{
attendance()
}
\value{
data.frame with one row per season
}
\description{
\code{attendance()} scrapes the attendance by season and game type.
}
\examples{
all_attendance <- attendance()
}
