% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Award.R
\name{award_winners}
\alias{award_winners}
\title{Access all the award winners/finalists}
\usage{
award_winners()
}
\value{
data.frame with one row per winner/finalist
}
\description{
\code{award_winners()} retrieves all the award winners/finalists as a \code{data.frame} where each row represents winner/finalist and includes detail on date/season filtering windows and chronological context, team identity, affiliation, and matchup-side context, and player identity, role, handedness, and biographical profile.
}
\examples{
all_award_winners <- award_winners()
}
