% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Award.R
\name{awards}
\alias{awards}
\title{Access all the awards}
\usage{
awards()
}
\value{
data.frame with one row per award
}
\description{
\code{awards()} retrieves all the awards as a \code{data.frame} where each row represents award and includes detail on recognition, leaderboard, and milestone-watch context.
}
\examples{
all_awards <- awards()
}
