% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Playoff.R
\name{bracket}
\alias{bracket}
\title{Access the playoff bracket for a season}
\usage{
bracket(season = season_now())
}
\arguments{
\item{season}{integer in YYYYYYYY (e.g., 20242025); see \code{\link[=seasons]{seasons()}} for
reference}
}
\value{
data.frame with one row per series
}
\description{
\code{bracket()} scrapes the playoff bracket for a given \code{season}.
}
\examples{
bracket_20242025 <- bracket(season = 20242025)
}
