% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Clean.R
\name{calculate_distance}
\alias{calculate_distance}
\title{Calculate the Euclidean distance from the attacking net for all the events
(plays) in a play-by-play}
\usage{
calculate_distance(
  data,
  x_coordinate_normalized_name = "xCoordNorm",
  y_coordinate_normalized_name = "yCoordNorm",
  distance_name = "distance"
)
}
\arguments{
\item{data}{data.frame of play-by-play(s); see \code{\link[=gc_play_by_play]{gc_play_by_play()}} and/or
\code{\link[=wsc_play_by_play]{wsc_play_by_play()}} for reference}

\item{x_coordinate_normalized_name}{name of column that contains normalized
x coordinate; see \code{\link[=normalize_coordinates]{normalize_coordinates()}} for reference}

\item{y_coordinate_normalized_name}{name of column that contains normalized
y coordinate; see \code{\link[=normalize_coordinates]{normalize_coordinates()}} for reference}

\item{distance_name}{name of column that you want contain distance}
}
\value{
data.frame with one row per event (play)
}
\description{
\code{calculate_distance()} calculates the Euclidean distance from the attacking
net for all the events (plays) in a play-by-play.
}
\examples{
# May take >5s, so skip.
\donttest{
  test                     <- gc_play_by_play()
  test_is_home_flagged     <- flag_is_home(test)
  test_coords_normalized   <- normalize_coordinates(test_is_home_flagged)
  test_distance_calculated <- calculate_distance(test_coords_normalized)
}
}
