% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Clean.R
\name{calculate_speed}
\alias{calculate_speed}
\title{Calculate event-to-event deltas and speeds in normalized x/y, distance, and angle for a play-by-play}
\usage{
calculate_speed(play_by_play)
}
\arguments{
\item{play_by_play}{data.frame of play-by-play(s); see \code{\link[=gc_play_by_play]{gc_play_by_play()}}, \code{\link[=gc_play_by_plays]{gc_play_by_plays()}}, \code{\link[=wsc_play_by_play]{wsc_play_by_play()}}, or \code{\link[=wsc_play_by_plays]{wsc_play_by_plays()}} for reference; the original columns must exist}
}
\value{
data.frame with one row per event (play) and added columns: \code{dXN}, \code{dYN}, \code{dD}, \code{dA}, \code{dT}, \code{dXNdT}, \code{dYNdT}, \code{dDdT}, \code{dAdT}
}
\description{
\code{calculate_speed()} calculates event-to-event deltas and speeds in normalized x/y, distance, and angle for a play-by-play.
}
