% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Coach.R
\name{coach_career_statistics}
\alias{coach_career_statistics}
\alias{coach_career_stats}
\title{Access the career statistics for all the coaches}
\usage{
coach_career_statistics()

coach_career_stats()
}
\value{
data.frame with one row per coach
}
\description{
\code{coach_career_statistics()} retrieves the career statistics for all the coaches as a \code{data.frame} where each row represents coach and includes detail on ranking movement, points pace, and division/conference position signals.
}
\examples{
coach_career_stats <- coach_career_statistics()
}
