% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Coach.R
\name{coach_franchise_statistics}
\alias{coach_franchise_statistics}
\alias{coach_franchise_stats}
\title{Access the statistics for all the coaches by franchise and game type}
\usage{
coach_franchise_statistics()

coach_franchise_stats()
}
\value{
data.frame with one row per franchise per coach per game type
}
\description{
\code{coach_franchise_statistics()} scrapes the statistics for all the coaches by
franchise and game type.
}
\examples{
coach_franchise_stats <- coach_franchise_statistics()
}
