% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Coach.R
\name{coaches}
\alias{coaches}
\title{Access all the coaches}
\usage{
coaches()
}
\value{
data.frame with one row per coach
}
\description{
\code{coaches()} retrieves all the coaches as a \code{data.frame} where each row represents coach and includes detail on team identity, affiliation, and matchup-side context plus player identity, role, handedness, and biographical profile.
}
\examples{
all_coaches <- coaches()
}
