% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Coach.R
\name{coaches}
\alias{coaches}
\title{Access all the coaches}
\usage{
coaches()
}
\value{
data.frame with one row per coach
}
\description{
\code{coaches()} scrapes all the coaches.
}
\examples{
all_coaches <- coaches()
}
