% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Other.R
\name{countries}
\alias{countries}
\title{Access all the countries}
\usage{
countries()
}
\value{
data.frame with one row per country
}
\description{
\code{countries()} retrieves all the countries as a \code{data.frame} where each row represents country and includes detail on reference metadata, regional context, and media availability detail.
}
\examples{
all_countries <- countries()
}
