% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Other.R
\name{countries}
\alias{countries}
\title{Access all the countries}
\usage{
countries()
}
\value{
data.frame with one row per country
}
\description{
\code{countries} scrapes all the countries.
}
\examples{
all_countries <- countries()
}
