% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utility.R
\name{.nhl_request}
\alias{.nhl_request}
\title{Build an NHL API request}
\usage{
.nhl_request(path, query = list(), type)
}
\arguments{
\item{path}{character}

\item{query}{list}

\item{type}{character of 'w' for web, 's' for stats, and 'r' for records}
}
\value{
httr2 request object
}
\description{
\code{.nhl_request()} constructs an NHL API request with the standard retry policy
used throughout the package.
}
\keyword{internal}
