% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Clean.R
\name{.on_ice_timing_scalar_column_names}
\alias{.on_ice_timing_scalar_column_names}
\title{Build scalar on-ice timing column names}
\usage{
.on_ice_timing_scalar_column_names(metric_suffix)
}
\arguments{
\item{metric_suffix}{scalar timing suffix}
}
\value{
character vector of column names
}
\description{
\code{.on_ice_timing_scalar_column_names()} returns the scalar goalie and skater
timing column names for a timing metric suffix.
}
\keyword{internal}
