% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Draft.R
\name{draft_picks}
\alias{draft_picks}
\title{Access all the draft picks}
\usage{
draft_picks()
}
\value{
data.frame with one row per pick
}
\description{
\code{draft_picks()} retrieves all the draft picks as a \code{data.frame} where each row represents pick and includes detail on team identity, affiliation, and matchup-side context plus player identity, role, handedness, and biographical profile.
}
\examples{
# May take >5s, so skip.
\donttest{all_draft_picks <- draft_picks()}
}
