% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Draft.R
\name{draft_picks}
\alias{draft_picks}
\title{Access all the draft picks}
\usage{
draft_picks()
}
\value{
data.frame with one row per pick
}
\description{
\code{draft_picks()} scrapes all the draft picks.
}
\examples{
# May take >5s, so skip.
\donttest{all_draft_picks <- draft_picks()}
}
