% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Draft.R
\name{draft_prospects}
\alias{draft_prospects}
\title{Access all the draft prospects}
\usage{
draft_prospects()
}
\value{
data.frame with one row per player
}
\description{
\code{draft_prospects()} scrapes all the draft prospects.
}
\examples{
# May take >5s, so skip.
\donttest{all_prospects <- draft_prospects()}
}
