% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Draft.R
\name{draft_tracker}
\alias{draft_tracker}
\title{Access the real-time draft tracker}
\usage{
draft_tracker()
}
\value{
data.frame with one row per player
}
\description{
\code{draft_tracker()} scrapes the real-time draft tracker.
}
\examples{
draft_tracker <- draft_tracker()
}
