% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Draft.R
\name{draft_tracker}
\alias{draft_tracker}
\title{Access the real-time draft tracker}
\usage{
draft_tracker()
}
\value{
data.frame with one row per player
}
\description{
\code{draft_tracker()} retrieves the real-time draft tracker as a \code{data.frame} where each row represents player and includes detail on team identity, affiliation, and matchup-side context, player identity, role, handedness, and biographical profile, and venue/location geography and regional metadata.
}
\examples{
draft_tracker <- draft_tracker()
}
