% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Draft.R
\name{drafts}
\alias{drafts}
\title{Access all the drafts}
\usage{
drafts()
}
\value{
data.frame with one row per draft
}
\description{
\code{drafts()} scrapes all the drafts.
}
\examples{
all_drafts <- drafts()
}
