% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Draft.R
\name{drafts}
\alias{drafts}
\title{Access all the drafts}
\usage{
drafts()
}
\value{
data.frame with one row per draft
}
\description{
\code{drafts()} retrieves all the drafts as a \code{data.frame} where each row represents draft and includes detail on venue/location geography and regional metadata.
}
\examples{
all_drafts <- drafts()
}
