% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utility.R
\name{espn_api}
\alias{espn_api}
\title{Call ESPN API}
\usage{
espn_api(path, query = list(), type)
}
\arguments{
\item{path}{String}

\item{query}{list}

\item{type}{integer where 1=site.api and 2=sports.core}
}
\value{
parsed JSON
}
\description{
Call ESPN API
}
\keyword{internal}
