% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utility.R
\name{espn_api}
\alias{espn_api}
\title{Call the ESPN API with 429 (rate limit) error-handling}
\usage{
espn_api(path, query = list(), type)
}
\arguments{
\item{path}{character}

\item{query}{list}

\item{type}{character of 'g' for general and 'c' for core}
}
\value{
parsed JSON (i.e., data.frame or list)
}
\description{
Call the ESPN API with 429 (rate limit) error-handling
}
\keyword{internal}
