% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utility.R
\name{espn_api}
\alias{espn_api}
\title{Internal: call ESPN API}
\usage{
espn_api(path, query = list(), type = 1)
}
\arguments{
\item{path}{String API path}

\item{query}{list Query parameters}

\item{type}{integer where 1=site.api and 2=sports.core}
}
\value{
parsed JSON
}
\description{
Internal: call ESPN API
}
\keyword{internal}
