% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ESPNOther.R
\name{espn_futures}
\alias{espn_futures}
\title{Access the ESPN futures for a season}
\usage{
espn_futures(season = season_now())
}
\arguments{
\item{season}{integer in YYYYYYYY (e.g., 20242025); see \code{\link[=seasons]{seasons()}} for
reference}
}
\value{
nested data.frame with one row per type (outer) and book (inner)
}
\description{
\code{espn_futures()} scrapes the ESPN futures for a given \code{season}.
}
\examples{
ESPN_futures_20252026 <- espn_futures(20252026)
}
