% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ESPNGame.R
\name{espn_game_odds}
\alias{espn_game_odds}
\title{Access the ESPN odds for a game}
\usage{
espn_game_odds(game = 401777460)
}
\arguments{
\item{game}{integer ID (e.g., 401777460); see \code{\link[=espn_games]{espn_games()}} for
reference}
}
\value{
data.frame with one row per provider
}
\description{
\code{espn_game_odds()} retrieves the ESPN odds for a game as a \code{data.frame} where each row represents provider and includes detail on team identity, affiliation, and matchup-side context plus betting market snapshots with side/total prices and provider variation.
}
\examples{
ESPN_odds_SCF_20242025 <- espn_game_odds(game = 401777460)
}
