% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ESPNGame.R
\name{espn_game_summary}
\alias{espn_game_summary}
\title{Access the ESPN summary for a game}
\usage{
espn_game_summary(game = 401777460)
}
\arguments{
\item{game}{integer ID (e.g., 401777460); see \code{\link[=espn_games]{espn_games()}} for
reference}
}
\value{
list with various items
}
\description{
\code{espn_game_summary()} scrapes the ESPN summary for a \code{game}.
}
\examples{
ESPN_summary_SCF_20242025 <- espn_game_summary(game = 401777460)
}
