% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ESPNGame.R
\name{espn_games}
\alias{espn_games}
\title{Access the ESPN games for a season}
\usage{
espn_games(season = season_now())
}
\arguments{
\item{season}{integer in YYYYYYYY (e.g., 20242025); see \code{\link[=seasons]{seasons()}} for
reference}
}
\value{
data.frame with one row per ESPN game
}
\description{
\code{espn_games()} scrapes the ESPN games for a given \code{season}.
}
\examples{
ESPN_games_20242025 <- espn_games(season = 20242025)
}
