% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ESPNOther.R
\name{espn_injuries}
\alias{espn_injuries}
\title{Access the real-time ESPN injury reports}
\usage{
espn_injuries()
}
\value{
nested data.frame with one row per team (outer) and player (inner)
}
\description{
\code{espn_injuries()} scrapes the real-time ESPN injury reports for all the
teams.
}
\examples{
ESPN_injuries_now <- espn_injuries()
}
