% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ESPNGame.R
\name{espn_play_by_play}
\alias{espn_play_by_play}
\alias{espn_pbp}
\title{Access the ESPN play-by-play for a game}
\usage{
espn_play_by_play(game = 401777460)

espn_pbp(game = 401777460)
}
\arguments{
\item{game}{integer ID (e.g., 401777460); see \code{\link[=espn_games]{espn_games()}} for
reference}
}
\value{
data.frame with one row per event (play)
}
\description{
\code{espn_play_by_play()} retrieves the ESPN play-by-play for a game as a \code{data.frame} where each row represents event and includes detail on game timeline state, period/clock progression, and matchup flow, team identity, affiliation, and matchup-side context, and situational splits across home/road, strength state, and overtime/shootout states.
}
\examples{
ESPN_pbp_SCF_20242025 <- espn_play_by_play(game = 401777460)
}
