% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ESPNPlayer.R
\name{espn_players}
\alias{espn_players}
\title{Access all the ESPN players}
\usage{
espn_players()
}
\value{
data.frame with one row per ESPN player
}
\description{
\code{espn_players()} scrapes all the ESPN players.
}
\examples{
all_ESPN_players <- espn_players()
}
