% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ESPNPlayer.R
\name{espn_players}
\alias{espn_players}
\title{Access all the ESPN players}
\usage{
espn_players()
}
\value{
data.frame with one row per ESPN player
}
\description{
\code{espn_players()} retrieves all the ESPN players as a \code{data.frame} where each row represents ESPN player and includes detail on person-level profile context and performance history with situational splits.
}
\examples{
all_ESPN_players <- espn_players()
}
