% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ESPNTeam.R
\name{espn_team_summary}
\alias{espn_team_summary}
\title{Access the ESPN summary for a team}
\usage{
espn_team_summary(team = 1)
}
\arguments{
\item{team}{integer ID (e.g., 1); see \code{\link[=espn_teams]{espn_teams()}} for
reference}
}
\value{
data.frame with one row
}
\description{
\code{espn_team_summary()} retrieves the ESPN summary for a team as a \code{data.frame} where each row represents one result and includes detail on game timing, matchup state, scoring flow, and situational event detail.
}
\examples{
ESPN_summary_Boston_Bruins <- espn_team_summary(team = 1)
}
