% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ESPNTeam.R
\name{espn_team_summary}
\alias{espn_team_summary}
\title{Access the ESPN summary for a team}
\usage{
espn_team_summary(team = 3988803)
}
\arguments{
\item{team}{integer ID (e.g., 1); see \code{\link[=espn_teams]{espn_teams()}} for
reference}
}
\value{
data.frame with one row
}
\description{
\code{espn_team_summary()} scrapes the ESPN summary for a \code{team}.
}
\examples{
ESPN_summary_Boston_Bruins <- espn_team_summary(team = 1)
}
