% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ESPNOther.R
\name{espn_transactions}
\alias{espn_transactions}
\title{Access the ESPN transactions for a season}
\usage{
espn_transactions(season = season_now())
}
\arguments{
\item{season}{integer in YYYYYYYY (e.g., 20242025); the summer of the latter
year is included}
}
\value{
data.frame with one row per transaction
}
\description{
\code{espn_transactions()} retrieves the ESPN transactions for a season as a \code{data.frame} where each row represents transaction and includes detail on availability and transaction-status tracking detail.
}
\examples{
ESPN_transactions_20242025 <- espn_transactions(season = 20242025)
}
