% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Draft.R
\name{expansion_draft_picks}
\alias{expansion_draft_picks}
\title{Access all the expansion draft picks}
\usage{
expansion_draft_picks()
}
\value{
data.frame with one row per pick
}
\description{
\code{expansion_draft_picks()} retrieves all the expansion draft picks as a \code{data.frame} where each row represents pick and includes detail on date/season filtering windows and chronological context plus team identity, affiliation, and matchup-side context.
}
\examples{
all_expansion_draft_picks <- expansion_draft_picks()
}
