% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Clean.R
\name{flag_is_rebound}
\alias{flag_is_rebound}
\title{Flag if the shot attempt is a rebound attempt or creates a rebound for all the shots in a play-by-play}
\usage{
flag_is_rebound(play_by_play)
}
\arguments{
\item{play_by_play}{data.frame of play-by-play(s); see \code{\link[=gc_play_by_play]{gc_play_by_play()}} and/or \code{\link[=wsc_play_by_play]{wsc_play_by_play()}} for reference}
}
\value{
data.frame with one row per event (play) and added columns: \code{createdRebound} and \code{isRebound}
}
\description{
\code{flag_is_rebound()} flags whether a shot attempt is a rebound attempt (i.e., taken within 3 seconds of a prior blocked, missed, or saved attempt with no stoppage in between), and whether a shot attempt creates a rebound under the same definition.
}
\keyword{internal}
