% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Clean.R
\name{flag_is_rush}
\alias{flag_is_rush}
\title{Flag if the shot attempt is a rush attempt or not for all the shots in a play-by-play}
\usage{
flag_is_rush(play_by_play)
}
\arguments{
\item{play_by_play}{data.frame of play-by-play(s); see \code{\link[=gc_play_by_play]{gc_play_by_play()}} and/or \code{\link[=wsc_play_by_play]{wsc_play_by_play()}} for reference}
}
\value{
data.frame with one row per event (play) and added \code{isRush} column
}
\description{
\code{flag_is_rush()} flags whether a shot attempt is a rush attempt, defined as any shot attempt occurring within 4 seconds of a prior event in the neutral or defensive zone with no stoppage in play in between.
}
\keyword{internal}
