% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Franchise.R
\name{franchise_season_statistics}
\alias{franchise_season_statistics}
\alias{franchise_season_stats}
\title{Access the statistics for all the franchises by season and game type}
\usage{
franchise_season_statistics()

franchise_season_stats()
}
\value{
data.frame with one row per franchise per season per game type
}
\description{
\code{franchise_season_statistics()} retrieves the statistics for all the franchises by season and game type as a \code{data.frame} where each row represents franchise per season per game type and includes detail on date/season filtering windows and chronological context, team identity, affiliation, and matchup-side context, and production, workload, efficiency, and result-level performance outcomes.
}
\examples{
# May take >5s, so skip.
\donttest{franchise_season_stats <- franchise_season_statistics()}
}
