% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Franchise.R
\name{franchise_statistics}
\alias{franchise_statistics}
\alias{franchise_stats}
\title{Access the all-time statistics for all the franchises by game type}
\usage{
franchise_statistics()

franchise_stats()
}
\value{
data.frame with one row per franchise per game type
}
\description{
\code{franchise_statistics()} scrapes the all-time statistics for all the
franchises by game type.
}
\examples{
franchise_stats <- franchise_statistics()
}
