% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Franchise.R
\name{franchise_team_statistics}
\alias{franchise_team_statistics}
\alias{franchise_team_stats}
\title{Access the all-time statistics for all the franchises by team and game type}
\usage{
franchise_team_statistics()

franchise_team_stats()
}
\value{
data.frame with one row per team per franchise per game type
}
\description{
\code{franchise_team_statistics()} retrieves the all-time statistics for all the franchises by team and game type as a \code{data.frame} where each row represents team per franchise per game type and includes detail on date/season filtering windows and chronological context, team identity, affiliation, and matchup-side context, and production, workload, efficiency, and result-level performance outcomes.
}
\examples{
franchise_team_stats <- franchise_team_statistics()
}
