% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Franchise.R
\name{franchise_versus_franchise}
\alias{franchise_versus_franchise}
\alias{franchise_vs_franchise}
\title{Access the all-time statistics versus other franchises for all the
franchises by game type}
\usage{
franchise_versus_franchise()

franchise_vs_franchise()
}
\value{
data.frame with one row per franchise per franchise per game type
}
\description{
\code{franchise_versus_franchise()} retrieves the all-time statistics versus other franchises for all the franchises by game type as a \code{data.frame} where each row represents franchise per franchise per game type and includes detail on date/season filtering windows and chronological context plus team identity, affiliation, and matchup-side context.
}
\examples{
# May take >5s, so skip.
\donttest{franchise_vs_franchise <- franchise_versus_franchise()}
}
