% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Franchise.R
\name{franchises}
\alias{franchises}
\title{Access all the franchises}
\usage{
franchises()
}
\value{
data.frame with one row per franchise
}
\description{
\code{franchises()} scrapes all the franchises.
}
\examples{
all_franchises <- franchises()
}
