% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Franchise.R
\name{franchises}
\alias{franchises}
\title{Access all the franchises}
\usage{
franchises()
}
\value{
data.frame with one row per franchise
}
\description{
\code{franchises()} retrieves all the franchises as a \code{data.frame} where each row represents franchise and includes detail on team identity, affiliation, and matchup-side context.
}
\examples{
all_franchises <- franchises()
}
