% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Game.R
\name{game_odds}
\alias{game_odds}
\title{Access the real-time game odds for a country by partnered bookmaker}
\usage{
game_odds(country = "US")
}
\arguments{
\item{country}{two-letter code (e.g., 'CA'); see \code{\link[=countries]{countries()}} for reference}
}
\value{
data.frame with one row per game
}
\description{
\code{game_odds()} retrieves the real-time game odds for a country by partnered bookmaker as a \code{data.frame} where each row represents game and includes detail on betting market lines, prices, and provider-level context.
}
\examples{
game_odds_CA <- game_odds(country = 'CA')
}
