% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Game.R
\name{games}
\alias{games}
\title{Access all the games}
\usage{
games()
}
\value{
data.frame with one row per game
}
\description{
\code{games()} retrieves all the games as a \code{data.frame} where each row represents game and includes detail on game timeline state, period/clock progression, and matchup flow plus date/season filtering windows and chronological context.
}
\examples{
# May take >5s, so skip.
\donttest{all_games <- games()}
}
