% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Game.R
\name{games}
\alias{games}
\title{Access all the games}
\usage{
games()
}
\value{
data.frame with one row per game
}
\description{
\code{games()} scrapes all the games.
}
\examples{
# May take >5s, so skip.
\donttest{all_games <- games()}
}
