% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Game.R
\name{gc_summary}
\alias{gc_summary}
\title{Access the GameCenter (GC) summary for a game}
\usage{
gc_summary(game = 2023030417)
}
\arguments{
\item{game}{integer ID (e.g., 2025020275); see \code{\link[=games]{games()}} for reference}
}
\value{
list of various items
}
\description{
\code{gc_summary()} scrapes the GC summary for a given \code{game}.
}
\examples{
gc_summary_Martin_Necas_legacy_game <- gc_summary(game = 2025020275)
}
