% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Manager.R
\name{general_managers}
\alias{general_managers}
\alias{gms}
\title{Access all the general managers}
\usage{
general_managers()

gms()
}
\value{
data.frame with one row per general manager
}
\description{
\code{general_managers()} retrieves all the general managers as a \code{data.frame} where each row represents general manager and includes detail on player identity, role, handedness, and biographical profile.
}
\examples{
all_GMs <- general_managers()
}
