% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Manager.R
\name{general_managers}
\alias{general_managers}
\alias{gms}
\title{Access all the general managers}
\usage{
general_managers()

gms()
}
\value{
data.frame with one row per general manager
}
\description{
\code{general_managers()} scrapes all the general managers.
}
\examples{
all_GMs <- general_managers()
}
