% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Other.R
\name{get_attendance}
\alias{get_attendance}
\title{Get attendance for all seasons}
\usage{
get_attendance()
}
\value{
tibble with one row per season
}
\description{
\code{get_attendance()} retrieves information on each season, including but not limited to their ID and regular and playoff attendance. May soon be merged with \code{get_seasons()}.
}
\examples{
all_attendance <- get_attendance()
}
