% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LeagueOld.R
\name{get_attendance}
\alias{get_attendance}
\title{Access the attendance by season and game type}
\usage{
get_attendance()
}
\value{
data.frame with one row per season
}
\description{
\code{get_attendance()} is deprecated. Use \code{\link[=attendance]{attendance()}} instead.
}
