% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Other.R
\name{get_attendance}
\alias{get_attendance}
\title{Get attendance for all seasons}
\usage{
get_attendance()
}
\value{
tibble with one row per season
}
\description{
Get attendance for all seasons
}
\examples{
all_attendance <- get_attendance()
}
