% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OtherOld.R
\name{get_attendance}
\alias{get_attendance}
\title{Get the attendance for all the seasons}
\usage{
get_attendance()
}
\description{
\code{get_attendance()} is deprecated. Use \code{\link[=ns_attendance]{ns_attendance()}} instead.
}
