% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AwardOld.R
\name{get_award_winners}
\alias{get_award_winners}
\title{Access all the award winners/finalists}
\usage{
get_award_winners()
}
\value{
data.frame with one row per winner/finalist
}
\description{
\code{get_award_winners()} is deprecated. Use \code{\link[=award_winners]{award_winners()}} instead.
}
