% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AwardOld.R
\name{get_award_winners}
\alias{get_award_winners}
\title{Get all award winners/finalists}
\usage{
get_award_winners()
}
\description{
\code{get_award_winners()} is deprecated. Use \code{\link[=ns_award_winners]{ns_award_winners()}} instead.
}
