% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Award.R
\name{get_award_winners}
\alias{get_award_winners}
\title{Get all award winners/finalists}
\usage{
get_award_winners()
}
\value{
tibble with one row per award
}
\description{
Get all award winners/finalists
}
\examples{
all_awards <- get_awards()
}
