% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Award.R
\name{get_awards}
\alias{get_awards}
\title{Get all awards}
\usage{
get_awards()
}
\value{
tibble with one row per award
}
\description{
\code{get_awards()} retrieves information on each award, including but not limited to their trophy ID, name, description, creation date, and image URL.
}
\examples{
all_awards <- get_awards()
}
