% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlayoffOld.R
\name{get_bracket}
\alias{get_bracket}
\title{Access the playoff bracket for a season}
\usage{
get_bracket(season = season_now())
}
\arguments{
\item{season}{integer in YYYYYYYY (e.g., 20242025); see \code{\link[=seasons]{seasons()}} for
reference}
}
\value{
data.frame with one row per series
}
\description{
\code{get_bracket()} is deprecated. Use \code{\link[=bracket]{bracket()}} instead.
}
