% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlayoffOld.R
\name{get_bracket}
\alias{get_bracket}
\title{Get the playoff bracket of a season}
\usage{
get_bracket(season = ns_season()$seasonId)
}
\description{
\code{get_bracket()} is deprecated. Use \code{\link[=ns_bracket]{ns_bracket()}} instead.
}
