% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OtherOld.R
\name{get_configuration}
\alias{get_configuration}
\title{Access the configurations for team, skater, and goalie reports}
\usage{
get_configuration()
}
\description{
\code{get_configuration()} is defunct. Use \code{\link[=team_report_configurations]{team_report_configurations()}},
\code{\link[=skater_report_configurations]{skater_report_configurations()}}, and/or \code{\link[=goalie_report_configurations]{goalie_report_configurations()}}
instead.
}
