% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Other.R
\name{get_configuration}
\alias{get_configuration}
\title{Get configuration for skater, goalie, and team statistics}
\usage{
get_configuration()
}
\value{
list with 5 items
}
\description{
Get configuration for skater, goalie, and team statistics
}
\examples{
config <- get_configuration()
}
