% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OtherOld.R
\name{get_configuration}
\alias{get_configuration}
\title{Access the configurations for skater, goalie, and team statistics reports}
\usage{
get_configuration()
}
\value{
list with 5 items
}
\description{
\code{get_configuration()} is deprecated. Use
\code{\link[=statistics_report_configuration]{statistics_report_configuration()}} instead.
}
