% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OtherOld.R
\name{get_configuration}
\alias{get_configuration}
\title{Access the configurations for team, skater, and goalie reports}
\usage{
get_configuration()
}
\value{
list with 5 items
}
\description{
\code{get_configuration()} is deprecated. Use
\code{\link[=report_configurations]{report_configurations()}} instead.
}
