% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OtherOld.R
\name{get_configuration}
\alias{get_configuration}
\title{Get the configuration for skater, goalie, and team statistics}
\usage{
get_configuration()
}
\description{
\code{get_configuration()} is deprecated. Use \code{\link[=ns_statistics_configuration]{ns_statistics_configuration()}} instead.
}
