% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Other.R
\name{get_countries}
\alias{get_countries}
\title{Get all countries}
\usage{
get_countries()
}
\value{
tibble with one row per country
}
\description{
\code{get_countries()} retrieves information on each country, including but not limited to their ID, name, 2-letter code, and 3-letter code.
}
\examples{
all_countries <- get_countries()
}
