% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DraftOld.R
\name{get_draft_picks}
\alias{get_draft_picks}
\title{Access all the draft picks}
\usage{
get_draft_picks()
}
\value{
data.frame with one row per pick
}
\description{
\code{get_draft_picks()} is deprecated. Use \code{\link[=draft_picks]{draft_picks()}} instead.
}
