% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Draft.R
\name{get_draft_picks}
\alias{get_draft_picks}
\title{Get draft picks by year (and round)}
\usage{
get_draft_picks(year = get_season_now()$seasonId\%\%10000 - 1, round = "all")
}
\arguments{
\item{year}{integer Year in YYYY}

\item{round}{integer or string Round of 1:7 or 'all'}
}
\value{
tibble with one row per pick
}
\description{
Get draft picks by year (and round)
}
\examples{
draft_picks_2024 <- get_draft_picks(year=2024, round='all')
}
