% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DraftOld.R
\name{get_draft_picks}
\alias{get_draft_picks}
\title{Get all the draft picks}
\usage{
get_draft_picks()
}
\description{
\code{get_draft_picks()} is deprecated. Use \code{\link[=ns_draft_picks]{ns_draft_picks()}} instead.
}
