% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Draft.R
\name{get_draft_tracker}
\alias{get_draft_tracker}
\title{Get draft tracker as of now}
\usage{
get_draft_tracker()
}
\value{
tibble with one row per pick
}
\description{
\code{get_draft_tracker()} retrieves information on the latest draft, including but not limited to each pick's team ID, name, and overall number and selected player's name and position.
}
\examples{
draft_tracker <- get_draft_tracker()
}
