% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Draft.R
\name{get_draft_tracker}
\alias{get_draft_tracker}
\title{Get draft tracker as of now}
\usage{
get_draft_tracker()
}
\value{
tibble with one row per pick
}
\description{
Get draft tracker as of now
}
\examples{
draft_tracker <- get_draft_tracker()
}
