% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DraftOld.R
\name{get_draft_tracker}
\alias{get_draft_tracker}
\title{Access the real-time draft tracker}
\usage{
get_draft_tracker()
}
\value{
data.frame with one row per player
}
\description{
\code{get_draft_tracker()} is deprecated. Use \code{\link[=draft_tracker]{draft_tracker()}} instead.
}
