% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Draft.R
\name{get_drafts}
\alias{get_drafts}
\title{Get all drafts}
\usage{
get_drafts()
}
\value{
tibble with one row per draft
}
\description{
\code{get_drafts()} retrieves information on each draft, including but not limited to their year, type, venue, minimum and maximum player ages, and number of rounds and picks.
}
\examples{
all_drafts <- get_drafts()
}
