% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ESPNAthlete.R
\name{get_espn_athlete}
\alias{get_espn_athlete}
\title{Get athlete (player) by season and ESPN Athlete (Player) ID}
\usage{
get_espn_athlete(
  season = get_season_now()$seasonId\%\%10000,
  athlete = 3988803
)
}
\arguments{
\item{season}{integer Season in YYYY}

\item{athlete}{integer ESPN Athlete (Player) ID}
}
\value{
list with various items
}
\description{
\code{get_espn_athlete()} retrieves information on an \code{athlete} for a given \code{season}, including but not limited to his name, bio-metrics, birth date and location, position, team(s), and jersey number.
}
\examples{
ESPN_Charlie_McAvoy_20242025 <- get_espn_athlete(season=2025, athlete=3988803)
}
