% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ESPNAthlete.R
\name{get_espn_athlete}
\alias{get_espn_athlete}
\title{Get the ESPN information on an athlete for a season}
\usage{
get_espn_athlete(athlete = 3988803, season = season()\%\%10000)
}
\description{
\code{get_espn_athlete()} is temporarily defunct while we re-evaluate the
practicality of ESPN API information.
}
