% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ESPNPlayerOld.R
\name{get_espn_athletes}
\alias{get_espn_athletes}
\title{Access all the ESPN athletes (players)}
\usage{
get_espn_athletes()
}
\value{
data.frame with one row per ESPN athlete (player)
}
\description{
\code{get_espn_athletes()} is deprecated. Use \code{\link[=espn_players]{espn_players()}} instead.
}
