% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ESPNAthlete.R
\name{get_espn_athletes}
\alias{get_espn_athletes}
\title{Get all ESPN athletes}
\usage{
get_espn_athletes()
}
\value{
tibble with one row per athlete
}
\description{
\code{get_espn_athletes()} retrieves ESPN hyperlinks for each athlete; the hyperlinks are formatted in \verb{base/athletes/\{ESPN Athlete ID\}?query}. May soon be reworked to only return the ESPN Athlete IDs.
}
\examples{
all_ESPN_athletes <- get_espn_athletes()
}
