% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ESPNAthlete.R
\name{get_espn_athletes}
\alias{get_espn_athletes}
\title{Get all the ESPN athletes}
\usage{
get_espn_athletes()
}
\value{
data.frame with one row per athlete
}
\description{
\code{get_espn_athletes()} retrieves the ESPN ID for each athlete. Temporarily
deprecated while we re-evaluate the practicality of ESPN API information.
Use \code{\link[=players]{players()}} instead.
}
\examples{
all_ESPN_athletes <- get_espn_athletes()
}
