% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ESPNCoach.R
\name{get_espn_coach}
\alias{get_espn_coach}
\title{Get coach by ESPN Coach ID (and season)}
\usage{
get_espn_coach(coach = 5033, season = "all")
}
\arguments{
\item{coach}{integer ESPN Coach ID}

\item{season}{integer/string in YYYY or 'all'}
}
\value{
list with various items
}
\description{
\code{get_espn_coach()} retrieves information on a \code{coach} for a given \code{season} or all seasons, including but not limited to his or her name and head-shot URL. Access \code{get_espn_coaches()} for \code{coach} and \code{get_seasons()} for \code{season} references.
}
\examples{
ESPN_Paul_Maurice <- get_espn_coach(coach=5033, season='all')
}
