% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ESPNCoach.R
\name{get_espn_coach}
\alias{get_espn_coach}
\title{Get the ESPN information on a coach for a season}
\usage{
get_espn_coach(coach = 5033, season = "all")
}
\description{
\code{get_espn_coach()} is temporarily defunct while we re-evaluate the
practicality of ESPN API information.
}
