% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ESPNCoach.R
\name{get_espn_coach}
\alias{get_espn_coach}
\title{Get coach by ESPN Coach ID (and season)}
\usage{
get_espn_coach(coach = 5033, season = "")
}
\arguments{
\item{coach}{integer ESPN Coach ID}

\item{season}{integer/string Season in YYYY or '' for all seasons}
}
\value{
list with various items
}
\description{
Get coach by ESPN Coach ID (and season)
}
\examples{
ESPN_Paul_Maurice <- get_espn_coach(coach=5033, season='')
}
