% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ESPNCoach.R
\name{get_espn_coach_career}
\alias{get_espn_coach_career}
\title{Get career coaching records by ESPN Coach ID and game-type}
\usage{
get_espn_coach_career(coach = 5033, game_type = 0)
}
\arguments{
\item{coach}{integer ESPN Coach ID}

\item{game_type}{integer where 0=total, 1=regular, and 2=playoffs}
}
\value{
tibble with one row per statistic
}
\description{
\code{get_espn_coach_career()} retrieves information on a \code{coach} for a given \code{game_type}, including but not limited to his or her number of wins, ties, and losses.
}
\examples{
ESPN_Paul_Maurice_career <- get_espn_coach_career(coach=5033, game_type=0)
}
