% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ESPNCoach.R
\name{get_espn_coach_career}
\alias{get_espn_coach_career}
\title{Get the ESPN career records of a coach for a game type}
\usage{
get_espn_coach_career(coach = 5033, game_type = 0)
}
\description{
\code{get_espn_coach_career()} is temporarily defunct while we re-evaluate the
practicality of ESPN API information.
}
