% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ESPNCoachOld.R
\name{get_espn_coaches}
\alias{get_espn_coaches}
\title{Access the ESPN coaches for a season}
\usage{
get_espn_coaches()
}
\description{
\code{get_espn_coaches()} is defunct. Use \code{\link[=coaches]{coaches()}} instead.
}
