% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ESPNCoach.R
\name{get_espn_coaches}
\alias{get_espn_coaches}
\title{Get ESPN coaches by season}
\usage{
get_espn_coaches(season = get_season_now()$seasonId\%\%10000)
}
\arguments{
\item{season}{integer Season in YYYY}
}
\value{
tibble with one row per coach
}
\description{
Get ESPN coaches by season
}
\examples{
ESPN_coaches_20242025 <- get_espn_coaches(2025)
}
