% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ESPNEvent.R
\name{get_espn_event}
\alias{get_espn_event}
\title{Get information on an ESPN event}
\usage{
get_espn_event(event = 401687600)
}
\description{
\code{get_espn_event()} is temporarily defunct while we re-evaluate the
practicality of ESPN API information.
}
