% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ESPNEvent.R
\name{get_espn_event}
\alias{get_espn_event}
\title{Get event (game) by ESPN ID}
\usage{
get_espn_event(event = 401687600)
}
\arguments{
\item{event}{integer ESPN Event (Game) ID}
}
\value{
list with various items
}
\description{
Get event (game) by ESPN ID
}
\examples{
NJD_BUF_2024_10_04 <- get_espn_event(event=401687600)
}
