% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ESPNGameOld.R
\name{get_espn_event_odds}
\alias{get_espn_event_odds}
\title{Access the ESPN odds for an event (game)}
\usage{
get_espn_event_odds(event = 401777460)
}
\arguments{
\item{event}{integer ID (e.g., 401777460); see \code{\link[=espn_games]{espn_games()}} for
reference}
}
\value{
data.frame with one row per provider
}
\description{
\code{get_espn_event_odds()} is deprecated. Use \code{\link[=espn_game_odds]{espn_game_odds()}}
instead.
}
