% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ESPNEvent.R
\name{get_espn_event_officials}
\alias{get_espn_event_officials}
\title{Get the officials of an event}
\usage{
get_espn_event_officials(event = 401687600)
}
\description{
\code{get_espn_event_officials()} is defunct. Use \code{\link[=gc_summary]{gc_summary()}} instead.
}
