% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ESPNEvent.R
\name{get_espn_event_play_by_play}
\alias{get_espn_event_play_by_play}
\title{Get event (game) play-by-play by ESPN Event (Game) ID}
\usage{
get_espn_event_play_by_play(event = 401687600)
}
\arguments{
\item{event}{integer ESPN Event (Game) ID}
}
\value{
tibble with one row per play
}
\description{
\code{get_espn_event_play_by_play()} retrieves ESPN-provided information on each play for a given \code{event}, including but not limited to their ID, type, time of occurrence, strength-state, participants, and X and Y coordinates. Access \code{get_espn_events()} for \code{event} reference.
}
\examples{
NJD_BUF_2024_10_04_pbp <- get_espn_event_play_by_play(event=401687600)
}
