% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ESPNEvent.R
\name{get_espn_event_play_by_play}
\alias{get_espn_event_play_by_play}
\title{Get event (game) play-by-play by ESPN Event (Game) ID}
\usage{
get_espn_event_play_by_play(event = 401687600)
}
\arguments{
\item{event}{integer ESPN Event (Game) ID}
}
\value{
tibble with one row per play
}
\description{
Get event (game) play-by-play by ESPN Event (Game) ID
}
\examples{
NJD_BUF_2024_10_04_pbp <- get_espn_event_play_by_play(event=401687600)
}
