% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ESPNGameOld.R
\name{get_espn_event_stars}
\alias{get_espn_event_stars}
\title{Access the three stars for an ESPN event (game)}
\usage{
get_espn_event_stars()
}
\description{
\code{get_espn_event_stars()} is defunct. Use \code{\link[=gc_summary]{gc_summary()}} and/or
\code{\link[=wsc_summary]{wsc_summary()}} instead.
}
