% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ESPNEvent.R
\name{get_espn_event_stars}
\alias{get_espn_event_stars}
\title{Get event (game) stars by ESPN Event (Game) ID}
\usage{
get_espn_event_stars(event = 401687600)
}
\arguments{
\item{event}{integer ESPN Event (Game) ID}
}
\value{
tibble with one row per athlete
}
\description{
Get event (game) stars by ESPN Event (Game) ID
}
\examples{
NJD_BUF_2024_10_04_stars <- get_espn_event_stars(event=401687600)
}
