% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ESPNGameOld.R
\name{get_espn_events}
\alias{get_espn_events}
\title{Access the ESPN events (games) by start and end dates}
\usage{
get_espn_events()
}
\description{
\code{get_espn_events()} is defunct. Use \code{\link[=espn_games]{espn_games()}} instead.
}
