% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ESPN.R
\name{get_espn_events}
\alias{get_espn_events}
\title{Get ESPN events (games) by season}
\usage{
get_espn_events(season = get_season_now()$seasonId\%/\%10000)
}
\arguments{
\item{season}{integer Season in YYYY}
}
\value{
tibble with one row per event (game)
}
\description{
Get ESPN events (games) by season
}
\examples{
ESPN_events_2024 <- get_espn_events(2024)
}
