% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ESPNEvent.R
\name{get_espn_events}
\alias{get_espn_events}
\title{Get the ESPN events for an interval of dates}
\usage{
get_espn_events(start_date = 20241004, end_date = 20250624)
}
\arguments{
\item{start_date}{integer in YYYYMMDD (e.g., 20241004)}

\item{end_date}{integer in YYYYMMDD (e.g., 20250624)}
}
\value{
data.frame with one row per ESPN event
}
\description{
\code{get_espn_events()} retrieves the ESPN ID for each event for an interval of
dates bound by a given set of \code{start_date} and \code{end_date}. Access
\code{seasons()} for \code{start_date} and \code{end_date} references. Note the date
format differs from the NHL API; will soon be fixed to accept both.
Temporarily deprecated while we re-evaluate the practicality of ESPN API
information. Use \code{\link[=games]{games()}} instead.
}
\examples{
ESPN_events_20242025 <- get_espn_events(
  start_date = 20241004, 
  end_date   = 20250624
)
}
