% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ESPNOtherOld.R
\name{get_espn_futures}
\alias{get_espn_futures}
\title{Access the ESPN futures for a season}
\usage{
get_espn_futures()
}
\description{
\code{get_espn_futures()} is defunct. Use \code{\link[=espn_futures]{espn_futures()}} instead.
}
