% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ESPN.R
\name{get_espn_futures}
\alias{get_espn_futures}
\title{Get ESPN futures by season}
\usage{
get_espn_futures(season = get_season_now()$seasonId\%/\%10000)
}
\arguments{
\item{season}{integer Season in YYYY}
}
\value{
tibble with one row per future
}
\description{
Get ESPN futures by season
}
\examples{
espn_futures_2024 <- get_espn_futures(2024)
}
