% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ESPNOther.R
\name{get_espn_futures}
\alias{get_espn_futures}
\title{Get the ESPN futures for a season}
\usage{
get_espn_futures(season = ns_season()\%\%10000)
}
\arguments{
\item{season}{integer in YYYY (e.g., 2026)}
}
\value{
nested data.frame with one row per type (outer) and book (inner)
}
\description{
\code{get_espn_futures()} retrieves real-time ESPN futures of various types for a
given \code{season}. Access \code{ns_seasons()} for \code{season} reference. Note the
season format differs from the NHL API; will soon be fixed to accept both.
}
\examples{
ESPN_futures_20252026 <- get_espn_futures(2026)
}
