% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ESPNOther.R
\name{get_espn_injuries}
\alias{get_espn_injuries}
\title{Get ESPN injury reports as of now}
\usage{
get_espn_injuries()
}
\value{
nested tibble with one row per team and player
}
\description{
\code{get_espn_injuries()} retrieves injury reports by team.
}
\examples{
ESPN_injuries_now <- get_espn_injuries()
}
