% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ESPNOtherOld.R
\name{get_espn_injuries}
\alias{get_espn_injuries}
\title{Access the real-time ESPN injury reports}
\usage{
get_espn_injuries()
}
\value{
nested data.frame with one row per team (outer) and player (inner)
}
\description{
\code{get_espn_injuries()} is deprecated. Use \code{\link[=espn_injuries]{espn_injuries()}} instead.
}
