% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ESPN.R
\name{get_espn_injuries}
\alias{get_espn_injuries}
\title{Get ESPN injury reports as of now}
\usage{
get_espn_injuries()
}
\value{
nested tibble with one row per team and one row per player
}
\description{
Get ESPN injury reports as of now
}
\examples{
espn_injuries_now <- get_espn_injuries()
}
