% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ESPNTeamOld.R
\name{get_espn_team}
\alias{get_espn_team}
\title{Access the ESPN summary for a team and season}
\usage{
get_espn_team()
}
\description{
\code{get_espn_team()} is defunct. Use \code{\link[=espn_team_summary]{espn_team_summary()}} instead.
}
