% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ESPNTeam.R
\name{get_espn_team}
\alias{get_espn_team}
\title{Get the ESPN information on a team for a season}
\usage{
get_espn_team(team = 1, season = ns_season()\%\%10000)
}
\description{
\code{get_espn_team()} is temporarily defunct while we re-evaluate the
practicality of ESPN API information.
}
