% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ESPNTeamOld.R
\name{get_espn_teams}
\alias{get_espn_teams}
\title{Access all the ESPN teams for a season}
\usage{
get_espn_teams()
}
\description{
\code{get_espn_teams()} is defunct. Use \code{\link[=espn_teams]{espn_teams()}} instead.
}
