% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ESPNOther.R
\name{get_espn_transactions}
\alias{get_espn_transactions}
\title{Get ESPN transactions by start and end dates}
\usage{
get_espn_transactions(start_date = 20241004, end_date = 20250624)
}
\arguments{
\item{start_date}{integer Start Date in YYYYMMDD}

\item{end_date}{integer End Date in YYYYMMDD}
}
\value{
tibble with one row per transaction
}
\description{
\code{get_espn_transactions()} retrieves information on each transaction for a given set of \code{start_date} and \code{end_date}, including but not limited to their date, description, and involved teams. Access \code{get_seasons()} for \code{start_season} and \code{end_season} references.
}
\examples{
ESPN_transactions_20242025 <- get_espn_transactions(start_date=20241004, end_date=20250624)
}
