% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ESPNOtherOld.R
\name{get_espn_transactions}
\alias{get_espn_transactions}
\title{Access the ESPN transactions by start and end dates}
\usage{
get_espn_transactions()
}
\description{
\code{get_espn_transactions()} is defunct. Use \code{\link[=espn_transactions]{espn_transactions()}} instead.
}
