% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ESPN.R
\name{get_espn_transactions}
\alias{get_espn_transactions}
\title{Get ESPN transactions by season}
\usage{
get_espn_transactions(season = get_season_now()$seasonId\%/\%10000)
}
\arguments{
\item{season}{integer Season in YYYY}
}
\value{
tibble with one row per transaction
}
\description{
Get ESPN transactions by season
}
\examples{
espn_transactions_2024 <- get_espn_transactions(2024)
}
