% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FranchiseOld.R
\name{get_franchise_season_by_season}
\alias{get_franchise_season_by_season}
\title{Access the statistics for all the franchises by season and game type}
\usage{
get_franchise_season_by_season()
}
\value{
data.frame with one row per franchise per season per game type
}
\description{
\code{get_franchise_season_by_season()} is deprecated. Use
\code{\link[=franchise_season_statistics]{franchise_season_statistics()}} instead.
}
