% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FranchiseOld.R
\name{get_franchise_team_totals}
\alias{get_franchise_team_totals}
\title{Access the all-time statistics for all the franchises by team and game type}
\usage{
get_franchise_team_totals()
}
\value{
data.frame with one row per team per franchise per game type
}
\description{
\code{get_franchise_team_totals()} is deprecated. Use
\code{\link[=franchise_team_statistics]{franchise_team_statistics()}} instead.
}
