% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Record.R
\name{get_franchise_team_totals}
\alias{get_franchise_team_totals}
\title{Get all franchises' teams' totals}
\usage{
get_franchise_team_totals()
}
\value{
tibble with one row per team
}
\description{
Get all franchises' teams' totals
}
\examples{
all_franchise_team_totals <- get_franchise_team_totals()
}
