% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RecordOld.R
\name{get_franchise_team_totals}
\alias{get_franchise_team_totals}
\title{Get all the franchises' teams' all-time totals}
\usage{
get_franchise_team_totals()
}
\description{
\code{get_franchise_team_totals()} is deprecated. Use
\code{\link[=ns_franchise_team_totals]{ns_franchise_team_totals()}} instead.
}
