% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FranchiseOld.R
\name{get_franchise_vs_franchise}
\alias{get_franchise_vs_franchise}
\title{Access the all-time statistics versus other franchises for all the
franchises by game type}
\usage{
get_franchise_vs_franchise()
}
\value{
data.frame with one row per franchise per franchise per game type
}
\description{
\code{get_franchise_vs_franchise()} is deprecated. Use
\code{\link[=franchise_versus_franchise]{franchise_versus_franchise()}} instead.
}
