% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Record.R
\name{get_franchise_vs_franchise}
\alias{get_franchise_vs_franchise}
\title{Get all franchises' all-time records versus other franchises}
\usage{
get_franchise_vs_franchise(game_type = 2)
}
\arguments{
\item{game_type}{integer Game-type where 2=regular and 3=playoffs}
}
\value{
tibble with one row per franchise versus franchise
}
\description{
Get all franchises' all-time records versus other franchises
}
\examples{
franchise_vs_franchise <- get_franchise_vs_franchise()
}
