% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RecordOld.R
\name{get_franchise_vs_franchise}
\alias{get_franchise_vs_franchise}
\title{Get all the franchises' all-time records versus other franchises}
\usage{
get_franchise_vs_franchise()
}
\description{
\code{get_franchise_vs_franchise()} is deprecated. Use
\code{\link[=ns_franchise_versus_franchise]{ns_franchise_versus_franchise()}} instead.
}
