% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FranchiseOld.R
\name{get_franchise_vs_franchise}
\alias{get_franchise_vs_franchise}
\title{Get the all-time statistics versus other franchises for all the franchises}
\usage{
get_franchise_vs_franchise()
}
\value{
data.frame with one row per franchise per franchise
}
\description{
\code{get_franchise_vs_franchise()} is deprecated. Use
\code{\link[=ns_franchise_versus_franchise]{ns_franchise_versus_franchise()}} instead.
}
