% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Team.R
\name{get_franchises}
\alias{get_franchises}
\title{Get all franchises}
\usage{
get_franchises()
}
\value{
tibble with one row per franchise
}
\description{
Get all franchises
}
\examples{
all_franchises <- get_franchises()
}
