% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FranchiseOld.R
\name{get_franchises}
\alias{get_franchises}
\title{Get all the franchises}
\usage{
get_franchises()
}
\value{
data.frame with one row per franchise
}
\description{
\code{get_franchises()} is deprecated. Use \code{\link[=ns_franchises]{ns_franchises()}} instead.
}
