% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Game.R
\name{get_game_boxscore}
\alias{get_game_boxscore}
\title{Get boxscore by game, team, and player-type}
\usage{
get_game_boxscore(game = 2024020602, team = "home", player_type = "forwards")
}
\arguments{
\item{game}{integer Game ID}

\item{team}{string Team whether 'home' or 'away'}

\item{player_type}{string Player-type of 'forwards', 'defensemen', or 'goalies'}
}
\value{
tibble with one row per player
}
\description{
Get boxscore by game, team, and player-type
}
