% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GameOld.R
\name{get_game_landing}
\alias{get_game_landing}
\title{Access the GameCenter (GC) summary for a game}
\usage{
get_game_landing(game = 2023030417)
}
\arguments{
\item{game}{integer ID (e.g., 2025020275); see \code{\link[=games]{games()}} for reference}
}
\value{
list of various items
}
\description{
\code{get_game_landing()} is deprecated. Use \code{\link[=gc_summary]{gc_summary()}} instead.
}
