% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Game.R
\name{get_game_landing}
\alias{get_game_landing}
\title{Get GameCenter (GC) game-landing by game}
\usage{
get_game_landing(game = 2024020602)
}
\arguments{
\item{game}{integer Game ID}
}
\value{
list of various items
}
\description{
\code{get_game_landing()} retrieves GC-provided information on a \code{game}, including but not limited to its type, venue, start time, clock, home and away teams, and TV broadcast(s). Access \code{get_games()} for \code{game} reference.
}
\examples{
game_landing_2024030411 <- get_game_landing(game=2024030411)
}
