% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Game.R
\name{get_game_landing}
\alias{get_game_landing}
\title{Get GC game landing by game}
\usage{
get_game_landing(game = 2024020602)
}
\arguments{
\item{game}{integer Game ID}
}
\value{
list of 24 items
}
\description{
Get GC game landing by game
}
\examples{
game_landing_2024030411 <- get_game_landing(game=2024030411)
}
