% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GameOld.R
\name{get_game_story}
\alias{get_game_story}
\title{Access the World Showcase (WSC) summary for a game}
\usage{
get_game_story(game = 2023030417)
}
\arguments{
\item{game}{integer ID (e.g., 2025020275); see \code{\link[=games]{games()}} for reference}
}
\value{
list of various items
}
\description{
\code{get_game_story()} is deprecated. Use \code{\link[=wsc_summary]{wsc_summary()}} instead.
}
