% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Game.R
\name{get_game_story}
\alias{get_game_story}
\title{Get World Showcase (WSC) game-story by game}
\usage{
get_game_story(game = 2024020602)
}
\arguments{
\item{game}{integer Game ID}
}
\value{
list of various items
}
\description{
\code{get_game_story()} retrieves WSC-provided information on a \code{game}, including but not limited to its type, venue, start time, clock, home and away teams, and TV broadcast(s). Access \code{get_games()} for \code{game} reference.
}
\examples{
game_story_2024030411 <- get_game_story(game=2024030411)
}
