% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GameOld.R
\name{get_game_story}
\alias{get_game_story}
\title{Get the World Showcase (WSC) summary of a game}
\usage{
get_game_story(game = 2025020275)
}
\description{
\code{get_game_story()} is deprecated. Use \code{\link[=ns_wsc_summary]{ns_wsc_summary()}} instead.
}
