% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Game.R
\name{get_game_story}
\alias{get_game_story}
\title{Get game story by game}
\usage{
get_game_story(game_id = 2024020602)
}
\arguments{
\item{game_id}{integer Game ID}
}
\value{
list of 24 items
}
\description{
Get game story by game
}
