% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Game.R
\name{get_games}
\alias{get_games}
\title{Get all games}
\usage{
get_games()
}
\value{
tibble with one row per game
}
\description{
\code{get_games()} retrieves information on each game, including but not limited to their ID, season, type, start date and time, and home and visiting teams' IDs and scores.
}
\examples{
# This may take >5s, so skip.
\donttest{
  all_games <- get_games()
}
}
