% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GameOld.R
\name{get_gc_play_by_play}
\alias{get_gc_play_by_play}
\title{Get the GameCenter (GC) play-by-play of a game}
\usage{
get_gc_play_by_play(game = 2025020275)
}
\description{
\code{get_gc_play_by_play()} is deprecated. Use \code{\link[=ns_gc_play_by_play]{ns_gc_play_by_play()}} instead.
}
