% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OtherOld.R
\name{get_glossary}
\alias{get_glossary}
\title{Access the glossary}
\usage{
get_glossary()
}
\value{
data.frame with one row per terminology
}
\description{
\code{get_glossary()} is deprecated. Use \code{\link[=glossary]{glossary()}} instead.
}
