% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Players.R
\name{get_goalie_leaders}
\alias{get_goalie_leaders}
\title{Get goalie stats leaders for a season}
\usage{
get_goalie_leaders(season = 20242025, game_type = 2, category = "wins")
}
\arguments{
\item{season}{string Season in 'YYYYYYYY' e.g. 20242025}

\item{game_type}{integer 2=regular, 3=playoffs}

\item{category}{string e.g. wins, shutouts, savePctg, goalsAgainstAverage
(note that some are only available in recent years)}
}
\value{
tibble with one row per skater
}
\description{
Get goalie stats leaders for a season
}
