% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GoalieOld.R
\name{get_goalie_leaders}
\alias{get_goalie_leaders}
\title{Get the goalie statistics leaders for a season, game type, and category}
\usage{
get_goalie_leaders(season = "current", game_type = "", category = "wins")
}
\description{
\code{get_goalie_leaders()} is deprecated. Use \code{\link[=ns_goalie_leaders]{ns_goalie_leaders()}} instead.
}
