% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Goalie.R
\name{get_goalie_milestones}
\alias{get_goalie_milestones}
\title{Get goalie milestones}
\usage{
get_goalie_milestones()
}
\value{
tibble with one row per goalie
}
\description{
\code{get_goalie_milestones()} retrieves information on each goalie close to a milestone, including but not limited to their ID, name, and statistics.
}
\examples{
goalie_milestones <- get_goalie_milestones()
}
