% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Goalie.R
\name{get_goalie_milestones}
\alias{get_goalie_milestones}
\title{Get goalie milestones}
\usage{
get_goalie_milestones()
}
\value{
tibble with one row per goalie
}
\description{
Get goalie milestones
}
\examples{
goalie_milestones <- get_goalie_milestones()
}
