% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GoalieOld.R
\name{get_goalie_milestones}
\alias{get_goalie_milestones}
\title{Get the goalies on milestone watch}
\usage{
get_goalie_milestones()
}
\description{
\code{get_goalie_milestones()} is deprecated. Use \code{\link[=goalie_milestones]{goalie_milestones()}}
instead.
}
