% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GoalieOld.R
\name{get_goalie_statistics}
\alias{get_goalie_statistics}
\title{Get goalie statistics}
\usage{
get_goalie_statistics()
}
\description{
\code{get_goalie_statistics()} is defunct. Use \code{\link[=goalie_statistics]{goalie_statistics()}} instead.
}
