% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GoalieOld.R
\name{get_goalie_statistics}
\alias{get_goalie_statistics}
\title{Get goalie statistics}
\usage{
get_goalie_statistics(
  season = 20242025,
  teams = 1:100,
  game_types = 1:3,
  dates = c("2025-01-01"),
  report = "summary",
  is_aggregate = FALSE,
  is_game = FALSE
)
}
\description{
\code{get_goalie_statistics()} is defunct. Use \code{\link[=ns_goalie_statistics]{ns_goalie_statistics()}} instead.
}
