% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GoalieOld.R
\name{get_goalies}
\alias{get_goalies}
\title{Access all the goalies for a range of seasons}
\usage{
get_goalies()
}
\description{
\code{get_goalies()} is defunct. Use \code{\link[=players]{players()}} instead.
}
