% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GoalieOld.R
\name{get_goalies}
\alias{get_goalies}
\title{Get the goalies for a range of seasons}
\usage{
get_goalies(start_season = 19171918, end_season = 20242025)
}
\description{
\code{get_goalies()} is defunct. Use \code{\link[=ns_players]{ns_players()}} instead.
}
