% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Goalie.R
\name{get_goalies}
\alias{get_goalies}
\title{Get all goalies' biographies from a range of years}
\usage{
get_goalies(start_year = 1917, end_year = 2025)
}
\arguments{
\item{start_year}{integer Year to start search}

\item{end_year}{integer Year to end search}
}
\value{
tibble with one row per goalie
}
\description{
Get all goalies' biographies from a range of years
}
