% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Goalie.R
\name{get_goalies}
\alias{get_goalies}
\title{Get all goalies' biographies by range of seasons}
\usage{
get_goalies(start_season = 19171918, end_season = get_season_now()$seasonId)
}
\arguments{
\item{start_season}{integer in YYYYYYYY}

\item{end_season}{integer in YYYYYYYY}
}
\value{
tibble with one row per goalie
}
\description{
\code{get_goalies()} retrieves information about each goalie for a specified \code{start_season} and \code{end_season}, including but not limited to their ID, name, bio-metrics, and career statistics. This will soon become deprecated as \code{get_players()} can list all players and their IDs much quicker.
}
\examples{
goalies_2000s <- get_goalies(start_season=20002001, end_season=20242025)
}
