% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Other.R
\name{get_officials}
\alias{get_officials}
\title{Get all officials}
\usage{
get_officials()
}
\value{
tibble with one row per official
}
\description{
\code{get_officials()} retrieves information on each official, including but not limited to their ID, name, and birth date and location.
}
\examples{
all_officials <- get_officials()
}
