% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OtherOld.R
\name{get_partner_odds}
\alias{get_partner_odds}
\title{Get the real-time game odds for a country by partnered bookmaker}
\usage{
get_partner_odds(country = "US")
}
\arguments{
\item{country}{two-letter code (e.g., 'CA'); see \code{\link[=countries]{countries()}} for
reference}
}
\value{
data.frame with one row per game
}
\description{
\code{get_partner_odds()} is deprecated. Use \code{\link[=game_partner_odds]{game_partner_odds()}} instead.
}
