% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Other.R
\name{get_partner_odds}
\alias{get_partner_odds}
\title{Get partner odds as of now}
\usage{
get_partner_odds(country = "US")
}
\arguments{
\item{country}{string 2-letter code}
}
\value{
tibble with one row per game
}
\description{
\code{get_partner_odds()} retrieves NHL-partnered information on each game for a specified \code{country}, including but not limited to their ID and home and away team odds.
}
\examples{
partner_odds_now_CA <- get_partner_odds(country='CA')
}
