% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OtherOld.R
\name{get_partner_odds}
\alias{get_partner_odds}
\title{Get partner odds as of now}
\usage{
get_partner_odds(country = "US")
}
\description{
\code{get_partner_odds()} is deprecated. Use \code{\link[=ns_partner_odds]{ns_partner_odds()}} instead.
}
