% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Other.R
\name{get_partner_odds}
\alias{get_partner_odds}
\title{Get partner odds now}
\usage{
get_partner_odds(country = "US")
}
\arguments{
\item{country}{string Country code e.g. 'US'}
}
\value{
???
}
\description{
Get partner odds now
}
