% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlayerOld.R
\name{get_player_game_log}
\alias{get_player_game_log}
\title{Get the game log of a player for a season and game type}
\usage{
get_player_game_log(player = 8480039, season = "now", game_type = "")
}
\description{
\code{get_player_game_log()} is deprecated. Use \code{\link[=ns_player_game_log]{ns_player_game_log()}} instead.
}
