% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Players.R
\name{get_player_game_log}
\alias{get_player_game_log}
\title{Get a player's game log for a season}
\usage{
get_player_game_log(player_id, season = 20242025, game_type = 2)
}
\arguments{
\item{player_id}{integer NHL player ID}

\item{season}{string Season in 'YYYYYYYY' e.g. 20242025}

\item{game_type}{integer 2=regular, 3=playoffs}
}
\value{
tibble with one row per game
}
\description{
Get a player's game log for a season
}
