% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Player.R
\name{get_player_landing}
\alias{get_player_landing}
\title{Get landing by player}
\usage{
get_player_landing(player = 8480039)
}
\arguments{
\item{player}{integer Player ID}
}
\value{
list with various items
}
\description{
\code{get_player_landing()} retrieves information on a \code{player}, including but not limited to his ID, name, bio-metrics, career statistics, and awards. Access \code{get_players()} for \code{player} reference.
}
\examples{
Mikko_Rantanen_landing <- get_player_landing(player=8478420)
}
