% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Player.R
\name{get_player_landing}
\alias{get_player_landing}
\title{Get a player's landing}
\usage{
get_player_landing(player_id = 8480039)
}
\arguments{
\item{player_id}{integer Player ID}
}
\value{
list with 36 items
}
\description{
Get a player's landing
}
