% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Player.R
\name{get_players}
\alias{get_players}
\title{Get all players}
\usage{
get_players()
}
\value{
tibble with one row per player
}
\description{
\code{get_players()} retrieves information on each player, including but not limited to their ID, name, bio-metrics, birth date and location, and hall-of-fame status.
}
\examples{
# This may take >5s, so skip.
\donttest{
  all_players <- get_players()
}
}
