% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlayerOld.R
\name{get_players}
\alias{get_players}
\title{Access all the players}
\usage{
get_players()
}
\value{
data.frame with one row per player
}
\description{
\code{get_players()} is deprecated. Use \code{\link[=players]{players()}} instead.
}
