% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Playoff.R
\name{get_playoff_bracket}
\alias{get_playoff_bracket}
\title{Get playoff bracket by year}
\usage{
get_playoff_bracket(year = 2025)
}
\arguments{
\item{year}{integer Year in YYYY}
}
\value{
tibble with one row per match-up
}
\description{
Get playoff bracket by year
}
