% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LeagueOld.R
\name{get_schedule}
\alias{get_schedule}
\title{Access the schedule for a date}
\usage{
get_schedule(date = "2025-01-01")
}
\arguments{
\item{date}{character in 'YYYY-MM-DD' (e.g., '2025-01-01'); see
\code{\link[=seasons]{seasons()}} for reference}
}
\value{
data.frame with one row per game
}
\description{
\code{get_schedule()} is deprecated. Use \code{\link[=schedule]{schedule()}} instead.
}
