% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/League.R
\name{get_schedule}
\alias{get_schedule}
\title{Get schedule by date}
\usage{
get_schedule(date = "2025-01-01")
}
\arguments{
\item{date}{string Date in 'YYYY-MM-DD'}
}
\value{
tibble with one row per game
}
\description{
Get schedule by date
}
\examples{
schedule_2025_01_02 <- get_schedule(date='2025-01-02')
}
