% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/League.R
\name{get_schedule}
\alias{get_schedule}
\title{Get schedule for week by date}
\usage{
get_schedule(date = "2025-01-01")
}
\arguments{
\item{date}{string Date in 'YYYY-MM-DD'}
}
\value{
list of 8 items
}
\description{
Get schedule for week by date
}
