% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LeagueOld.R
\name{get_schedule}
\alias{get_schedule}
\title{Get schedule by date}
\usage{
get_schedule(date = "2025-01-01")
}
\description{
\code{get_schedule()} is deprecated. Use \code{\link[=ns_schedule]{ns_schedule()}} instead.
}
