% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Game.R
\name{get_scores}
\alias{get_scores}
\title{Get score(s) by date}
\usage{
get_scores(date = "2025-01-01")
}
\arguments{
\item{date}{string in 'YYYY-MM-DD'}
}
\value{
tibble with one row per game
}
\description{
\code{get_scores()} retrieves information on each game for a given \code{date}, including but not limited to their ID; type; venue; start time; period and intermission clocks; and home and away teams' IDs, names, and scores. Access \code{get_seasons()} for \code{date} reference.
}
\examples{
scores_2025_01_02 <- get_scores(date='2025-01-02')
}
