% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Game.R
\name{get_scores}
\alias{get_scores}
\title{Get score(s) by date}
\usage{
get_scores(date = "2025-01-01")
}
\arguments{
\item{date}{string Date in 'YYYY-MM-DD'}
}
\value{
tibble with one row per game
}
\description{
Get score(s) by date
}
\examples{
scores_2025_01_02 <- get_scores(date='2025-01-02')
}
