% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GameOld.R
\name{get_scores}
\alias{get_scores}
\title{Get the score(s) of the game(s) for a date}
\usage{
get_scores(date = "now")
}
\description{
\code{get_scores()} is deprecated. Use \code{\link[=ns_scores]{ns_scores()}} instead.
}
