% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Other.R
\name{get_season_now}
\alias{get_season_now}
\title{Get season as of now}
\usage{
get_season_now()
}
\value{
tibble with one row
}
\description{
\code{get_season_now()} retrieves information on the current season, including but not limited to its ID and game-type.
}
\examples{
season_now <- get_season_now()
}
