% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LeagueOld.R
\name{get_season_now}
\alias{get_season_now}
\title{Access the season and game type as of now}
\usage{
get_season_now()
}
\description{
\code{get_season_now()} is defunct. Use \code{\link[=season_now]{season_now()}} and/or \code{\link[=game_type_now]{game_type_now()}}
instead.
}
