% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LeagueOld.R
\name{get_seasons}
\alias{get_seasons}
\title{Get all the seasons}
\usage{
get_seasons()
}
\value{
data.frame with one row per season
}
\description{
\code{get_seasons()} is deprecated. Use \code{\link[=ns_seasons]{ns_seasons()}} instead.
}
