% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlayoffOld.R
\name{get_series_schedule}
\alias{get_series_schedule}
\title{Access the playoff schedule for a season and series}
\usage{
get_series_schedule(season = season_now(), series = "a")
}
\arguments{
\item{season}{integer in YYYYYYYY (e.g., 20242025); see \code{\link[=seasons]{seasons()}} for
reference}

\item{series}{one-letter code (e.g., 'O'); see \code{\link[=series]{series()}} and/or
\code{\link[=bracket]{bracket()}} for reference}
}
\value{
data.frame with one row per game
}
\description{
\code{get_series_schedule()} is deprecated. Use \code{\link[=series_schedule]{series_schedule()}} instead.
}
