% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Playoff.R
\name{get_series_schedule}
\alias{get_series_schedule}
\title{Get playoff series schedule}
\usage{
get_series_schedule(season = get_season_now()$seasonId, series = "a")
}
\arguments{
\item{season}{integer Season in YYYYYYYY}

\item{series}{string Series code e.g. 'a'}
}
\value{
tibble with one row per game
}
\description{
Get playoff series schedule
}
