% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Game.R
\name{get_shift_charts}
\alias{get_shift_charts}
\title{Get shift charts}
\usage{
get_shift_charts(game = 2024020602)
}
\arguments{
\item{game}{integer Game ID}
}
\value{
tibble with one row per shift
}
\description{
Get shift charts
}
\examples{
shift_charts_2024030411 <- get_shift_charts(game=2024030411)
}
