% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Game.R
\name{get_shift_charts}
\alias{get_shift_charts}
\title{Get shift charts}
\usage{
get_shift_charts(game = 2024020602)
}
\arguments{
\item{game}{integer Game ID}
}
\value{
tibble with one row per shift
}
\description{
\code{get_shift_charts()} retrieves information on each shift for a given \code{game}, including but not limited to their period, start and end times, and player's ID and name. Access \code{get_games()} for \code{game} reference.
}
\examples{
shift_charts_2024030411 <- get_shift_charts(game=2024030411)
}
