% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GameOld.R
\name{get_shift_charts}
\alias{get_shift_charts}
\title{Access the shift charts for a game}
\usage{
get_shift_charts(game = 2023030417)
}
\arguments{
\item{game}{integer ID (e.g., 2025020275); see \code{\link[=games]{games()}} for reference}
}
\value{
data.frame with one row per shift
}
\description{
\code{get_shift_charts()} is deprecated. Use \code{\link[=shifts]{shifts()}} instead.
}
