% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Players.R
\name{get_skater_leaders}
\alias{get_skater_leaders}
\title{Get skater stats leaders for a season}
\usage{
get_skater_leaders(
  season = 20242025,
  game_type = 2,
  category = "points",
  limit = -1
)
}
\arguments{
\item{season}{string Season in 'YYYYYYYY' e.g. 20242025}

\item{game_type}{integer 2=regular, 3=playoffs}
}
\value{
tibble with one row per skater
}
\description{
Get skater stats leaders for a season
}
