% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Skater.R
\name{get_skater_leaders}
\alias{get_skater_leaders}
\title{Get skater stats leaders for a season}
\usage{
get_skater_leaders(season = 20242025, game_type = 2, category = "points")
}
\arguments{
\item{season}{integer Season in YYYYYYYY}

\item{game_type}{integer GameType where 2=regular and 3=playoffs}

\item{category}{string e.g. assists, goals, goalsSh, goalsPp, points, penaltyMins, toi, plusMinus, faceoffLeaders}
}
\value{
tibble with one row per skater
}
\description{
Get skater stats leaders for a season
}
