% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Skater.R
\name{get_skater_milestones}
\alias{get_skater_milestones}
\title{Get skater milestones}
\usage{
get_skater_milestones()
}
\value{
tibble with one row per skater
}
\description{
\code{get_skater_milestones()} retrieves information on each skater close to a milestone, including but not limited to their ID, name, and statistics.
}
\examples{
skater_milestones <- get_skater_milestones()
}
