% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SkaterOld.R
\name{get_skater_statistics}
\alias{get_skater_statistics}
\title{Get skater statistics}
\usage{
get_skater_statistics()
}
\description{
\code{get_skater_statistics()} is defunct. Use \code{\link[=skater_statistics]{skater_statistics()}} instead.
}
