% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Skater.R
\name{get_skater_statistics}
\alias{get_skater_statistics}
\title{Get skater statistics by season}
\usage{
get_skater_statistics(
  season = 20242025,
  report = "summary",
  teams = 1:100,
  is_aggregate = F,
  is_game = F,
  dates = c("2025-01-01"),
  game_types = 1:3
)
}
\arguments{
\item{season}{integer Season in YYYYYYYY}

\item{report}{string Report e.g. 'summary' and 'bios' (forced to 'summary' if
\code{is_game}=T)}

\item{teams}{vector of integers TeamID(s)}

\item{is_aggregate}{boolean isAggregate where T=regular and playoffs combined
from multiple teams, if applicable}

\item{is_game}{boolean isGame where T=rows by games and F=rows by players}

\item{dates}{vector of strings Date(s) in 'YYYY-MM-DD' (only if paired with
\code{is_game})}
}
\value{
tibble with one row per skater or game
}
\description{
Get skater statistics by season
}
