% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Skater.R
\name{get_skaters}
\alias{get_skaters}
\title{Get skaters' biographies by range of seasons}
\usage{
get_skaters(start_season = 19171918, end_season = get_season_now()$seasonId)
}
\arguments{
\item{start_season}{integer in YYYYYYYY}

\item{end_season}{integer in YYYYYYYY}
}
\value{
tibble with one row per skater
}
\description{
\code{get_skaters()} retrieves information on each skater for a given set of \code{start_season} and \code{end_season}, including but not limited to their ID, name, bio-metrics, and career statistics. Access \code{get_seasons()} for \code{start_season} and \code{end_season} references. Will soon be deprecated as \code{get_players()} can list all players and their IDs much more efficiently.
}
\examples{
skaters_2000s <- get_skaters(start_season=20002001, end_season=20242025)
}
