% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Skater.R
\name{get_skaters}
\alias{get_skaters}
\title{Get all skaters' biographies from a range of years}
\usage{
get_skaters(start_year = 1917, end_year = 2025)
}
\arguments{
\item{start_year}{integer Year to start search}

\item{end_year}{integer Year to end search}
}
\value{
tibble with one row per skater
}
\description{
Get all skaters' biographies from a range of years
}
