% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Skater.R
\name{get_skaters}
\alias{get_skaters}
\title{Get all skaters' biographies from a range of seasons}
\usage{
get_skaters(start_season = 19171918, end_season = get_season_now()$seasonId)
}
\arguments{
\item{start_season}{integer Season to start search in YYYYYYYY}

\item{end_season}{integer Season to end search in YYYYYYYY}
}
\value{
tibble with one row per skater
}
\description{
Get all skaters' biographies from a range of seasons
}
