% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Player.R
\name{get_spotlight_players}
\alias{get_spotlight_players}
\title{Get 'spotlight' players as of now}
\usage{
get_spotlight_players()
}
\value{
tibble with one row per player
}
\description{
Get 'spotlight' players as of now
}
\examples{
spotlight_players_now <- get_spotlight_players()
}
