% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LeagueOld.R
\name{get_standings}
\alias{get_standings}
\title{Access the standings for a date}
\usage{
get_standings(date = "2025-01-01")
}
\arguments{
\item{date}{character in 'YYYY-MM-DD' (e.g., '2025-01-01'); see
\code{\link[=seasons]{seasons()}} for reference}
}
\value{
data.frame with one row per team
}
\description{
\code{get_standings()} is deprecated. Use \code{\link[=standings]{standings()}} instead.
}
