% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/League.R
\name{get_standings}
\alias{get_standings}
\title{Get standings by date}
\usage{
get_standings(date = "2025-01-01")
}
\arguments{
\item{date}{string Date in 'YYYY-MM-DD'}
}
\value{
tibble with one row per team
}
\description{
Get standings by date
}
