% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LeagueOld.R
\name{get_standings}
\alias{get_standings}
\title{Get standings by date}
\usage{
get_standings(date = "2025-01-01")
}
\description{
\code{get_standings()} is deprecated. Use \code{\link[=ns_standings]{ns_standings()}} instead.
}
