% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/League.R
\name{get_standings_information}
\alias{get_standings_information}
\title{Get standings information for all seasons}
\usage{
get_standings_information()
}
\value{
tibble with one row per season
}
\description{
\code{get_standings_information()} retrieves information on each season, including but not limited to their ID; start and end dates for standings; and conference-division, win-tie-loss, and wildcard regulations. May soon be merged with \code{get_seasons()}.
}
\examples{
standings_info <- get_standings_information()
}
