% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LeagueOld.R
\name{get_standings_information}
\alias{get_standings_information}
\title{Access the standings rules by season}
\usage{
get_standings_information()
}
\value{
data.frame with one row per season
}
\description{
\code{get_standings_information()} is deprecated. Use \code{\link[=standings_rules]{standings_rules()}}
instead.
}
