% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LeagueOld.R
\name{get_standings_information}
\alias{get_standings_information}
\title{Get standings information for all seasons}
\usage{
get_standings_information()
}
\description{
\code{get_standings_information()} is deprecated. Use
\code{\link[=ns_standings_information]{ns_standings_information()}} instead.
}
