% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OtherOld.R
\name{get_streams}
\alias{get_streams}
\title{Access all the streams}
\usage{
get_streams()
}
\value{
data.frame with one row per stream
}
\description{
\code{get_streams()} is deprecated. Use \code{\link[=streams]{streams()}} instead.
}
