% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Team.R
\name{get_team_prospects}
\alias{get_team_prospects}
\title{Get prospects by team and player-type}
\usage{
get_team_prospects(team = "BOS", player_type = "forwards")
}
\arguments{
\item{team}{string 3-letter Code}

\item{player_type}{string of 'forwards', 'defensemen', or 'goalies'}
}
\value{
tibble with one row per player
}
\description{
\code{get_team_prospects()} retrieves information on each prospect for a given set of \code{team} and \code{player_type}, including but not limited to their ID, name, bio-metrics, and birth date and location. Access \code{get_teams()} for \code{team} reference.
}
\examples{
COL_defensemen_prospects <- get_team_prospects(
  team='COL',
  player_type='defensemen'
)
}
