% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TeamOld.R
\name{get_team_prospects}
\alias{get_team_prospects}
\title{Access the prospects for a team and position}
\usage{
get_team_prospects(team = "NJD", player_type = "forwards")
}
\arguments{
\item{team}{three-letter code (e.g., 'COL'); see \code{\link[=teams]{teams()}} for reference}

\item{player_type}{character of 'forwards', 'defensemen', or 'goalies'}
}
\value{
data.frame with one row per player
}
\description{
\code{get_team_prospects()} is deprecated. Use \code{\link[=team_prospects]{team_prospects()}} instead.
}
