% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Team.R
\name{get_team_prospects}
\alias{get_team_prospects}
\title{Get team prospects by player-type}
\usage{
get_team_prospects(team = "BOS", player_type = "forwards")
}
\arguments{
\item{team}{string 3-letter team code}

\item{player_type}{string Player-type of 'forwards', 'defensemen', or 'goalies'}
}
\value{
tibble with one row per player
}
\description{
Get team prospects by player-type
}
