% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Team.R
\name{get_team_roster}
\alias{get_team_roster}
\title{Get roster by team, season, and player-type}
\usage{
get_team_roster(
  team = "BOS",
  season = get_season_now()$seasonId,
  player_type = "forwards"
)
}
\arguments{
\item{team}{string 3-letter team code}

\item{season}{integer Season in YYYYYYYY}

\item{player_type}{string Player-type of 'forwards', 'defensemen', or
'goalies'}
}
\value{
tibble with one row per player
}
\description{
Get roster by team, season, and player-type
}
\examples{
COL_defensemen_20242025 <- get_team_roster(
  team='COL',
  season=20242025,
  player_type='defensemen'
)
}
