% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TeamOld.R
\name{get_team_roster}
\alias{get_team_roster}
\title{Get the roster of a team for a season and position}
\usage{
get_team_roster(team = 1, season = "current", player_type = "forwards")
}
\description{
\code{get_team_roster()} is deprecated. Use \code{\link[=ns_roster]{ns_roster()}} instead.
}
