% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TeamOld.R
\name{get_team_roster_statistics}
\alias{get_team_roster_statistics}
\title{Access the roster statistics for a team, season, game type, and player type}
\usage{
get_team_roster_statistics(
  team = "NJD",
  season = "now",
  game_type = 2,
  player_type = "skaters"
)
}
\arguments{
\item{team}{three-letter code (e.g., 'COL'); see \code{\link[=teams]{teams()}} for reference}

\item{season}{integer in YYYYYYYY (e.g., 20242025); see \code{\link[=seasons]{seasons()}} for
reference}

\item{game_type}{integer in 1:3 (where 1 = pre-season, 2 = regular season, 3
= playoff/post-season) OR character of 'pre', 'regular', or
playoff'/'post'; see \code{\link[=seasons]{seasons()}} for reference; most functions will NOT
support pre-season}

\item{player_type}{character of 'skaters' or 'goalies'}
}
\value{
data.frame with one row per player
}
\description{
\code{get_team_roster_statistics()} is deprecated. Use \code{\link[=roster_statistics]{roster_statistics()}}
instead.
}
