% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Team.R
\name{get_team_roster_statistics}
\alias{get_team_roster_statistics}
\title{Get roster statistics by team, season, game-type, and player-type}
\usage{
get_team_roster_statistics(
  team = "BOS",
  season = get_season_now()$seasonId,
  game_type = 2,
  player_type = "skaters"
)
}
\arguments{
\item{team}{string 3-letter Code}

\item{season}{integer in YYYYYYYY}

\item{game_type}{integer where 2=regular and 3=playoffs}

\item{player_type}{string of 'skaters' or 'goalies'}
}
\value{
tibble with one row per player
}
\description{
\code{get_team_roster_statistics()} retrieves information on each player for a given set of \code{team}, \code{season}, \code{game_type} and \code{player_type}, including but not limited to their ID, name, and statistics. Access \code{get_teams()} for \code{team} and \code{get_team_seasons()} for \code{season} references.
}
\examples{
regular_COL_goalies_statistics_20242025 <- get_team_roster_statistics(
  team='COL',
  season=20242025,
  game_type=2,
  player_type='goalies'
)
}
