% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TeamOld.R
\name{get_team_roster_statistics}
\alias{get_team_roster_statistics}
\title{Get the statistics of a roster for a season, game type, and position}
\usage{
get_team_roster_statistics(
  team = 1,
  season = "now",
  game_type = 2,
  player_type = "skaters"
)
}
\description{
\code{get_team_roster_statistics()} is deprecated. Use \code{\link[=ns_roster_statistics]{ns_roster_statistics()}}
instead.
}
