% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Team.R
\name{get_team_schedule}
\alias{get_team_schedule}
\title{Get schedule by team and season}
\usage{
get_team_schedule(team = "BOS", season = get_season_now()$seasonId)
}
\arguments{
\item{team}{string 3-letter team code}

\item{season}{integer Season in YYYYYYYY}
}
\value{
tibble with one row per game
}
\description{
Get schedule by team and season
}
\examples{
COL_schedule_20242025 <- get_team_schedule(team='COL', season=20242025)
}
