% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TeamOld.R
\name{get_team_schedule}
\alias{get_team_schedule}
\title{Get the schedule of a team for a season}
\usage{
get_team_schedule(team = 1, season = "now")
}
\description{
\code{get_team_schedule()} is deprecated. Use \code{\link[=ns_team_schedule_season]{ns_team_schedule_season()}} instead.
}
