% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TeamOld.R
\name{get_team_schedule}
\alias{get_team_schedule}
\title{Access the schedule for a team and season}
\usage{
get_team_schedule(team = "NJD", season = "now")
}
\arguments{
\item{team}{three-letter code (e.g., 'COL'); see \code{\link[=ns_teams]{ns_teams()}} for reference}

\item{season}{integer in YYYYYYYY (e.g., 20242025); see \code{\link[=ns_seasons]{ns_seasons()}} for
reference}
}
\value{
data.frame with one row per game
}
\description{
\code{get_team_schedule()} is deprecated. Use \code{\link[=ns_team_season_schedule]{ns_team_season_schedule()}} instead.
}
