% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Team.R
\name{get_team_scoreboard}
\alias{get_team_scoreboard}
\title{Get team scoreboard as of now}
\usage{
get_team_scoreboard(team = "BOS")
}
\arguments{
\item{team}{string 3-letter team code}
}
\value{
tibble with one row per game
}
\description{
Get team scoreboard as of now
}
\examples{
FLA_scoreboard_now <- get_team_scoreboard(team='FLA')
}
