% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Team.R
\name{get_team_scoreboard}
\alias{get_team_scoreboard}
\title{Get team scoreboard now}
\usage{
get_team_scoreboard(team = "BOS")
}
\arguments{
\item{team}{string Team in 3-letter code}
}
\value{
???
}
\description{
Get team scoreboard now
}
