% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Team.R
\name{get_team_seasons}
\alias{get_team_seasons}
\title{Get season(s) for which team played in regular season and/or playoffs}
\usage{
get_team_seasons(team = "BOS")
}
\arguments{
\item{team}{3-letter Code}
}
\value{
tibble with one row per season
}
\description{
\code{get_team_seasons()} retrieves information on each season for a given \code{team}, including but not limited to their ID and game-type(s). Access \code{get_teams()} for \code{team} reference.
}
\examples{
COL_seasons <- get_team_seasons(team='COL')
}
