% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TeamOld.R
\name{get_team_seasons}
\alias{get_team_seasons}
\title{Access the season(s) and game type(s) in which a team played}
\usage{
get_team_seasons(team = "NJD")
}
\arguments{
\item{team}{three-letter code (e.g., 'COL'); see \code{\link[=teams]{teams()}} for reference}
}
\value{
data.frame with one row per season
}
\description{
\code{get_team_seasons()} is deprecated. Use \code{\link[=team_seasons]{team_seasons()}} instead.
}
