% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TeamOld.R
\name{get_team_seasons}
\alias{get_team_seasons}
\title{Get the season(s) and game type(s) of which a team played in}
\usage{
get_team_seasons(team = 1)
}
\description{
\code{get_team_seasons()} is deprecated. Use \code{\link[=ns_team_seasons]{ns_team_seasons()}} instead.
}
