% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TeamOld.R
\name{get_team_statistics}
\alias{get_team_statistics}
\title{Access various reports for all the teams by season or game}
\usage{
get_team_statistics()
}
\description{
\code{get_team_statistics()} is defunct. Use \code{\link[=team_season_report]{team_season_report()}} and/or
\code{\link[=team_game_report]{team_game_report()}} instead.
}
