% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TeamOld.R
\name{get_team_statistics}
\alias{get_team_statistics}
\title{Get team statistics}
\usage{
get_team_statistics(
  season = 20242025,
  report = "summary",
  is_aggregate = FALSE,
  is_game = FALSE,
  dates = c("2025-01-01"),
  game_types = 1:3
)
}
\description{
\code{get_team_statistics()} is defunct. Use \code{\link[=ns_team_statistics]{ns_team_statistics()}} instead.
}
