% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Team.R
\name{get_team_statistics}
\alias{get_team_statistics}
\title{Get team statistics by season}
\usage{
get_team_statistics(
  season = 20242025,
  is_game = F,
  dates = c("2025-01-01"),
  game_types = 1:3
)
}
\arguments{
\item{season}{integer Season in YYYYYYYY}

\item{is_game}{boolean isGame where T=rows by games and F=rows by players}

\item{dates}{vector of strings Date(s) in 'YYYY-MM-DD' (only if paired with
\code{is_game})}

\item{is_aggregate}{boolean isAggregate where T=regular and playoffs combined
from multiple teams, if applicable}
}
\value{
tibble with one row per skater or game
}
\description{
Get team statistics by season
}
