% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Team.R
\name{get_team_statistics}
\alias{get_team_statistics}
\title{Get team statistics by season}
\usage{
get_team_statistics(
  season = get_season_now()$seasonId,
  report = "summary",
  is_aggregate = FALSE,
  is_game = FALSE,
  dates = c("2025-01-01"),
  game_types = 1:3
)
}
\arguments{
\item{season}{integer in YYYYYYYY}

\item{report}{string}

\item{is_aggregate}{boolean}

\item{is_game}{boolean}

\item{dates}{vector of strings in 'YYYY-MM-DD' *only if paired with \code{is_game}}

\item{game_types}{vector of integers where 1=pre-season, 2=regular, and
3=playoffs}
}
\value{
tibble with one row per team or game
}
\description{
Get team statistics by season
}
\examples{
playoff_team_stf_20242025 <- get_team_statistics(
  season=20242025,
  report='scoretrailfirst',
  game_types=c(3)
)

}
