% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Team.R
\name{get_team_statistics}
\alias{get_team_statistics}
\title{Get team statistics by season, game-type, and player-type}
\usage{
get_team_statistics(
  team = "BOS",
  season = 20242025,
  game_type = 2,
  player_type = "skaters"
)
}
\arguments{
\item{team}{string Team in 3-letter code}

\item{season}{integer Season in YYYYYYYY}

\item{game_type}{integer Game-type where 2=regular and 3=playoffs}

\item{player_type}{string Player-type of 'skaters' or 'goalies'}
}
\value{
tibble with one row per player
}
\description{
Get team statistics by season, game-type, and player-type
}
