% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TeamOld.R
\name{get_team_statistics}
\alias{get_team_statistics}
\title{Get various reports for all the teams by season or game}
\usage{
get_team_statistics(
  season = 20242025,
  report = "summary",
  is_aggregate = FALSE,
  is_game = FALSE,
  dates = c("2025-01-01"),
  game_types = 1:3
)
}
\description{
\code{get_team_statistics()} is defunct. Use \code{\link[=ns_team_season_report]{ns_team_season_report()}} or
\code{\link[=ns_team_game_report]{ns_team_game_report()}} instead.
}
