% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TeamOld.R
\name{get_teams}
\alias{get_teams}
\title{Access all the teams}
\usage{
get_teams()
}
\value{
data.frame with one row per team
}
\description{
\code{get_teams()} is deprecated. Use \code{\link[=teams]{teams()}} instead.
}
