% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OtherOld.R
\name{get_tv_schedule}
\alias{get_tv_schedule}
\title{Access the NHL Network TV schedule for a date}
\usage{
get_tv_schedule(date = "now")
}
\arguments{
\item{date}{character in 'YYYY-MM-DD' (e.g., '2025-01-01'); see
\code{\link[=seasons]{seasons()}} for reference}
}
\value{
data.frame with one row per program
}
\description{
\code{get_tv_schedule()} is deprecated. Use \code{\link[=tv_schedule]{tv_schedule()}} instead.
}
