% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Other.R
\name{get_tv_schedule}
\alias{get_tv_schedule}
\title{Get TV schedule by date}
\usage{
get_tv_schedule(date = "2025-01-01")
}
\arguments{
\item{date}{string Date in 'YYYY-MM-DD'}
}
\value{
tibble with one row per program
}
\description{
Get TV schedule by date
}
\examples{
tv_schedule_2025_01_02 <- get_tv_schedule(date='2025-01-02')
}
