% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Other.R
\name{get_venues}
\alias{get_venues}
\title{Get all venues}
\usage{
get_venues()
}
\value{
tibble with one row per venue
}
\description{
\code{get_venues()} retrieves information on each venue, including but not limited to their ID, name, and location.
}
\examples{
all_venues <- get_venues()
}
