% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LeagueOld.R
\name{get_venues}
\alias{get_venues}
\title{Access all the venues}
\usage{
get_venues()
}
\value{
data.frame with one row per venue
}
\description{
\code{get_venues()} is deprecated. Use \code{\link[=ns_venues]{ns_venues()}} instead.
}
