% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GameOld.R
\name{get_wsc_play_by_play}
\alias{get_wsc_play_by_play}
\title{Access the World Showcase (WSC) play-by-play for a game}
\usage{
get_wsc_play_by_play(game = 2023030417)
}
\arguments{
\item{game}{integer ID (e.g., 2025020275); see \code{\link[=games]{games()}} for reference}
}
\value{
data.frame with one row per event (play)
}
\description{
\code{get_wsc_play_by_play()} is deprecated. Use \code{\link[=wsc_play_by_play]{wsc_play_by_play()}} instead.
}
