% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GameOld.R
\name{get_wsc_play_by_play}
\alias{get_wsc_play_by_play}
\title{Get the World Showcase (WSC) play-by-play of a game}
\usage{
get_wsc_play_by_play(game = 2025020275)
}
\description{
\code{get_wsc_play_by_play()} is deprecated. Use \code{\link[=ns_wsc_play_by_play]{ns_wsc_play_by_play()}} instead.
}
