% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Game.R
\name{get_wsc_play_by_play}
\alias{get_wsc_play_by_play}
\title{Get World Showcase (WSC) play-by-play by game}
\usage{
get_wsc_play_by_play(game_id = 2024020602)
}
\arguments{
\item{game_id}{integer Game ID}
}
\value{
tibble with one row per play
}
\description{
Get World Showcase (WSC) play-by-play by game
}
