% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Other.R
\name{glossary}
\alias{glossary}
\title{Access the glossary}
\usage{
glossary()
}
\value{
data.frame with one row per terminology
}
\description{
\code{glossary()} retrieves the glossary as a \code{data.frame} where each row represents terminology and includes detail on reference definitions and rules-framework information.
}
\examples{
glossary <- glossary()
}
