% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GoalieEDGE.R
\name{goalie_edge_seasons}
\alias{goalie_edge_seasons}
\title{Access the season(s) and game type(s) in which there exists goalie EDGE
statistics}
\usage{
goalie_edge_seasons()
}
\value{
data.frame with one row per season
}
\description{
\code{goalie_edge_seasons()} retrieves the season(s) and game type(s) in which there exists goalie EDGE statistics as a \code{data.frame} where each row represents season and includes detail on date/season filtering windows and chronological context plus NHL EDGE style tracking outputs and relative-performance context.
}
\examples{
goalie_EDGE_seasons <- goalie_edge_seasons()
}
