% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GoalieEDGE.R
\name{goalie_edge_seasons}
\alias{goalie_edge_seasons}
\title{Get the season(s) and game type(s) for which the NHL recorded goalie EDGE
statistics}
\usage{
goalie_edge_seasons()
}
\value{
data.frame with one row per season
}
\description{
\code{goalie_edge_seasons} retrieves information on each
}
\examples{
goalie_EDGE_seasons <- goalie_edge_seasons()
}
