% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GoalieEDGE.R
\name{goalie_edge_shot_location}
\alias{goalie_edge_shot_location}
\title{Get the EDGE shot location statistics of a goalie for a season, game type,
and report type}
\usage{
goalie_edge_shot_location(
  player = 8478406,
  season = "now",
  game_type = "",
  report_type = "details"
)
}
\arguments{
\item{player}{integer ID (e.g., 8478406)}

\item{season}{integer in YYYYYYYY (e.g., 20242025)}

\item{game_type}{integer in 1:3 (where 1 = pre-season, 2 = regular season, 3
= playoff/post-season) OR character of 'pre', 'regular', or 'playoff'/'post'}

\item{report_type}{character of 'd'/details' or 't'/'totals'}
}
\value{
data.frame with one row per shot location
}
\description{
\code{goalie_edge_shot_location()} retrieves information on each
}
\examples{
Mackenzie_Blackwood_shot_location_totals_regular_20242025 <- 
  goalie_edge_shot_location(
    player      = 8478406,
    season      = 20242025,
    game_type   = 2,
    report_type = 'T'
  )
}
