% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Goalie.R
\name{goalie_game_scoring}
\alias{goalie_game_scoring}
\title{Access the scoring statistics for all the goalies by game}
\usage{
goalie_game_scoring()
}
\value{
data.frame with one row per player per game
}
\description{
\code{goalie_game_scoring()} scrapes the scoring statistics for all the goalies
by game.
}
\examples{
goalie_game_scoring <- goalie_game_scoring()
}
