% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Goalie.R
\name{goalie_game_statistics}
\alias{goalie_game_statistics}
\alias{goalie_game_stats}
\title{Access the statistics for all the goalies by game}
\usage{
goalie_game_statistics()

goalie_game_stats()
}
\value{
data.frame with one row per goalie per game
}
\description{
\code{goalie_game_statistics()} scrapes the statistics for all the goalies by
game.
}
\examples{
\donttest{goalie_game_stats <- goalie_game_statistics()}
}
