% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Goalie.R
\name{goalie_milestones}
\alias{goalie_milestones}
\title{Access the goalies on milestone watch}
\usage{
goalie_milestones()
}
\value{
data.frame with one row per player
}
\description{
\code{goalie_milestones()} scrapes the goalies on milestone watch.
}
\examples{
goalie_milestones <- goalie_milestones()
}
