% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Goalie.R
\name{goalie_milestones}
\alias{goalie_milestones}
\title{Access the goalies on milestone watch}
\usage{
goalie_milestones()
}
\value{
data.frame with one row per player
}
\description{
\code{goalie_milestones()} retrieves the goalies on milestone watch as a \code{data.frame} where each row represents player and includes detail on date/season filtering windows and chronological context, player identity, role, handedness, and biographical profile, and ranking movement, points pace, and division/conference position signals.
}
\examples{
goalie_milestones <- goalie_milestones()
}
