% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Goalie.R
\name{goalie_milestones}
\alias{goalie_milestones}
\title{Get the goalies on milestone watch}
\usage{
goalie_milestones()
}
\value{
data.frame with one row per goalie
}
\description{
\code{get_goalie_milestones()} retrieves information on each goalie close to a
milestone, including but not limited to their ID, name, and statistics.
}
\examples{
goalie_milestones <- goalie_milestones()
}
