% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Goalie.R
\name{goalie_report_configurations}
\alias{goalie_report_configurations}
\alias{goalie_report_configs}
\title{Access the configurations for goalie reports}
\usage{
goalie_report_configurations()

goalie_report_configs()
}
\value{
list with various items
}
\description{
\code{goalie_report_configurations()} scrapes the configurations for
\code{\link[=goalie_season_report]{goalie_season_report()}} and \code{\link[=goalie_game_report]{goalie_game_report()}}.
}
\examples{
goalie_report_configs <- goalie_report_configurations()
}
