% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Goalie.R
\name{goalie_scoring}
\alias{goalie_scoring}
\title{Access the career scoring statistics for all the goalies}
\usage{
goalie_scoring()
}
\value{
data.frame with one row per player
}
\description{
\code{goalie_scoring()} retrieves the career scoring statistics for all the goalies as a \code{data.frame} where each row represents player and includes detail on date/season filtering windows and chronological context, team identity, affiliation, and matchup-side context, and player identity, role, handedness, and biographical profile.
}
\examples{
goalie_scoring <- goalie_scoring()
}
