% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Goalie.R
\name{goalie_season_statistics}
\alias{goalie_season_statistics}
\alias{goalie_season_stats}
\title{Access the statistics for all the goalies by season, game type, and team.}
\usage{
goalie_season_statistics()

goalie_season_stats()
}
\value{
data.frame with one row per player per season per game type,
separated by team if applicable
}
\description{
\code{goalie_season_statistics()} retrieves the statistics for all the goalies by season, game type, and team as a \code{data.frame} where each row represents player per season per game type, separated by team if applicable and includes detail on date/season filtering windows and chronological context, team identity, affiliation, and matchup-side context, and player identity, role, handedness, and biographical profile.
}
\examples{
goalie_season_stats <- goalie_season_statistics()
}
