% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Goalie.R
\name{goalie_season_statistics}
\alias{goalie_season_statistics}
\alias{goalie_season_stats}
\title{Get the goalie statistics for all the seasons}
\usage{
goalie_season_statistics()

goalie_season_stats()
}
\value{
data.frame with one row per goalie per season
}
\description{
\code{goalie_season_statistics()} retrieves information on ...
}
\examples{
goalie_season_stats <- goalie_season_statistics()
}
