% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Goalie.R
\name{goalie_series_statistics}
\alias{goalie_series_statistics}
\alias{goalie_series_stats}
\title{Get the statistics for all the goalies by playoff series}
\usage{
goalie_series_statistics()

goalie_series_stats()
}
\value{
data.frame with one row per goalie per series
}
\description{
\code{goalie_series_statistics()} retrieves information on ...
}
\examples{
goalie_series_stats <- goalie_series_statistics()
}
