% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Goalie.R
\name{goalie_statistics}
\alias{goalie_statistics}
\alias{goalie_stats}
\title{Access the career statistics for all the goalies}
\usage{
goalie_statistics()

goalie_stats()
}
\value{
data.frame with one row per player
}
\description{
\code{goalie_statistics()} retrieves the career statistics for all the goalies as a \code{data.frame} where each row represents player and includes detail on team identity, affiliation, and matchup-side context, player identity, role, handedness, and biographical profile, and production, workload, efficiency, and result-level performance outcomes.
}
\examples{
goalie_stats <- goalie_statistics()
}
