% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Share.R
\name{ig_game_cumulative_expected_goals}
\alias{ig_game_cumulative_expected_goals}
\alias{ig_game_cum_xG}
\title{Save an Instagram (IG) share-able cumulative expected goals (xG) time-series
plot for a game}
\usage{
ig_game_cumulative_expected_goals(game = 2023030417, model = NULL, save = TRUE)

ig_game_cum_xG(game = 2023030417, model = NULL)
}
\arguments{
\item{game}{integer ID (e.g., 2025020275); see \code{\link[=games]{games()}} for reference}

\item{model}{deprecated legacy model selector; ignored}

\item{save}{logical only FALSE for tests}
}
\value{
\code{NULL}
}
\description{
\code{ig_game_cumulative_expected_goals()} saves an IG share-able cumulative xG
time-series plot for a given \code{game} as a PNG.
}
\examples{
# May take >5s, so skip.
\donttest{ig_game_cumulative_expected_goals(
  game  = 2023030417, 
  save  = FALSE
)}
}
