% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Share.R
\name{ig_game_shot_locations}
\alias{ig_game_shot_locations}
\alias{ig_game_shot_locs}
\title{Save an Instagram (IG) share-able shot-location plot for a game}
\usage{
ig_game_shot_locations(
  game = 2023030417,
  team = "home",
  model = NULL,
  save = TRUE
)

ig_game_shot_locs(game = 2023030417, team = "home", model = NULL)
}
\arguments{
\item{game}{integer ID (e.g., 2025020275); see \code{\link[=games]{games()}} for reference}

\item{team}{character of 'h'/'home' or 'a'/'away'}

\item{model}{deprecated legacy model selector; ignored}

\item{save}{logical only FALSE for tests}
}
\value{
\code{NULL}
}
\description{
\code{ig_game_shot_locations()} saves an IG share-able shot location plot for a
given \code{game}.
}
\examples{
# May take >5s, so skip.
\donttest{ig_game_shot_locations(
  game  = 2023030417, 
  team  = 'H', 
  save  = FALSE
)}
}
