% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Other.R
\name{location}
\alias{location}
\title{Access the location for a zip code}
\usage{
location(zip = 10001)
}
\arguments{
\item{zip}{integer (e.g., 48304)}
}
\value{
data.frame with one row per team
}
\description{
\code{location()} retrieves the location for a zip code as a \code{data.frame} where each row represents team and includes detail on venue/location geography and regional metadata.
}
\examples{
Cranbrook_Schools <- location(48304)
}
