% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Draft.R
\name{lottery_odds}
\alias{lottery_odds}
\title{Access the draft lottery odds}
\usage{
lottery_odds()
}
\value{
data.frame with one row per draft lottery
}
\description{
\code{lottery_odds()} scrapes the draft lottery odds.
}
\examples{
lottery_odds <- lottery_odds()
}
