% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Draft.R
\name{lottery_odds}
\alias{lottery_odds}
\title{Access the draft lottery odds}
\usage{
lottery_odds()
}
\value{
data.frame with one row per draft lottery
}
\description{
\code{lottery_odds()} retrieves the draft lottery odds as a \code{data.frame} where each row represents draft lottery and includes detail on draft-cycle evaluation, ranking, and selection tracking detail.
}
\examples{
lottery_odds <- lottery_odds()
}
