% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utility.R
\name{nhl_api}
\alias{nhl_api}
\title{Call the NHL API with 429 (rate limit) error-handling}
\usage{
nhl_api(path, query = list(), type)
}
\arguments{
\item{path}{character}

\item{query}{list}

\item{type}{character of 'w' for web, 's' for stats, and 'r' for records}
}
\value{
parsed JSON (i.e., data.frame or list)
}
\description{
Call the NHL API with 429 (rate limit) error-handling
}
\keyword{internal}
