% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utility.R
\name{nhl_api}
\alias{nhl_api}
\title{Internal: call NHL API}
\usage{
nhl_api(path, query = list(), stats_rest = FALSE)
}
\arguments{
\item{path}{String API path}

\item{query}{list Query parameters}

\item{stats_rest}{boolean isStatsRest where TRUE=NHL's REST API and FALSE=
NHL's standard API}
}
\value{
parsed JSON
}
\description{
Internal: call NHL API
}
\keyword{internal}
