% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utils.R
\name{nhl_api}
\alias{nhl_api}
\title{Internal: call NHL API}
\usage{
nhl_api(path, query = list(), stats_rest = F)
}
\arguments{
\item{path}{string API path}

\item{query}{list Query parameters}
}
\value{
parsed JSON
}
\description{
Internal: call NHL API
}
\keyword{internal}
