% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Clean.R
\name{normalize_coordinates}
\alias{normalize_coordinates}
\title{Normalize the x and y coordinates for all the events (plays) in a
play-by-play}
\usage{
normalize_coordinates(
  data,
  is_home_name = "isHome",
  home_team_defending_side_name = "homeTeamDefendingSide",
  x_coordinate_name = "xCoord",
  y_coordinate_name = "yCoord",
  x_coordinate_normalized_name = "xCoordNorm",
  y_coordinate_normalized_name = "yCoordNorm"
)
}
\arguments{
\item{data}{data.frame of play-by-play(s); see \code{\link[=gc_play_by_play]{gc_play_by_play()}} and/or
\code{\link[=wsc_play_by_play]{wsc_play_by_play()}} for reference}

\item{is_home_name}{name of column that contains home/not logical
indicator; see \code{\link[=flag_is_home]{flag_is_home()}} for reference}

\item{home_team_defending_side_name}{name of column that contains home team
defending side}

\item{x_coordinate_name}{name of column that contains x coordinate}

\item{y_coordinate_name}{name of column that contains y coordinate}

\item{x_coordinate_normalized_name}{name of column that you want contain
normalized x coordinate}

\item{y_coordinate_normalized_name}{name of column that you want contain
normalized y coordinate}
}
\value{
data.frame with one row per event (play)
}
\description{
\code{normalize_coordinates()} normalizes the x and y coordinates for all the
events (plays) in a play-by-play such that they all attack towards +x.
}
\examples{
# May take >5s, so skip.
\donttest{
  test                   <- gc_play_by_play()
  test_is_home_flagged   <- flag_is_home(test)
  test_coords_normalized <- normalize_coordinates(test_is_home_flagged)
}
}
