% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/League.R
\name{ns_attendance}
\alias{ns_attendance}
\title{Access the attendance by season and game type}
\usage{
ns_attendance()
}
\value{
data.frame with one row per season
}
\description{
\code{ns_attendance()} scrapes the attendance by season and game type.
}
\examples{
all_attendance <- ns_attendance()
}
