% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/League.R
\name{ns_attendance}
\alias{ns_attendance}
\title{Get the attendance for all the seasons}
\usage{
ns_attendance()
}
\value{
data.frame with one row per season
}
\description{
\code{ns_attendance()} retrieves information on each season, including but not
limited to their ID and regular and playoff attendance.
}
\examples{
all_attendance <- ns_attendance()
}
