% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Award.R
\name{ns_award_winners}
\alias{ns_award_winners}
\title{Access all the award winners/finalists}
\usage{
ns_award_winners()
}
\value{
data.frame with one row per winner/finalist
}
\description{
\code{ns_award_winners()} scrapes information on all the award winners/finalists.
}
\examples{
all_award_winners <- ns_award_winners()
}
