% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Award.R
\name{ns_awards}
\alias{ns_awards}
\title{Access all the awards}
\usage{
ns_awards()
}
\value{
data.frame with one row per award
}
\description{
\code{ns_awards()} scrapes information on all the awards.
}
\examples{
all_awards <- ns_awards()
}
