% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Game.R
\name{ns_boxscore}
\alias{ns_boxscore}
\title{Get the boxscore of a game for a team and position}
\usage{
ns_boxscore(game = 2025020275, team = "home", position = "forwards")
}
\arguments{
\item{game}{integer ID (e.g., 2025020275)}

\item{team}{character of 'h'/'home' or 'a'/'away'}

\item{position}{character of 'f'/'forwards', 'd'/'defensemen', or
'g'/goalies'}
}
\value{
data.frame with one row per player
}
\description{
\code{ns_boxscore()} retrieves information on each player for a given set
of \code{game}, \code{team}, and \code{player_type}, including but not limited to their ID,
name, sweater number, goals, assists, +/-, hits, blocks, shots-on-goal,
giveaways, takeaways, time on ice, and number of shifts. Access
\code{get_games()} for \code{game} reference.
}
\examples{
boxscore_SCF_game_7_20232024_FLA_defensemen <- ns_boxscore(
  game     = 2023030417,
  team     = 'A',
  position = 'D'
)
}
