% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Playoff.R
\name{ns_bracket}
\alias{ns_bracket}
\title{Get the playoff bracket of a season}
\usage{
ns_bracket(season = ns_season())
}
\arguments{
\item{season}{integer in YYYYYYYY (e.g., 20242025)}
}
\value{
data.frame with one row per series
}
\description{
\code{ns_bracket()} retrieves information on each series for a given \code{season},
including but not limited to their title, abbreviation, 1-letter code,
round, top and bottom seeds, and winning and losing teams' IDs. Access
\code{get_seasons()} for \code{season} reference.
}
\examples{
bracket_20242025 <- ns_bracket(season = 20242025)
}
