% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Other.R
\name{ns_countries}
\alias{ns_countries}
\title{Get all the countries}
\usage{
ns_countries()
}
\value{
data.frame with one row per country
}
\description{
\code{ns_countries} retrieves information on each country, including but not
limited to their ID, name, 2-letter code, and 3-letter code.
}
\examples{
all_countries <- ns_countries()
}
