% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Draft.R
\name{ns_draft_picks}
\alias{ns_draft_picks}
\title{Get all the draft picks}
\usage{
ns_draft_picks()
}
\value{
data.frame with one row per pick
}
\description{
\code{ns_draft_picks()} retrieves information on each selection, including but
not limited to their player ID, name, draft year, overall number,
bio-metrics, and the pick's team history.
}
\examples{
all_draft_picks <- ns_draft_picks()
}
