% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Draft.R
\name{ns_draft_prospects}
\alias{ns_draft_prospects}
\title{Get all the draft prospects}
\usage{
ns_draft_prospects()
}
\value{
data.frame with one row per prospect
}
\description{
\code{ns_draft_prospects()} retrieves information on ...
}
\examples{
# This may take >5s, so skip.
\donttest{all_draft_prospects <- ns_draft_prospects()}
}
