% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Draft.R
\name{ns_draft_tracker}
\alias{ns_draft_tracker}
\title{Access the real-time draft tracker}
\usage{
ns_draft_tracker()
}
\value{
data.frame with one row per player
}
\description{
\code{ns_draft_tracker()} scrapes the real-time draft tracker.
}
\examples{
draft_tracker <- ns_draft_tracker()
}
