% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Draft.R
\name{ns_drafts}
\alias{ns_drafts}
\title{Get all the drafts}
\usage{
ns_drafts()
}
\value{
data.frame with one row per draft
}
\description{
\code{ns_drafts()} retrieves information on each draft, including but not
limited to their year, type, venue, minimum and maximum player ages, and
number of rounds and picks.
}
\examples{
all_drafts <- ns_drafts()
}
