% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Event.R
\name{ns_event_replay}
\alias{ns_event_replay}
\title{Access the replay for an event}
\usage{
ns_event_replay(game = 2023030417, event = 866)
}
\arguments{
\item{game}{integer ID (e.g., 2025020262); see \code{\link[=ns_games]{ns_games()}} for reference}

\item{event}{integer ID (e.g., 751); see \code{\link[=gc_play_by_play]{gc_play_by_play()}} and/or
\code{\link[=gc_wsc_play_by_play]{gc_wsc_play_by_play()}} for reference; must be a 'goal' event}
}
\value{
data.frame with one row per decisecond
}
\description{
\code{ns_event_replay()} scrapes the replay for a given \code{event}.
}
\examples{
Gabriel_Landeskog_first_regular_goal_back_replay <- ns_event_replay(
  game  = 2025020262,
  event = 751
)
}
