% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Draft.R
\name{ns_expansion_draft_picks}
\alias{ns_expansion_draft_picks}
\title{Access all the expansion draft picks}
\usage{
ns_expansion_draft_picks()
}
\value{
data.frame with one row per pick
}
\description{
\code{ns_expansion_draft_picks()} scrapes information on all the expansion draft
picks
}
\examples{
all_expansion_draft_picks <- ns_expansion_draft_picks()
}
