% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Franchise.R
\name{ns_franchise_season_statistics}
\alias{ns_franchise_season_statistics}
\alias{ns_franchise_season_stats}
\title{Get statistics for all the franchises by season and game type}
\usage{
ns_franchise_season_statistics()

ns_franchise_season_stats()
}
\value{
data.frame with one row per franchise per game type per season
}
\description{
\code{ns_franchise_season_statistics()} returns information on statistics for all the franchises by season and game type, including but not limited to each franchise's ID, wins, ties, and losses.
}
\examples{
franchise_season_statistics <- ns_franchise_season_statistics()
}
