% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Franchise.R
\name{ns_franchise_statistics}
\alias{ns_franchise_statistics}
\alias{ns_franchise_stats}
\title{Access the all-time statistics for all the franchises by game type}
\usage{
ns_franchise_statistics()

ns_franchise_stats()
}
\value{
data.frame with one row per franchise per game type
}
\description{
\code{ns_franchise_statistics()} scrapes the all-time statistics for all the
franchises by game type.
}
\examples{
franchise_stats <- ns_franchise_statistics()
}
