% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Franchise.R
\name{ns_franchise_versus_franchise}
\alias{ns_franchise_versus_franchise}
\alias{ns_franchise_vs_franchise}
\title{Get all the franchises' all-time records versus other franchises}
\usage{
ns_franchise_versus_franchise()

ns_franchise_vs_franchise()
}
\value{
data.frame with one row per franchise versus another franchise
}
\description{
\code{ns_franchise_versus_franchise()} retrieves information on each franchise
versus another franchise, including but not limited to their IDs, game-type
ID, and all-time statistics.
}
\examples{
franchise_vs_franchise <- ns_franchise_versus_franchise()
}
