% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Franchise.R
\name{ns_franchises}
\alias{ns_franchises}
\title{Access all the franchises}
\usage{
ns_franchises()
}
\value{
data.frame with one row per franchise
}
\description{
\code{ns_franchises()} scrapes information on all the franchises.
}
\examples{
all_franchises <- ns_franchises()
}
