% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Franchise.R
\name{ns_franchises}
\alias{ns_franchises}
\title{Get all the franchises}
\usage{
ns_franchises()
}
\value{
data.frame with one row per franchise
}
\description{
\code{ns_franchises()} retrieves information on each franchise, including but
not limited to their ID; first and last seasons' IDs; captain, coach, and
general manager histories; and retired numbers.
}
\examples{
all_franchises <- ns_franchises()
}
