% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Game.R
\name{ns_game_rosters}
\alias{ns_game_rosters}
\title{Get the rosters of a game}
\usage{
ns_game_rosters(game = 2025020275)
}
\arguments{
\item{game}{integer ID (e.g., 2025020275)}
}
\value{
data.frame with one row per player
}
\description{
\code{ns_game_rosters()} retrieves ...
}
\examples{
rosters_SCF_game_7_20232024 <- ns_game_rosters(game = 2023030417)
}
