% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/League.R
\name{ns_game_type}
\alias{ns_game_type}
\title{Get the game type as of now}
\usage{
ns_game_type()
}
\value{
integer in 1:3 (where 1 = pre-season, 2 = regular season, 3
= playoff/post-season)
}
\description{
\code{ns_game_type()} returns the ID of the current game type (i.e., whether we are in pre-season, regular season, or the playoffs).
}
\examples{
game_type_now <- ns_game_type()
}
