% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Game.R
\name{ns_games}
\alias{ns_games}
\title{Get all the games}
\usage{
ns_games()
}
\value{
data.frame with one row per game
}
\description{
\code{ns_games()} retrieves information on each game, including but not limited
to their ID, season, type, start date and time, and home and visiting teams'
IDs and scores.
}
\examples{
# This may take >5s, so skip.
\donttest{
  all_games <- ns_games()
}
}
