% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Other.R
\name{ns_glossary}
\alias{ns_glossary}
\title{Access the glossary}
\usage{
ns_glossary()
}
\value{
data.frame with one row per terminology
}
\description{
\code{ns_glossary()} scrapes the glossary.
}
\examples{
glossary <- ns_glossary()
}
