% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Other.R
\name{ns_location}
\alias{ns_location}
\title{Access the location for a zip code}
\usage{
ns_location(zip = 10001)
}
\arguments{
\item{zip}{integer (e.g., 48304)}
}
\value{
data.frame with one row per team
}
\description{
\code{ns_location()} scrapes the location for a given \code{zip} code.
}
\examples{
Cranbrook_Schools <- ns_location(48304)
}
