% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Draft.R
\name{ns_lottery_odds}
\alias{ns_lottery_odds}
\title{Access the draft lottery odds}
\usage{
ns_lottery_odds()
}
\value{
data.frame with one row per draft lottery
}
\description{
\code{ns_lottery_odds()} scrapes the draft lottery odds.
}
\examples{
lottery_odds <- ns_lottery_odds()
}
