% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Other.R
\name{ns_officials}
\alias{ns_officials}
\title{Get all the officials}
\usage{
ns_officials()
}
\value{
data.frame with one row per official
}
\description{
\code{ns_officials()} retrieves information on each official, including but not
limited to their ID, name, and birth date and location.
}
\examples{
all_officials <- ns_officials()
}
