% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Event.R
\name{ns_penalty_shots}
\alias{ns_penalty_shots}
\alias{ns_ps}
\title{Get all the penalty shots}
\usage{
ns_penalty_shots()

ns_ps()
}
\value{
data.frame with one row per team
}
\description{
\code{ns_penalty_shots()} retrieves ...
}
\examples{
all_ps <- ns_penalty_shots()
}
