% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Player.R
\name{ns_player_game_log}
\alias{ns_player_game_log}
\title{Get the game log of a player for a season and game type}
\usage{
ns_player_game_log(player = 8480039, season = "now", game_type = "")
}
\arguments{
\item{player}{integer ID (e.g., 8480039)}

\item{season}{integer in YYYYYYYY (e.g., 20242025)}

\item{game_type}{integer in 1:3 (where 1 = pre-season, 2 = regular season, 3
= playoff/post-season) OR character of 'pre', 'regular', or 'playoff'/'post'}
}
\value{
data.frame with one row per game
}
\description{
\code{ns_player_game_log()} retrieves information on each game for a given set of
\code{player}, \code{season}, and \code{game_type}, including but not limited to their ID,
date, and statistics. Access \code{get_players()} for \code{player} and \code{get_seasons()}
for \code{season} references.
}
\examples{
Martin_Necas_gl_regular_20242025 <- ns_player_game_log(
  player    = 8480039,
  season    = 20242025,
  game_type = 2
)
}
