% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Player.R
\name{ns_player_seasons}
\alias{ns_player_seasons}
\title{Get the season(s) and game type(s) of which a player played in}
\usage{
ns_player_seasons(player = 8480039)
}
\arguments{
\item{player}{integer ID (e.g., 8480039)}
}
\value{
data.frame with one row per season
}
\description{
\code{ns_player_seasons()} retrieves information ...
}
\examples{
Martin_Necas_seasons <- ns_player_seasons(player = 8480039)
}
