% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Player.R
\name{ns_player_summary}
\alias{ns_player_summary}
\title{Get the summary of a player}
\usage{
ns_player_summary(player = 8480039)
}
\arguments{
\item{player}{integer ID (e.g., 8480039)}
}
\value{
list with various items
}
\description{
\code{ns_player_summary()} retrieves information on a \code{player}, including but not
limited to his ID, name, bio-metrics, career statistics, and awards. Access
\code{get_players()} for \code{player} reference.
}
\examples{
Martin_Necas_landing <- ns_player_summary(player = 8480039)
}
