% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Player.R
\name{ns_players}
\alias{ns_players}
\title{Get all the players}
\usage{
ns_players()
}
\value{
data.frame with one row per player
}
\description{
\code{ns_players()} retrieves information on each player, including but not
limited to their ID, name, bio-metrics, birth date and location, and
hall-of-fame status.
}
\examples{
# This may take >5s, so skip.
\donttest{all_players <- ns_players()}
}
