% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/League.R
\name{ns_schedule}
\alias{ns_schedule}
\title{Access the schedule for a date}
\usage{
ns_schedule(date = Sys.Date())
}
\arguments{
\item{date}{character in 'YYYY-MM-DD' (e.g., '2025-01-01'); see
\code{\link[=ns_seasons]{ns_seasons()}} for reference}
}
\value{
data.frame with one row per game
}
\description{
\code{ns_schedule()} scrapes the schedule for a given \code{date}.
}
\examples{
schedule_Halloween_2025 <- ns_schedule(date = '2025-10-31')
}
