% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/League.R
\name{ns_schedule}
\alias{ns_schedule}
\title{Get the schedule for a date}
\usage{
ns_schedule(date = Sys.Date())
}
\arguments{
\item{date}{character in 'YYYY-MM-DD' (e.g., '2025-01-01')}
}
\value{
data.frame with one row per game
}
\description{
\code{ns_schedule()} returns the schedule for a given \code{date}. Use \code{\link[=ns_seasons]{ns_seasons()}}
for \code{date} reference.
}
\examples{
schedule_Halloween_2025 <- ns_schedule(date = '2025-10-31')
}
