% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Game.R
\name{ns_scores}
\alias{ns_scores}
\title{Get the score(s) of the game(s) for a date}
\usage{
ns_scores(date = "now")
}
\arguments{
\item{date}{character in 'YYYY-MM-DD' (e.g., '2025-01-01')}
}
\value{
data.frame with one row per game
}
\description{
\code{ns_scores()} retrieves information on each game for a given \code{date},
including but not limited to their ID; type; venue; start time; period and
intermission clocks; and home and away teams' IDs, names, and scores. Access
\code{get_seasons()} for \code{date} reference.
}
\examples{
scores_Halloween_2025 <- ns_scores(date = '2025-10-31')
}
