% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/League.R
\name{ns_season}
\alias{ns_season}
\title{Get the season as of now}
\usage{
ns_season()
}
\value{
integer in YYYYYYYY (e.g., 20242025)
}
\description{
\code{ns_season()} returns the ID of the current season.
}
\examples{
season_now <- ns_season()
}
