% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/League.R
\name{ns_seasons}
\alias{ns_seasons}
\title{Get all seasons}
\usage{
ns_seasons()
}
\value{
data.frame with one row per season
}
\description{
\code{ns_seasons()} retrieves information on each season, including but not
limited to their ID; start and end dates; number of regular season and
playoff games; Stanley Cup owner; Olympics participation; entry and
supplemental draft, conference-division, win-tie-loss, and wildcard
regulations.
}
\examples{
all_seasons <- ns_seasons()
}
