% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/League.R
\name{ns_seasons}
\alias{ns_seasons}
\title{Access all the seasons}
\usage{
ns_seasons()
}
\value{
data.frame with one row per season
}
\description{
\code{ns_seasons()} scrapes information on all the seasons.
}
\examples{
all_seasons <- ns_seasons()
}
