% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/League.R
\name{ns_seasons}
\alias{ns_seasons}
\title{Get all the seasons}
\usage{
ns_seasons()
}
\value{
data.frame with one row per season
}
\description{
\code{ns_seasons()} returns information on all the seasons, including but not
limited to the start & end dates, minimum required games for goalie
statistics leaders, Olympics participation, and the use of the following:
all-star games, entry drafts, conferences, divisions, and points for OT
losses.
}
\examples{
all_seasons <- ns_seasons()
}
