% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Playoff.R
\name{ns_series}
\alias{ns_series}
\title{Get all the playoff series (games)}
\usage{
ns_series()
}
\value{
data.frame with one row per game per series
}
\description{
\code{ns_series()} retrieves ...
}
\examples{
all_series <- ns_series()
}
