% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Playoff.R
\name{ns_series_schedule}
\alias{ns_series_schedule}
\title{Access the playoff schedule for a season and series}
\usage{
ns_series_schedule(season = ns_season(), series = "a")
}
\arguments{
\item{season}{integer in YYYYYYYY (e.g., 20242025); see \code{\link[=ns_seasons]{ns_seasons()}} for
reference}

\item{series}{one-letter code (e.g., 'O'); see \code{\link[=ns_series]{ns_series()}} and/or
\code{\link[=ns_bracket]{ns_bracket()}} for reference}
}
\value{
data.frame with one row per game
}
\description{
\code{ns_series_schedule()} scrapes the playoff schedule for a given set of
\code{season} and \code{series}.
}
\examples{
SCF_schedule_20212022 <- ns_series_schedule(
  season = 20212022, 
  series = 'O'
)
}
