% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Game.R
\name{ns_shifts}
\alias{ns_shifts}
\title{Access the shift charts for a game}
\usage{
ns_shifts(game = 2023030417)
}
\arguments{
\item{game}{integer ID (e.g., 2025020275); see \code{\link[=ns_games]{ns_games()}} for reference}
}
\value{
data.frame with one row per shift
}
\description{
\code{ns_shifts()} scrapes the shift charts for a given \code{game}.
}
\examples{
shifts_Martin_Necas_legacy_game <- ns_shifts(game = 2025020275)
}
