% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Player.R
\name{ns_spotlight_players}
\alias{ns_spotlight_players}
\title{Get the spotlight players}
\usage{
ns_spotlight_players()
}
\value{
data.frame with one row per player
}
\description{
\code{ns_spotlight_players()} returns the spotlight players, including but not limited to each player's ID, name, and team.
}
\examples{
spotlight_players <- ns_spotlight_players()
}
