% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/League.R
\name{ns_standings}
\alias{ns_standings}
\title{Access the standings for a date}
\usage{
ns_standings(date = "now")
}
\arguments{
\item{date}{character in 'YYYY-MM-DD' (e.g., '2025-01-01'); see
\code{\link[=ns_seasons]{ns_seasons()}} for reference}
}
\value{
data.frame with one row per team
}
\description{
\code{ns_standings()} scrapes the standings for a given \code{date}.
}
\examples{
standings_Halloween_2025 <- ns_standings(date = '2025-10-31')
}
