% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/League.R
\name{ns_standings_rules}
\alias{ns_standings_rules}
\title{Get the standings rules for all the seasons}
\usage{
ns_standings_rules()
}
\value{
data.frame with one row per season
}
\description{
\code{ns_standings_rules()} returns information about the standings for all the seasons, including but not limited to each season's ID and each standing's start & end dates and rules.
}
\examples{
standings_rules <- ns_standings_rules()
}
