% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/League.R
\name{ns_standings_rules}
\alias{ns_standings_rules}
\title{Get the standings rules by season}
\usage{
ns_standings_rules()
}
\value{
data.frame with one row per season
}
\description{
\code{ns_standings_rules()} returns the standings rules by season, including but
not limited to the use of the following: conferences, divisions, wildcards,
points for regulation wins, points for OT losses, and ties.
}
\examples{
standings_rules <- ns_standings_rules()
}
