% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Other.R
\name{ns_streams}
\alias{ns_streams}
\title{Get all the streams}
\usage{
ns_streams()
}
\value{
data.frame with one row per stream
}
\description{
\code{ns_streams()} retrieves information on each stream, including but not
limited to their ID, name, and URL.
}
\examples{
all_streams <- ns_streams()
}
