% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EDGETeam.R
\name{ns_team_edge_seasons}
\alias{ns_team_edge_seasons}
\title{Get the season(s) and game type(s) of which the NHL recorded team EDGE
statistics}
\usage{
ns_team_edge_seasons()
}
\value{
data.frame with one row per season
}
\description{
\code{ns_team_edge_seasons} retrieves information on each
}
\examples{
team_EDGE_seasons <- ns_team_edge_seasons()
}
