% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Team.R
\name{ns_team_prospects}
\alias{ns_team_prospects}
\title{Access the prospects for a team and position}
\usage{
ns_team_prospects(team = 1, position = "forwards")
}
\arguments{
\item{team}{integer ID (e.g., 21), character full name (e.g., 'Colorado
Avalanche'), OR three-letter code (e.g., 'COL'); see \code{\link[=ns_teams]{ns_teams()}} for
reference}

\item{position}{character of 'f'/'forwards', 'd'/'defensemen', or
'g'/'goalies'}
}
\value{
data.frame with one row per player
}
\description{
\code{ns_team_prospects()} scrapes the prospects for a given set of \code{team} and
\code{position}.
}
\examples{
COL_forward_prospects <- ns_team_prospects(
  team     = 21,
  position = 'F'
)
}
