% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Team.R
\name{ns_team_season_schedule}
\alias{ns_team_season_schedule}
\title{Access the schedule for a team and season}
\usage{
ns_team_season_schedule(team = 1, season = "now")
}
\arguments{
\item{team}{integer ID (e.g., 21), character full name (e.g., 'Colorado
Avalanche'), OR three-letter code (e.g., 'COL'); see \code{\link[=ns_teams]{ns_teams()}} for
reference}

\item{season}{integer in YYYYYYYY (e.g., 20242025); see \code{\link[=ns_seasons]{ns_seasons()}} for
reference}
}
\value{
data.frame with one row per game
}
\description{
\code{ns_team_season_schedule()} scrapes the schedule for a given set of \code{team}
and \code{season}.
}
\examples{
COL_schedule_20252026 <- ns_team_season_schedule(
  team   = 21, 
  season = 20252026
)
}
