% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Team.R
\name{ns_team_season_statistics}
\alias{ns_team_season_statistics}
\alias{ns_team_season_stats}
\title{Get the statistics for all the teams by season and game type}
\usage{
ns_team_season_statistics()

ns_team_season_stats()
}
\value{
data.frame with one row per team per game type per season
}
\description{
\code{ns_team_season_statistics()} retrieves information on ...
}
\examples{
team_season_stats <- ns_team_season_statistics()
}
