% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Team.R
\name{ns_team_seasons}
\alias{ns_team_seasons}
\title{Access the season(s) and game type(s) in which a team played}
\usage{
ns_team_seasons(team = 1)
}
\arguments{
\item{team}{integer ID (e.g., 21), character full name (e.g., 'Colorado
Avalanche'), OR three-letter code (e.g., 'COL'); see \code{\link[=ns_teams]{ns_teams()}} for
reference}
}
\value{
data.frame with one row per season
}
\description{
\code{ns_team_seasons()} scrapes the season(s) and game type(s) in which a team
played in the NHL.
}
\examples{
COL_seasons <- ns_team_seasons(team = 21)
}
