% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Team.R
\name{ns_teams}
\alias{ns_teams}
\title{Access all the teams}
\usage{
ns_teams()
}
\value{
data.frame with one row per team
}
\description{
\code{ns_teams()} scrapes information on all the teams.
}
\examples{
all_teams <- ns_teams()
}
