% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/League.R
\name{ns_venues}
\alias{ns_venues}
\title{Access all the venues}
\usage{
ns_venues()
}
\value{
data.frame with one row per venue
}
\description{
\code{ns_venues()} scrapes information on all the venues.
}
\examples{
all_venues <- ns_venues()
}
