% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Game.R
\name{ns_wsc_play_by_play}
\alias{ns_wsc_play_by_play}
\alias{ns_wsc_pbp}
\title{Access the World Showcase (WSC) play-by-play for a game}
\usage{
ns_wsc_play_by_play(game = 2023030417)

ns_wsc_pbp(game = 2023030417)
}
\arguments{
\item{game}{integer ID (e.g., 2025020275); see \code{\link[=ns_games]{ns_games()}} for reference}
}
\value{
data.frame with one row per event (play)
}
\description{
\code{ns_wsc_play_by_play()} scrapes the WSC play-by-play for given \code{game}.
}
\examples{
wsc_pbp_Martin_Necas_legacy_game <- ns_wsc_play_by_play(game = 2025020275)
}
