% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Game.R
\name{ns_wsc_summary}
\alias{ns_wsc_summary}
\title{Access the World Showcase (WSC) summary for a game}
\usage{
ns_wsc_summary(game = 2023030417)
}
\arguments{
\item{game}{integer ID (e.g., 2025020275); see \code{\link[=ns_games]{ns_games()}} for reference}
}
\value{
list of various items
}
\description{
\code{ns_wsc_summary()} scrapes the WSC summary for a given \code{game}.
}
\examples{
wsc_summary_Martin_Necas_legacy_game <- ns_wsc_summary(game = 2025020275)
}
