% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Official.R
\name{officials}
\alias{officials}
\title{Access all the officials}
\usage{
officials()
}
\value{
data.frame with one row per official
}
\description{
\code{officials()} scrapes all the officials.
}
\examples{
all_officials <- officials()
}
