% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Official.R
\name{officials}
\alias{officials}
\title{Access all the officials}
\usage{
officials()
}
\value{
data.frame with one row per official
}
\description{
\code{officials()} retrieves all the officials as a \code{data.frame} where each row represents official and includes detail on player identity, role, handedness, and biographical profile plus coach/management/officiating identity and assignment history.
}
\examples{
all_officials <- officials()
}
