% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Event.R
\name{penalty_shots}
\alias{penalty_shots}
\alias{ps}
\title{Access all the penalty shots}
\usage{
penalty_shots()

ps()
}
\value{
data.frame with one row per penalty shot
}
\description{
\code{penalty_shots()} scrapes all the penalty shots.
}
\examples{
all_ps <- penalty_shots()
}
