% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Player.R
\name{player_summary}
\alias{player_summary}
\title{Access the summary for a player}
\usage{
player_summary(player = 8478402)
}
\arguments{
\item{player}{integer ID (e.g., 8480039); see \code{\link[=players]{players()}} for reference}
}
\value{
list with various items
}
\description{
\code{player_summary()} scrapes the summary for a given \code{player}.
}
\examples{
Martin_Necas_summary <- player_summary(player = 8480039)
}
