% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Player.R
\name{players}
\alias{players}
\title{Access all the players}
\usage{
players()
}
\value{
data.frame with one row per player
}
\description{
\code{players()} scrapes all the players.
}
\examples{
# May take >5s, so skip.
\donttest{all_players <- players()}
}
