% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Load.R
\name{replays}
\alias{replays}
\title{Access the replays for a season}
\usage{
replays(season = 20242025)
}
\arguments{
\item{season}{integer in YYYYYYYY (e.g., 20242025); see \code{\link[=seasons]{seasons()}} for
reference}
}
\value{
data.frame with one row per decisecond
}
\description{
\code{replays()} loads the replays for a given \code{season}.
}
\examples{
# May take >5s, so skip.
\donttest{replays_20252026 <- replays(season = 20252026)}
}
