% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Game.R
\name{scores}
\alias{scores}
\title{Access the scores for a date}
\usage{
scores(date = "now")
}
\arguments{
\item{date}{character in 'YYYY-MM-DD' (e.g., '2025-01-01'); see
\code{\link[=seasons]{seasons()}} for reference}
}
\value{
data.frame with one row per game
}
\description{
\code{scores()} scrapes the scores for a given \code{date}.
}
\examples{
scores_Halloween_2025 <- scores(date = '2025-10-31')
}
