% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/League.R
\name{season_now}
\alias{season_now}
\title{Access the season as of now}
\usage{
season_now()
}
\value{
integer in YYYYYYYY (e.g., 20242025)
}
\description{
\code{season_now()} retrieves the season as of now and returns a scalar integer used as the current-context default in season/game-type dependent wrappers.
}
\examples{
season_now <- season_now()
}
