% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/League.R
\name{seasons}
\alias{seasons}
\title{Access all the seasons}
\usage{
seasons()
}
\value{
data.frame with one row per season
}
\description{
\code{seasons()} retrieves all the seasons as a \code{data.frame} where each row represents season and includes detail on date/season filtering windows and chronological context.
}
\examples{
all_seasons <- seasons()
}
