% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Playoff.R
\name{series}
\alias{series}
\title{Access all the playoff series by game}
\usage{
series()
}
\value{
data.frame with one row per game per series
}
\description{
\code{series()} scrapes all the playoff series by game.
}
\examples{
# May take >5s, so skip.
\donttest{all_series <- series()}
}
