% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Playoff.R
\name{series_schedule}
\alias{series_schedule}
\title{Access the playoff schedule for a season and series}
\usage{
series_schedule(season = season_now(), series = "a")
}
\arguments{
\item{season}{integer in YYYYYYYY (e.g., 20242025); see \code{\link[=seasons]{seasons()}} for
reference}

\item{series}{one-letter code (e.g., 'O'); see \code{\link[=series]{series()}} and/or
\code{\link[=bracket]{bracket()}} for reference}
}
\value{
data.frame with one row per game
}
\description{
\code{series_schedule()} scrapes the playoff schedule for a given set of
\code{season} and \code{series}.
}
\examples{
SCF_schedule_20212022 <- series_schedule(
  season = 20212022, 
  series = 'O'
)
}
