% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Game.R
\name{shift_chart}
\alias{shift_chart}
\title{Access the shift chart for a game}
\usage{
shift_chart(game = 2023030417)
}
\arguments{
\item{game}{integer ID (e.g., 2025020275); see \code{\link[=games]{games()}} for reference}
}
\value{
data.frame with one row per shift
}
\description{
\code{shift_chart()} retrieves the shift chart for a game as a \code{data.frame} where each row represents shift and includes detail on game timeline state, period/clock progression, and matchup flow, date/season filtering windows and chronological context, and team identity, affiliation, and matchup-side context.
}
\examples{
shifts_Martin_Necas_legacy_game <- shift_chart(game = 2025020275)
}
