% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Load.R
\name{shift_charts}
\alias{shift_charts}
\title{Access the shift charts for a season}
\usage{
shift_charts(season = 20242025)
}
\arguments{
\item{season}{integer in YYYYYYYY (e.g., 20242025); see \code{\link[=seasons]{seasons()}} for
reference}
}
\value{
data.frame with one row per event (play) per game
}
\description{
\code{shift_charts()} loads the shift charts for a given \code{season}.
}
\examples{
# May take >5s, so skip.
\donttest{shift_charts_20212022 <- shift_charts(season = 20212022)}
}
