% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Game.R
\name{shifts}
\alias{shifts}
\title{Access the shift charts for a game}
\usage{
shifts(game = 2023030417)
}
\arguments{
\item{game}{integer ID (e.g., 2025020275); see \code{\link[=games]{games()}} for reference}
}
\value{
data.frame with one row per shift
}
\description{
\code{shifts()} scrapes the shift charts for a given \code{game}.
}
\examples{
shifts_Martin_Necas_legacy_game <- shifts(game = 2025020275)
}
