% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SkaterEDGE.R
\name{skater_edge_seasons}
\alias{skater_edge_seasons}
\title{Access the season(s) and game type(s) in which there exists skater EDGE
statistics}
\usage{
skater_edge_seasons()
}
\value{
data.frame with one row per season
}
\description{
\code{skater_edge_seasons()} retrieves the season(s) and game type(s) in which there exists skater EDGE statistics as a \code{data.frame} where each row represents season and includes detail on date/season filtering windows and chronological context plus NHL EDGE style tracking outputs and relative-performance context.
}
\examples{
skater_EDGE_seasons <- skater_edge_seasons()
}
