% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SkaterEDGE.R
\name{skater_edge_seasons}
\alias{skater_edge_seasons}
\title{Get the season(s) and game type(s) for which the NHL recorded skater EDGE
statistics}
\usage{
skater_edge_seasons()
}
\value{
data.frame with one row per season
}
\description{
\code{skater_edge_seasons} retrieves information on each
}
\examples{
skater_EDGE_seasons <- skater_edge_seasons()
}
