% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Skater.R
\name{skater_milestones}
\alias{skater_milestones}
\title{Access the skaters on milestone watch}
\usage{
skater_milestones()
}
\value{
data.frame with one row per player
}
\description{
\code{skater_milestones()} scrapes the skaters on milestone watch.
}
\examples{
skater_milestones <- skater_milestones()
}
