% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Skater.R
\name{skater_milestones}
\alias{skater_milestones}
\title{Get the skaters on milestone watch}
\usage{
skater_milestones()
}
\value{
data.frame with one row per skater
}
\description{
\code{skater_milestones()} retrieves information on each skater close to a
milestone, including but not limited to their ID, name, and statistics.
}
\examples{
skater_milestones <- skater_milestones()
}
