% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Skater.R
\name{skater_regular_statistics}
\alias{skater_regular_statistics}
\alias{skater_regular_stats}
\title{Access the career regular season statistics for all the skaters}
\usage{
skater_regular_statistics()

skater_regular_stats()
}
\value{
data.frame with one row per player
}
\description{
\code{skater_regular_statistics()} scrapes the career regular season statistics
for all the skaters.
}
\examples{
skater_regular_stats <- skater_regular_statistics()
}
