% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Skater.R
\name{skater_report_configurations}
\alias{skater_report_configurations}
\alias{skater_report_configs}
\title{Access the configurations for skater reports}
\usage{
skater_report_configurations()

skater_report_configs()
}
\value{
list with various items
}
\description{
\code{skater_report_configurations()} scrapes the configurations for
\code{\link[=skater_season_report]{skater_season_report()}} and \code{\link[=skater_game_report]{skater_game_report()}}.
}
\examples{
skater_report_configs <- skater_report_configurations()
}
