% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Skater.R
\name{skater_season_statistics}
\alias{skater_season_statistics}
\alias{skater_season_stats}
\title{Access the statistics for all the skaters by season, game type, and team}
\usage{
skater_season_statistics()

skater_season_stats()
}
\value{
data.frame with one row per player per season per game type,
separated by team if applicable
}
\description{
\code{skater_season_statistics()} retrieves the statistics for all the skaters by season, game type, and team as a \code{data.frame} where each row represents player per season per game type, separated by team if applicable and includes detail on date/season filtering windows and chronological context, team identity, affiliation, and matchup-side context, and player identity, role, handedness, and biographical profile.
}
\examples{
# May take >5s, so skip.
\donttest{skater_season_stats <- skater_season_statistics()}
}
