% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Skater.R
\name{skater_season_statistics}
\alias{skater_season_statistics}
\alias{skater_season_stats}
\title{Get the statistics for all the skaters by season, game type, and team}
\usage{
skater_season_statistics()

skater_season_stats()
}
\value{
data.frame with one row per skater per game type per season,
separated by team when necessary
}
\description{
\code{skater_season_statistics()} retrieves information on ...
}
\examples{
skater_season_stats <- skater_season_statistics()
}
