% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Skater.R
\name{skater_series_statistics}
\alias{skater_series_statistics}
\alias{skater_series_stats}
\title{Access the playoff statistics for all the skaters by series}
\usage{
skater_series_statistics()

skater_series_stats()
}
\value{
data.frame with one row per player per series
}
\description{
\code{skater_series_statistics()} retrieves the playoff statistics for all the skaters by series as a \code{data.frame} where each row represents player per series and includes detail on date/season filtering windows and chronological context, team identity, affiliation, and matchup-side context, and player identity, role, handedness, and biographical profile.
}
\examples{
# May take >5s, so skip.
\donttest{skater_series_stats <- skater_series_statistics()}
}
