% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Skater.R
\name{skater_statistics}
\alias{skater_statistics}
\alias{skater_stats}
\title{Access the career statistics for all the skaters}
\usage{
skater_statistics()

skater_stats()
}
\value{
data.frame with one row per player
}
\description{
\code{skater_statistics()} scrapes the career statistics for all the skaters.
}
\examples{
skater_stats <- skater_statistics()
}
