% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Player.R
\name{spotlight_players}
\alias{spotlight_players}
\title{Access the spotlight players}
\usage{
spotlight_players()
}
\value{
data.frame with one row per player
}
\description{
\code{spotlight_players()} retrieves the spotlight players as a \code{data.frame} where each row represents player and includes detail on team identity, affiliation, and matchup-side context plus player identity, role, handedness, and biographical profile.
}
\examples{
spotlight_players <- spotlight_players()
}
