% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Other.R
\name{streams}
\alias{streams}
\title{Access all the streams}
\usage{
streams()
}
\value{
data.frame with one row per stream
}
\description{
\code{streams()} retrieves all the streams as a \code{data.frame} where each row represents stream and includes detail on reference metadata, regional context, and media availability detail.
}
\examples{
all_streams <- streams()
}
