% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Other.R
\name{streams}
\alias{streams}
\title{Access all the streams}
\usage{
streams()
}
\value{
data.frame with one row per stream
}
\description{
\code{streams()} scrapes all the streams.
}
\examples{
all_streams <- streams()
}
