% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TeamEDGE.R
\name{team_edge_seasons}
\alias{team_edge_seasons}
\title{Access the season(s) and game type(s) in which there exists team EDGE
statistics}
\usage{
team_edge_seasons()
}
\value{
data.frame with one row per season
}
\description{
\code{team_edge_seasons()} retrieves the season(s) and game type(s) in which there exists team EDGE statistics as a \code{data.frame} where each row represents season and includes detail on date/season filtering windows and chronological context plus NHL EDGE style tracking outputs and relative-performance context.
}
\examples{
team_EDGE_seasons <- team_edge_seasons()
}
