% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TeamEDGE.R
\name{team_edge_skating_distance}
\alias{team_edge_skating_distance}
\title{Access the EDGE skating distance statistics for a team, season, game type,
and report type}
\usage{
team_edge_skating_distance(
  team = 1,
  season = "now",
  game_type = "",
  report_type = "details"
)
}
\arguments{
\item{team}{integer ID (e.g., 21), character full name (e.g., 'Colorado
Avalanche'), OR three-letter code (e.g., 'COL'); see \code{\link[=teams]{teams()}} for
reference}

\item{season}{integer in YYYYYYYY (e.g., 20242025); see
\code{\link[=team_edge_seasons]{team_edge_seasons()}} for reference}

\item{game_type}{integer in 1:3 (where 1 = pre-season, 2 = regular season, 3
= playoff/post-season) OR character of 'pre', 'regular', or
'playoff'/'post'; see \code{\link[=team_edge_seasons]{team_edge_seasons()}} for reference}

\item{report_type}{character of 'd'/'details' or 'l'/'l10'/'last 10'}
}
\value{
data.frame with one row per combination of strength state and
position (report_type = 'details') or game (report_type = 'last 10')
game
}
\description{
\code{team_edge_skating_distance()} scrapes the EDGE skating distance
statistics for a given set of \code{team}, \code{season}, \code{game_type}, and
\code{report_type}.
}
\examples{
COL_L10_skating_distance_regular_20242025 <- team_edge_skating_distance(
  team        = 21,
  season      = 20242025,
  game_type   = 2,
  report_type = 'L'
)
}
