% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Team.R
\name{team_logos}
\alias{team_logos}
\title{Access all the team logos}
\usage{
team_logos()
}
\value{
data.frame with one row per logo
}
\description{
\code{team_logos()} scrapes all the team logos.
}
\examples{
all_team_logos <- team_logos()
}
