% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Team.R
\name{team_logos}
\alias{team_logos}
\title{Access all the team logos}
\usage{
team_logos()
}
\value{
data.frame with one row per logo
}
\description{
\code{team_logos()} retrieves all the team logos as a \code{data.frame} where each row represents logo and includes detail on team identity, affiliation, and matchup-side context.
}
\examples{
all_team_logos <- team_logos()
}
