% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Team.R
\name{team_report_configurations}
\alias{team_report_configurations}
\alias{team_report_configs}
\title{Access the configurations for team reports}
\usage{
team_report_configurations()

team_report_configs()
}
\value{
list with various items
}
\description{
\code{team_report_configurations()} retrieves the configurations for team reports as a nested \code{list} that separates summary and detail blocks for situational splits across home/road, strength state, and overtime/shootout states plus configuration catalogs for valid report categories and filters.
}
\examples{
team_report_configs <- team_report_configurations()
}
