% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Team.R
\name{team_season_statistics}
\alias{team_season_statistics}
\alias{team_season_stats}
\title{Access the statistics for all the teams by season and game type}
\usage{
team_season_statistics()

team_season_stats()
}
\value{
data.frame with one row per team per season per game type
}
\description{
\code{team_season_statistics()} retrieves the statistics for all the teams by season and game type as a \code{data.frame} where each row represents team per season per game type and includes detail on date/season filtering windows and chronological context, team identity, affiliation, and matchup-side context, and production, workload, efficiency, and result-level performance outcomes.
}
\examples{
# May take >5s, so skip.
\donttest{team_season_statistics <- team_season_statistics()}
}
