% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Team.R
\name{teams}
\alias{teams}
\title{Access all the teams}
\usage{
teams()
}
\value{
data.frame with one row per team
}
\description{
\code{teams()} retrieves all the teams as a \code{data.frame} where each row represents team and includes detail on team identity, affiliation, and matchup-side context.
}
\examples{
all_teams <- teams()
}
