% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Other.R
\name{tv_schedule}
\alias{tv_schedule}
\title{Access the NHL Network TV schedule for a date}
\usage{
tv_schedule(date = "now")
}
\arguments{
\item{date}{character in 'YYYY-MM-DD' (e.g., '2025-01-01'); see
\code{\link[=seasons]{seasons()}} for reference}
}
\value{
data.frame with one row per program
}
\description{
\code{tv_schedule()} scrapes the NHL Network TV schedule for a given \code{date}.
}
\examples{
tv_schedule_Halloween_2025 <- tv_schedule(date = '2025-10-31')
}
