% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/League.R
\name{venues}
\alias{venues}
\title{Access all the venues}
\usage{
venues()
}
\value{
data.frame with one row per venue
}
\description{
\code{venues()} scrapes all the venues.
}
\examples{
all_venues <- venues()
}
