% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/League.R
\name{venues}
\alias{venues}
\title{Access all the venues}
\usage{
venues()
}
\value{
data.frame with one row per venue
}
\description{
\code{venues()} retrieves all the venues as a \code{data.frame} where each row represents venue and includes detail on venue/location geography and regional metadata.
}
\examples{
all_venues <- venues()
}
