% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Game.R
\name{wsc_play_by_play}
\alias{wsc_play_by_play}
\alias{wsc_pbp}
\title{Access the World Showcase (WSC) play-by-play for a game}
\usage{
wsc_play_by_play(game = 2023030417)

wsc_pbp(game = 2023030417)
}
\arguments{
\item{game}{integer ID (e.g., 2025020275); see \code{\link[=games]{games()}} for reference}
}
\value{
data.frame with one row per event (play)
}
\description{
\code{wsc_play_by_play()} retrieves the World Showcase (WSC) play-by-play for a game as a \code{data.frame} where each row represents event and includes detail on game timeline state, period/clock progression, and matchup flow, date/season filtering windows and chronological context, and player identity, role, handedness, and biographical profile.
}
\examples{
wsc_pbp_Martin_Necas_legacy_game <- wsc_play_by_play(game = 2025020275)
}
