% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Share.R
\name{x_game_shot_locations}
\alias{x_game_shot_locations}
\alias{x_game_shot_locs}
\title{Save an X (Twitter) share-able shot-location plot for a game}
\usage{
x_game_shot_locations(game = 2023030417, team = "home", model = 1, save = TRUE)

x_game_shot_locs(game = 2023030417, team = "home", model = 1)
}
\arguments{
\item{game}{integer ID (e.g., 2025020275); see \code{\link[=games]{games()}} for reference}

\item{team}{character of 'h'/'home' or 'a'/'away'}

\item{model}{integer in 1:3 indicating which expected goals model to use
(e.g., 1); see \code{\link[=calculate_expected_goals_v1]{calculate_expected_goals_v1()}},
\code{\link[=calculate_expected_goals_v2]{calculate_expected_goals_v2()}}, and/or \code{\link[=calculate_expected_goals_v3]{calculate_expected_goals_v3()}} for
reference}

\item{save}{logical only FALSE for tests}
}
\value{
\code{NULL}
}
\description{
\code{x_game_shot_locations()} saves an X share-able shot-location plot for a
given \code{game}.
}
\examples{
# May take >5s, so skip.
\donttest{x_game_shot_locations(
  game  = 2023030417, 
  model = 1, 
  team  = 'H',
  save  = FALSE
)}
}
